Reorganised structure of files and types using barrel files
