+++ js/features/Navigation/types.ts
@@ -1,4 +1,1 @@
-export type NavItemData = {
-  name: string;
-  key: string;
-};
+export type { NavItemData } from '../../types';
