+++ js/features/Policy/types.ts
@@ -1,7 +1,1 @@
-import { LinkData } from "../../ui/Link/types";
-
-export type PolicyData = {
-  id: string;
-  type: string;
-  link: LinkData;
-};
+export type { PolicyData } from '../../types';
