+++ js/features/Select/types.ts
@@ -1,4 +1,1 @@
-export type SelectItemData = {
-  name: string;
-  key: string;
-}
+export type { SelectItemData } from '../../types';
