+++ js/features/index.js
@@ -1,0 +1,9 @@
+export { Burger } from './Burger/index.js';
+export { Form } from './Form/index.js';
+export { Input } from './Input/index.js';
+export { Lang } from './Lang/index.js';
+export { Logo } from './Logo/index.js';
+export { Navigation } from './Navigation/index.js';
+export { Policy } from './Policy/index.js';
+export { Select } from './Select/index.js';
+export { Theme } from './Theme/index.js';
