+++ js/ui/Column/types.ts
@@ -1,6 +1,1 @@
-import { LinkData } from '../Link/types';
-
-export type ColumnData = {
-  name: string;
-  links: LinkData[];
-};
+export type { ColumnData } from '../../types';
