+++ js/ui/Link/types.ts
@@ -1,6 +1,2 @@
-export type LinkData = {
-  name: string;
-  url: string;
-  type: string;
-  content: string;
-};
+export type { LinkData } from '../../types';
+
