+++ js/ui/icons/IconApple/IconApple.js
@@ -1,0 +1,15 @@
+/* eslint-disable */
+
+/**
+ * @function IconApple
+ * @returns {string} HTML
+ */
+
+export const IconApple = () => `
+  <svg width="150" height="44" viewBox="0 0 150 44" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <rect x="0.5" y="0.5" width="149" height="43" rx="4.5" stroke="#A6A6A6"/>
+    <path d="M62.87 32.614C62.942 32.77 62.978 32.926 62.978 33.082C62.978 33.358 62.864 33.592 62.636 33.784C62.42 33.976 62.168 34.072 61.88 34.072C61.688 34.072 61.508 34.024 61.34 33.928C61.172 33.82 61.04 33.658 60.944 33.442L59.864 31.03H53.528L52.448 33.442C52.352 33.658 52.22 33.82 52.052 33.928C51.884 34.024 51.698 34.072 51.494 34.072C51.218 34.072 50.966 33.976 50.738 33.784C50.51 33.592 50.396 33.358 50.396 33.082C50.396 32.926 50.432 32.77 50.504 32.614L55.472 21.94C55.58 21.7 55.742 21.52 55.958 21.4C56.186 21.268 56.426 21.202 56.678 21.202C56.93 21.202 57.164 21.268 57.38 21.4C57.608 21.52 57.776 21.7 57.884 21.94L62.87 32.614ZM54.32 29.23H59.072L56.696 23.884L54.32 29.23ZM69.5874 24.982C70.3434 24.982 71.0154 25.174 71.6034 25.558C72.2034 25.942 72.6654 26.488 72.9894 27.196C73.3254 27.892 73.4934 28.696 73.4934 29.608C73.4934 30.52 73.3254 31.318 72.9894 32.002C72.6654 32.686 72.2094 33.214 71.6214 33.586C71.0334 33.958 70.3554 34.144 69.5874 34.144C68.9634 34.144 68.4054 34.012 67.9134 33.748C67.4214 33.484 67.0434 33.118 66.7794 32.65V36.322C66.7794 36.634 66.6774 36.88 66.4734 37.06C66.2694 37.24 65.9994 37.33 65.6634 37.33C65.3274 37.33 65.0514 37.234 64.8354 37.042C64.6314 36.85 64.5294 36.598 64.5294 36.286V26.08C64.5294 25.756 64.6254 25.498 64.8174 25.306C65.0214 25.114 65.2914 25.018 65.6274 25.018C65.9634 25.018 66.2334 25.114 66.4374 25.306C66.6414 25.498 66.7434 25.756 66.7434 26.08V26.53C67.0074 26.038 67.3854 25.66 67.8774 25.396C68.3814 25.12 68.9514 24.982 69.5874 24.982ZM68.9934 32.398C69.7014 32.398 70.2534 32.158 70.6494 31.678C71.0454 31.186 71.2434 30.496 71.2434 29.608C71.2434 28.696 71.0454 27.988 70.6494 27.484C70.2534 26.98 69.7014 26.728 68.9934 26.728C68.2734 26.728 67.7154 26.974 67.3194 27.466C66.9354 27.958 66.7434 28.66 66.7434 29.572C66.7434 30.472 66.9354 31.168 67.3194 31.66C67.7154 32.152 68.2734 32.398 68.9934 32.398ZM80.3979 24.982C81.1539 24.982 81.8259 25.174 82.4139 25.558C83.0139 25.942 83.4759 26.488 83.7999 27.196C84.1359 27.892 84.3039 28.696 84.3039 29.608C84.3039 30.52 84.1359 31.318 83.7999 32.002C83.4759 32.686 83.0199 33.214 82.4319 33.586C81.8439 33.958 81.1659 34.144 80.3979 34.144C79.7739 34.144 79.2159 34.012 78.7239 33.748C78.2319 33.484 77.8539 33.118 77.5899 32.65V36.322C77.5899 36.634 77.4879 36.88 77.2839 37.06C77.0799 37.24 76.8099 37.33 76.4739 37.33C76.1379 37.33 75.8619 37.234 75.6459 37.042C75.4419 36.85 75.3399 36.598 75.3399 36.286V26.08C75.3399 25.756 75.4359 25.498 75.6279 25.306C75.8319 25.114 76.1019 25.018 76.4379 25.018C76.7739 25.018 77.0439 25.114 77.2479 25.306C77.4519 25.498 77.5539 25.756 77.5539 26.08V26.53C77.8179 26.038 78.1959 25.66 78.6879 25.396C79.1919 25.12 79.7619 24.982 80.3979 24.982ZM79.8039 32.398C80.5119 32.398 81.0639 32.158 81.4599 31.678C81.8559 31.186 82.0539 30.496 82.0539 29.608C82.0539 28.696 81.8559 27.988 81.4599 27.484C81.0639 26.98 80.5119 26.728 79.8039 26.728C79.0839 26.728 78.5259 26.974 78.1299 27.466C77.7459 27.958 77.5539 28.66 77.5539 29.572C77.5539 30.472 77.7459 31.168 78.1299 31.66C78.5259 32.152 79.0839 32.398 79.8039 32.398ZM95.6272 34.162C94.7032 34.162 93.8272 34.042 92.9992 33.802C92.1832 33.55 91.5172 33.214 91.0012 32.794C90.7252 32.59 90.5872 32.302 90.5872 31.93C90.5872 31.678 90.6592 31.462 90.8032 31.282C90.9592 31.09 91.1392 30.994 91.3432 30.994C91.5592 30.994 91.8052 31.078 92.0812 31.246C92.5972 31.618 93.1372 31.888 93.7012 32.056C94.2652 32.224 94.8892 32.308 95.5732 32.308C96.4492 32.308 97.1212 32.158 97.5892 31.858C98.0572 31.546 98.2912 31.096 98.2912 30.508C98.2912 30.052 98.0692 29.704 97.6252 29.464C97.1932 29.212 96.4672 28.972 95.4472 28.744C94.3912 28.516 93.5332 28.252 92.8732 27.952C92.2132 27.652 91.7152 27.268 91.3792 26.8C91.0432 26.332 90.8752 25.738 90.8752 25.018C90.8752 24.274 91.0852 23.608 91.5052 23.02C91.9252 22.432 92.5072 21.976 93.2512 21.652C94.0072 21.316 94.8532 21.148 95.7892 21.148C97.5052 21.148 98.9032 21.604 99.9832 22.516C100.139 22.648 100.247 22.78 100.307 22.912C100.379 23.032 100.415 23.188 100.415 23.38C100.415 23.632 100.337 23.854 100.181 24.046C100.037 24.226 99.8632 24.316 99.6592 24.316C99.5392 24.316 99.4252 24.298 99.3172 24.262C99.2212 24.226 99.0892 24.16 98.9212 24.064C98.4292 23.704 97.9492 23.44 97.4812 23.272C97.0252 23.092 96.4612 23.002 95.7892 23.002C94.9732 23.002 94.3252 23.17 93.8452 23.506C93.3772 23.83 93.1432 24.286 93.1432 24.874C93.1432 25.366 93.3532 25.744 93.7732 26.008C94.1932 26.272 94.9072 26.518 95.9152 26.746C96.9832 26.986 97.8472 27.256 98.5072 27.556C99.1792 27.844 99.6892 28.216 100.037 28.672C100.397 29.128 100.577 29.698 100.577 30.382C100.577 31.126 100.367 31.786 99.9472 32.362C99.5392 32.926 98.9572 33.37 98.2012 33.694C97.4572 34.006 96.5992 34.162 95.6272 34.162ZM106.965 32.434C107.565 32.47 107.865 32.74 107.865 33.244C107.865 33.532 107.745 33.754 107.505 33.91C107.277 34.054 106.947 34.114 106.515 34.09L106.029 34.054C104.013 33.91 103.005 32.83 103.005 30.814V26.89H102.105C101.781 26.89 101.529 26.818 101.349 26.674C101.181 26.53 101.097 26.32 101.097 26.044C101.097 25.768 101.181 25.558 101.349 25.414C101.529 25.27 101.781 25.198 102.105 25.198H103.005V23.542C103.005 23.218 103.107 22.96 103.311 22.768C103.515 22.576 103.791 22.48 104.139 22.48C104.475 22.48 104.745 22.576 104.949 22.768C105.153 22.96 105.255 23.218 105.255 23.542V25.198H106.785C107.109 25.198 107.355 25.27 107.523 25.414C107.703 25.558 107.793 25.768 107.793 26.044C107.793 26.32 107.703 26.53 107.523 26.674C107.355 26.818 107.109 26.89 106.785 26.89H105.255V30.976C105.255 31.864 105.663 32.338 106.479 32.398L106.965 32.434ZM113.051 34.144C112.151 34.144 111.359 33.958 110.675 33.586C110.003 33.214 109.481 32.68 109.109 31.984C108.737 31.288 108.551 30.478 108.551 29.554C108.551 28.63 108.737 27.826 109.109 27.142C109.481 26.446 110.003 25.912 110.675 25.54C111.359 25.168 112.151 24.982 113.051 24.982C113.951 24.982 114.737 25.168 115.409 25.54C116.093 25.912 116.615 26.446 116.975 27.142C117.347 27.826 117.533 28.63 117.533 29.554C117.533 30.478 117.347 31.288 116.975 31.984C116.615 32.68 116.093 33.214 115.409 33.586C114.737 33.958 113.951 34.144 113.051 34.144ZM113.033 32.398C113.765 32.398 114.323 32.158 114.707 31.678C115.091 31.198 115.283 30.49 115.283 29.554C115.283 28.63 115.091 27.928 114.707 27.448C114.323 26.956 113.771 26.71 113.051 26.71C112.331 26.71 111.773 26.956 111.377 27.448C110.993 27.928 110.801 28.63 110.801 29.554C110.801 30.49 110.993 31.198 111.377 31.678C111.761 32.158 112.313 32.398 113.033 32.398ZM124.305 25.018C124.617 24.994 124.863 25.06 125.043 25.216C125.223 25.372 125.313 25.606 125.313 25.918C125.313 26.242 125.235 26.482 125.079 26.638C124.923 26.794 124.641 26.896 124.233 26.944L123.693 26.998C122.985 27.07 122.463 27.31 122.127 27.718C121.803 28.126 121.641 28.636 121.641 29.248V33.028C121.641 33.376 121.533 33.646 121.317 33.838C121.101 34.018 120.831 34.108 120.507 34.108C120.183 34.108 119.913 34.018 119.697 33.838C119.493 33.646 119.391 33.376 119.391 33.028V26.062C119.391 25.726 119.493 25.468 119.697 25.288C119.913 25.108 120.177 25.018 120.489 25.018C120.801 25.018 121.053 25.108 121.245 25.288C121.437 25.456 121.533 25.702 121.533 26.026V26.746C121.761 26.218 122.097 25.81 122.541 25.522C122.997 25.234 123.501 25.072 124.053 25.036L124.305 25.018ZM133.072 31.588C133.276 31.588 133.438 31.666 133.558 31.822C133.69 31.978 133.756 32.188 133.756 32.452C133.756 32.824 133.534 33.136 133.09 33.388C132.682 33.616 132.22 33.802 131.704 33.946C131.188 34.078 130.696 34.144 130.228 34.144C128.812 34.144 127.69 33.736 126.862 32.92C126.034 32.104 125.62 30.988 125.62 29.572C125.62 28.672 125.8 27.874 126.16 27.178C126.52 26.482 127.024 25.942 127.672 25.558C128.332 25.174 129.076 24.982 129.904 24.982C130.696 24.982 131.386 25.156 131.974 25.504C132.562 25.852 133.018 26.344 133.342 26.98C133.666 27.616 133.828 28.366 133.828 29.23C133.828 29.746 133.6 30.004 133.144 30.004H127.834C127.906 30.832 128.14 31.444 128.536 31.84C128.932 32.224 129.508 32.416 130.264 32.416C130.648 32.416 130.984 32.368 131.272 32.272C131.572 32.176 131.908 32.044 132.28 31.876C132.64 31.684 132.904 31.588 133.072 31.588ZM129.958 26.566C129.346 26.566 128.854 26.758 128.482 27.142C128.122 27.526 127.906 28.078 127.834 28.798H131.902C131.878 28.066 131.698 27.514 131.362 27.142C131.026 26.758 130.558 26.566 129.958 26.566Z" fill="white"/>
+    <path d="M51.45 16C51.23 16 51.06 15.9433 50.94 15.83C50.8267 15.71 50.77 15.54 50.77 15.32V9.63C50.77 9.41 50.8267 9.24333 50.94 9.13C51.06 9.01 51.23 8.95 51.45 8.95H53.39C54.57 8.95 55.48 9.25333 56.12 9.86C56.7667 10.4667 57.09 11.3367 57.09 12.47C57.09 13.0367 57.0067 13.54 56.84 13.98C56.68 14.4133 56.44 14.78 56.12 15.08C55.8067 15.38 55.42 15.61 54.96 15.77C54.5067 15.9233 53.9833 16 53.39 16H51.45ZM52.05 14.94H53.31C53.7233 14.94 54.08 14.89 54.38 14.79C54.6867 14.6833 54.94 14.53 55.14 14.33C55.3467 14.1233 55.5 13.8667 55.6 13.56C55.7 13.2467 55.75 12.8833 55.75 12.47C55.75 11.6433 55.5467 11.0267 55.14 10.62C54.7333 10.2133 54.1233 10.01 53.31 10.01H52.05V14.94ZM60.4972 16.1C59.9905 16.1 59.5505 15.9967 59.1772 15.79C58.8039 15.5833 58.5139 15.29 58.3072 14.91C58.1005 14.5233 57.9972 14.07 57.9972 13.55C57.9972 13.1567 58.0539 12.8067 58.1672 12.5C58.2872 12.1867 58.4572 11.92 58.6772 11.7C58.8972 11.4733 59.1605 11.3033 59.4672 11.19C59.7739 11.07 60.1172 11.01 60.4972 11.01C61.0039 11.01 61.4439 11.1133 61.8172 11.32C62.1905 11.5267 62.4805 11.82 62.6872 12.2C62.8939 12.58 62.9972 13.03 62.9972 13.55C62.9972 13.9433 62.9372 14.2967 62.8172 14.61C62.7039 14.9233 62.5372 15.1933 62.3172 15.42C62.0972 15.64 61.8339 15.81 61.5272 15.93C61.2205 16.0433 60.8772 16.1 60.4972 16.1ZM60.4972 15.15C60.7439 15.15 60.9605 15.09 61.1472 14.97C61.3339 14.85 61.4772 14.6733 61.5772 14.44C61.6839 14.2 61.7372 13.9033 61.7372 13.55C61.7372 13.0167 61.6239 12.62 61.3972 12.36C61.1705 12.0933 60.8705 11.96 60.4972 11.96C60.2505 11.96 60.0339 12.02 59.8472 12.14C59.6605 12.2533 59.5139 12.43 59.4072 12.67C59.3072 12.9033 59.2572 13.1967 59.2572 13.55C59.2572 14.0767 59.3705 14.4767 59.5972 14.75C59.8239 15.0167 60.1239 15.15 60.4972 15.15ZM65.8213 16.08C65.6546 16.08 65.5113 16.04 65.3913 15.96C65.2713 15.8733 65.1746 15.74 65.1013 15.56L63.6513 11.86C63.5913 11.7 63.5713 11.56 63.5913 11.44C63.6179 11.3133 63.6813 11.2133 63.7813 11.14C63.8813 11.0667 64.0146 11.03 64.1813 11.03C64.3279 11.03 64.4479 11.0667 64.5413 11.14C64.6346 11.2067 64.7146 11.3367 64.7813 11.53L65.9913 14.86H65.7613L67.0113 11.45C67.0646 11.3033 67.1346 11.1967 67.2213 11.13C67.3146 11.0633 67.4346 11.03 67.5813 11.03C67.7279 11.03 67.8479 11.0667 67.9413 11.14C68.0346 11.2067 68.1046 11.31 68.1513 11.45L69.3813 14.86H69.1713L70.3913 11.5C70.4579 11.32 70.5413 11.1967 70.6413 11.13C70.7479 11.0633 70.8646 11.03 70.9913 11.03C71.1513 11.03 71.2746 11.07 71.3613 11.15C71.4479 11.23 71.4979 11.3333 71.5113 11.46C71.5246 11.58 71.5013 11.7133 71.4413 11.86L70.0013 15.56C69.9346 15.7333 69.8379 15.8633 69.7113 15.95C69.5913 16.0367 69.4479 16.08 69.2813 16.08C69.1146 16.08 68.9679 16.0367 68.8413 15.95C68.7213 15.8633 68.6279 15.7333 68.5612 15.56L67.2513 12.09H67.8213L66.5413 15.55C66.4746 15.73 66.3813 15.8633 66.2613 15.95C66.1413 16.0367 65.9946 16.08 65.8213 16.08ZM73.0766 16.08C72.8766 16.08 72.7233 16.0267 72.6166 15.92C72.51 15.8067 72.4566 15.6467 72.4566 15.44V11.66C72.4566 11.4533 72.51 11.2967 72.6166 11.19C72.7233 11.0833 72.8733 11.03 73.0666 11.03C73.26 11.03 73.41 11.0833 73.5166 11.19C73.6233 11.2967 73.6766 11.4533 73.6766 11.66V12.34L73.5666 12.09C73.7133 11.7367 73.94 11.47 74.2466 11.29C74.56 11.1033 74.9133 11.01 75.3066 11.01C75.7 11.01 76.0233 11.0833 76.2766 11.23C76.53 11.3767 76.72 11.6 76.8466 11.9C76.9733 12.1933 77.0366 12.5667 77.0366 13.02V15.44C77.0366 15.6467 76.9833 15.8067 76.8766 15.92C76.77 16.0267 76.6166 16.08 76.4166 16.08C76.2166 16.08 76.06 16.0267 75.9466 15.92C75.84 15.8067 75.7866 15.6467 75.7866 15.44V13.08C75.7866 12.7 75.7133 12.4233 75.5666 12.25C75.4266 12.0767 75.2066 11.99 74.9066 11.99C74.54 11.99 74.2466 12.1067 74.0266 12.34C73.8133 12.5667 73.7066 12.87 73.7066 13.25V15.44C73.7066 15.8667 73.4966 16.08 73.0766 16.08ZM79.9763 16.1C79.4229 16.1 79.0063 15.9433 78.7263 15.63C78.4463 15.31 78.3063 14.8433 78.3063 14.23V9.5C78.3063 9.29333 78.3596 9.13667 78.4663 9.03C78.5729 8.92333 78.7263 8.87 78.9263 8.87C79.1263 8.87 79.2796 8.92333 79.3863 9.03C79.4996 9.13667 79.5563 9.29333 79.5563 9.5V14.17C79.5563 14.4767 79.6196 14.7033 79.7463 14.85C79.8796 14.9967 80.0663 15.07 80.3063 15.07C80.3596 15.07 80.4096 15.07 80.4563 15.07C80.5029 15.0633 80.5496 15.0567 80.5962 15.05C80.6896 15.0367 80.7529 15.0633 80.7863 15.13C80.8196 15.19 80.8363 15.3167 80.8363 15.51C80.8363 15.6767 80.8029 15.8067 80.7363 15.9C80.6696 15.9933 80.5596 16.05 80.4063 16.07C80.3396 16.0767 80.2696 16.0833 80.1963 16.09C80.1229 16.0967 80.0496 16.1 79.9763 16.1ZM83.6808 16.1C83.1741 16.1 82.7341 15.9967 82.3608 15.79C81.9874 15.5833 81.6974 15.29 81.4908 14.91C81.2841 14.5233 81.1808 14.07 81.1808 13.55C81.1808 13.1567 81.2374 12.8067 81.3508 12.5C81.4708 12.1867 81.6408 11.92 81.8608 11.7C82.0808 11.4733 82.3441 11.3033 82.6508 11.19C82.9574 11.07 83.3008 11.01 83.6808 11.01C84.1874 11.01 84.6274 11.1133 85.0008 11.32C85.3741 11.5267 85.6641 11.82 85.8708 12.2C86.0774 12.58 86.1808 13.03 86.1808 13.55C86.1808 13.9433 86.1208 14.2967 86.0008 14.61C85.8874 14.9233 85.7208 15.1933 85.5008 15.42C85.2808 15.64 85.0174 15.81 84.7108 15.93C84.4041 16.0433 84.0608 16.1 83.6808 16.1ZM83.6808 15.15C83.9274 15.15 84.1441 15.09 84.3308 14.97C84.5174 14.85 84.6608 14.6733 84.7608 14.44C84.8674 14.2 84.9208 13.9033 84.9208 13.55C84.9208 13.0167 84.8074 12.62 84.5808 12.36C84.3541 12.0933 84.0541 11.96 83.6808 11.96C83.4341 11.96 83.2174 12.02 83.0308 12.14C82.8441 12.2533 82.6974 12.43 82.5908 12.67C82.4908 12.9033 82.4408 13.1967 82.4408 13.55C82.4408 14.0767 82.5541 14.4767 82.7808 14.75C83.0074 15.0167 83.3074 15.15 83.6808 15.15ZM88.8225 16.1C88.4692 16.1 88.1525 16.0333 87.8725 15.9C87.5992 15.76 87.3825 15.5733 87.2225 15.34C87.0692 15.1067 86.9925 14.8433 86.9925 14.55C86.9925 14.19 87.0858 13.9067 87.2725 13.7C87.4592 13.4867 87.7625 13.3333 88.1825 13.24C88.6025 13.1467 89.1658 13.1 89.8725 13.1H90.3725V13.82H89.8825C89.4692 13.82 89.1392 13.84 88.8925 13.88C88.6458 13.92 88.4692 13.99 88.3625 14.09C88.2625 14.1833 88.2125 14.3167 88.2125 14.49C88.2125 14.71 88.2892 14.89 88.4425 15.03C88.5958 15.17 88.8092 15.24 89.0825 15.24C89.3025 15.24 89.4958 15.19 89.6625 15.09C89.8358 14.9833 89.9725 14.84 90.0725 14.66C90.1725 14.48 90.2225 14.2733 90.2225 14.04V12.89C90.2225 12.5567 90.1492 12.3167 90.0025 12.17C89.8558 12.0233 89.6092 11.95 89.2625 11.95C89.0692 11.95 88.8592 11.9733 88.6325 12.02C88.4125 12.0667 88.1792 12.1467 87.9325 12.26C87.8058 12.32 87.6925 12.3367 87.5925 12.31C87.4992 12.2833 87.4258 12.23 87.3725 12.15C87.3192 12.0633 87.2925 11.97 87.2925 11.87C87.2925 11.77 87.3192 11.6733 87.3725 11.58C87.4258 11.48 87.5158 11.4067 87.6425 11.36C87.9492 11.2333 88.2425 11.1433 88.5225 11.09C88.8092 11.0367 89.0692 11.01 89.3025 11.01C89.7825 11.01 90.1758 11.0833 90.4825 11.23C90.7958 11.3767 91.0292 11.6 91.1825 11.9C91.3358 12.1933 91.4125 12.5733 91.4125 13.04V15.44C91.4125 15.6467 91.3625 15.8067 91.2625 15.92C91.1625 16.0267 91.0192 16.08 90.8325 16.08C90.6458 16.08 90.4992 16.0267 90.3925 15.92C90.2925 15.8067 90.2425 15.6467 90.2425 15.44V14.96H90.3225C90.2758 15.1933 90.1825 15.3967 90.0425 15.57C89.9092 15.7367 89.7392 15.8667 89.5325 15.96C89.3258 16.0533 89.0892 16.1 88.8225 16.1ZM94.5613 16.1C94.1346 16.1 93.7579 15.9967 93.4313 15.79C93.1113 15.5833 92.8613 15.29 92.6813 14.91C92.5013 14.5233 92.4113 14.07 92.4113 13.55C92.4113 13.0233 92.5013 12.5733 92.6813 12.2C92.8613 11.82 93.1113 11.5267 93.4313 11.32C93.7579 11.1133 94.1346 11.01 94.5613 11.01C94.9679 11.01 95.3213 11.11 95.6213 11.31C95.9279 11.51 96.1346 11.7733 96.2413 12.1H96.1313V9.5C96.1313 9.29333 96.1846 9.13667 96.2913 9.03C96.3979 8.92333 96.5513 8.87 96.7513 8.87C96.9513 8.87 97.1046 8.92333 97.2113 9.03C97.3246 9.13667 97.3813 9.29333 97.3813 9.5V15.44C97.3813 15.6467 97.3279 15.8067 97.2213 15.92C97.1146 16.0267 96.9613 16.08 96.7613 16.08C96.5613 16.08 96.4079 16.0267 96.3013 15.92C96.1946 15.8067 96.1413 15.6467 96.1413 15.44V14.66L96.2513 14.96C96.1513 15.3067 95.9479 15.5833 95.6413 15.79C95.3346 15.9967 94.9746 16.1 94.5613 16.1ZM94.9113 15.15C95.1579 15.15 95.3746 15.09 95.5613 14.97C95.7479 14.85 95.8913 14.6733 95.9913 14.44C96.0979 14.2 96.1513 13.9033 96.1513 13.55C96.1513 13.0167 96.0379 12.62 95.8113 12.36C95.5846 12.0933 95.2846 11.96 94.9113 11.96C94.6646 11.96 94.4479 12.02 94.2613 12.14C94.0746 12.2533 93.9279 12.43 93.8213 12.67C93.7213 12.9033 93.6713 13.1967 93.6713 13.55C93.6713 14.0767 93.7846 14.4767 94.0113 14.75C94.2379 15.0167 94.5379 15.15 94.9113 15.15ZM103.622 16.1C103.116 16.1 102.676 15.9967 102.302 15.79C101.929 15.5833 101.639 15.29 101.432 14.91C101.226 14.5233 101.122 14.07 101.122 13.55C101.122 13.1567 101.179 12.8067 101.292 12.5C101.412 12.1867 101.582 11.92 101.802 11.7C102.022 11.4733 102.286 11.3033 102.592 11.19C102.899 11.07 103.242 11.01 103.622 11.01C104.129 11.01 104.569 11.1133 104.942 11.32C105.316 11.5267 105.606 11.82 105.812 12.2C106.019 12.58 106.122 13.03 106.122 13.55C106.122 13.9433 106.062 14.2967 105.942 14.61C105.829 14.9233 105.662 15.1933 105.442 15.42C105.222 15.64 104.959 15.81 104.652 15.93C104.346 16.0433 104.002 16.1 103.622 16.1ZM103.622 15.15C103.869 15.15 104.086 15.09 104.272 14.97C104.459 14.85 104.602 14.6733 104.702 14.44C104.809 14.2 104.862 13.9033 104.862 13.55C104.862 13.0167 104.749 12.62 104.522 12.36C104.296 12.0933 103.996 11.96 103.622 11.96C103.376 11.96 103.159 12.02 102.972 12.14C102.786 12.2533 102.639 12.43 102.532 12.67C102.432 12.9033 102.382 13.1967 102.382 13.55C102.382 14.0767 102.496 14.4767 102.722 14.75C102.949 15.0167 103.249 15.15 103.622 15.15ZM107.774 16.08C107.574 16.08 107.421 16.0267 107.314 15.92C107.207 15.8067 107.154 15.6467 107.154 15.44V11.66C107.154 11.4533 107.207 11.2967 107.314 11.19C107.421 11.0833 107.571 11.03 107.764 11.03C107.957 11.03 108.107 11.0833 108.214 11.19C108.321 11.2967 108.374 11.4533 108.374 11.66V12.34L108.264 12.09C108.411 11.7367 108.637 11.47 108.944 11.29C109.257 11.1033 109.611 11.01 110.004 11.01C110.397 11.01 110.721 11.0833 110.974 11.23C111.227 11.3767 111.417 11.6 111.544 11.9C111.671 12.1933 111.734 12.5667 111.734 13.02V15.44C111.734 15.6467 111.681 15.8067 111.574 15.92C111.467 16.0267 111.314 16.08 111.114 16.08C110.914 16.08 110.757 16.0267 110.644 15.92C110.537 15.8067 110.484 15.6467 110.484 15.44V13.08C110.484 12.7 110.411 12.4233 110.264 12.25C110.124 12.0767 109.904 11.99 109.604 11.99C109.237 11.99 108.944 12.1067 108.724 12.34C108.511 12.5667 108.404 12.87 108.404 13.25V15.44C108.404 15.8667 108.194 16.08 107.774 16.08ZM117.958 16.1C117.525 16.1 117.162 16.0267 116.868 15.88C116.582 15.7333 116.368 15.52 116.228 15.24C116.088 14.9533 116.018 14.6 116.018 14.18V12.05H115.498C115.338 12.05 115.215 12.01 115.128 11.93C115.042 11.8433 114.998 11.7267 114.998 11.58C114.998 11.4267 115.042 11.31 115.128 11.23C115.215 11.15 115.338 11.11 115.498 11.11H116.018V10.2C116.018 9.99333 116.072 9.83667 116.178 9.73C116.292 9.62333 116.448 9.57 116.648 9.57C116.848 9.57 117.002 9.62333 117.108 9.73C117.215 9.83667 117.268 9.99333 117.268 10.2V11.11H118.328C118.488 11.11 118.612 11.15 118.698 11.23C118.785 11.31 118.828 11.4267 118.828 11.58C118.828 11.7267 118.785 11.8433 118.698 11.93C118.612 12.01 118.488 12.05 118.328 12.05H117.268V14.11C117.268 14.43 117.338 14.67 117.478 14.83C117.618 14.99 117.845 15.07 118.158 15.07C118.272 15.07 118.372 15.06 118.458 15.04C118.545 15.02 118.622 15.0067 118.688 15C118.768 14.9933 118.835 15.02 118.888 15.08C118.942 15.1333 118.968 15.2467 118.968 15.42C118.968 15.5533 118.945 15.6733 118.898 15.78C118.858 15.88 118.782 15.95 118.668 15.99C118.582 16.0167 118.468 16.04 118.328 16.06C118.188 16.0867 118.065 16.1 117.958 16.1ZM120.303 16.08C120.103 16.08 119.95 16.0267 119.843 15.92C119.737 15.8067 119.683 15.6467 119.683 15.44V9.5C119.683 9.29333 119.737 9.13667 119.843 9.03C119.95 8.92333 120.103 8.87 120.303 8.87C120.503 8.87 120.657 8.92333 120.763 9.03C120.877 9.13667 120.933 9.29333 120.933 9.5V12.09H120.793C120.94 11.7367 121.167 11.47 121.473 11.29C121.787 11.1033 122.14 11.01 122.533 11.01C122.927 11.01 123.25 11.0833 123.503 11.23C123.757 11.3767 123.947 11.6 124.073 11.9C124.2 12.1933 124.263 12.5667 124.263 13.02V15.44C124.263 15.6467 124.21 15.8067 124.103 15.92C123.997 16.0267 123.843 16.08 123.643 16.08C123.443 16.08 123.287 16.0267 123.173 15.92C123.067 15.8067 123.013 15.6467 123.013 15.44V13.08C123.013 12.7 122.94 12.4233 122.793 12.25C122.653 12.0767 122.433 11.99 122.133 11.99C121.767 11.99 121.473 12.1067 121.253 12.34C121.04 12.5667 120.933 12.87 120.933 13.25V15.44C120.933 15.8667 120.723 16.08 120.303 16.08ZM127.943 16.1C127.389 16.1 126.913 15.9967 126.513 15.79C126.113 15.5833 125.803 15.29 125.583 14.91C125.369 14.53 125.263 14.08 125.263 13.56C125.263 13.0533 125.366 12.61 125.573 12.23C125.786 11.85 126.076 11.5533 126.443 11.34C126.816 11.12 127.239 11.01 127.713 11.01C128.059 11.01 128.369 11.0667 128.643 11.18C128.923 11.2933 129.159 11.4567 129.353 11.67C129.553 11.8833 129.703 12.1433 129.803 12.45C129.909 12.75 129.963 13.09 129.963 13.47C129.963 13.59 129.919 13.6833 129.833 13.75C129.753 13.81 129.636 13.84 129.483 13.84H126.273V13.12H129.083L128.923 13.27C128.923 12.9633 128.876 12.7067 128.783 12.5C128.696 12.2933 128.566 12.1367 128.393 12.03C128.226 11.9167 128.016 11.86 127.763 11.86C127.483 11.86 127.243 11.9267 127.043 12.06C126.849 12.1867 126.699 12.37 126.593 12.61C126.493 12.8433 126.443 13.1233 126.443 13.45V13.52C126.443 14.0667 126.569 14.4767 126.823 14.75C127.083 15.0167 127.463 15.15 127.963 15.15C128.136 15.15 128.329 15.13 128.543 15.09C128.763 15.0433 128.969 14.9667 129.163 14.86C129.303 14.78 129.426 14.7467 129.533 14.76C129.639 14.7667 129.723 14.8033 129.783 14.87C129.849 14.9367 129.889 15.02 129.903 15.12C129.916 15.2133 129.896 15.31 129.843 15.41C129.796 15.51 129.713 15.5967 129.593 15.67C129.359 15.8167 129.089 15.9267 128.783 16C128.483 16.0667 128.203 16.1 127.943 16.1Z" fill="white"/>
+    <path d="M35.0457 22.8903C35.0123 19.3528 38.0969 17.6319 38.2377 17.552C36.4907 15.1339 33.7822 14.8038 32.8309 14.7774C30.5568 14.55 28.3504 16.069 27.1925 16.069C26.0106 16.069 24.2265 14.7994 22.3036 14.8363C19.8285 14.8731 17.5146 16.2323 16.2447 18.344C13.6242 22.6498 15.5777 28.9768 18.088 32.4572C19.344 34.1615 20.8113 36.065 22.7324 35.9983C24.611 35.9245 25.3131 34.8613 27.5807 34.8613C29.8269 34.8613 30.4864 35.9983 32.4455 35.9553C34.4621 35.9245 35.7311 34.2431 36.9427 32.5231C38.3942 30.5695 38.9768 28.6458 39 28.5465C38.9528 28.5316 35.0836 27.1303 35.0447 22.8903H35.0457ZM31.3451 12.4884C32.3557 11.2882 33.0476 9.65682 32.855 8C31.3933 8.06146 29.5638 8.96055 28.5106 10.1336C27.5779 11.1688 26.7461 12.8634 26.96 14.4561C28.6033 14.572 30.2901 13.6694 31.3442 12.4884H31.3451Z" fill="white"/>
+  </svg>
+`;
