+++ js/ui/icons/IconApple/index.js
@@ -1,0 +1,1 @@
+export { IconApple } from './IconApple.js';
