+++ js/ui/icons/IconArrowDown/IconArrowDown.js
@@ -1,0 +1,12 @@
+/* eslint-disable */
+
+/**
+ * @function IconArrowDown
+ * @returns {string} HTML
+ */
+
+export const IconArrowDown = () => `
+  <svg viewBox="0 0 9 6" xmlns="http://www.w3.org/2000/svg">
+    <path d="M4.5 6L0 1.49973L1.50071 0L4.5 3.00053L7.49929 0L9 1.49973L4.5 6Z" fill="currentColor" />
+  </svg>
+`;
