+++ js/ui/icons/IconArrowDown/index.js
@@ -1,0 +1,1 @@
+export { IconArrowDown } from './IconArrowDown.js';
