+++ js/ui/icons/IconClose/IconClose.js
@@ -1,0 +1,13 @@
+/* eslint-disable */
+
+/**
+ * @function IconClose
+ * @returns {string} HTML
+ */
+
+export const IconClose = () => `
+  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <rect width="31.8192" height="2.12128" transform="matrix(0.707107 -0.707106 0.707107 0.707106 0.000488281 22.4996)" fill="white"/>
+    <rect width="31.8192" height="2.12128" transform="matrix(0.707107 0.707106 -0.707107 0.707106 1.5 0.000488281)" fill="white"/>
+  </svg>
+`;
