+++ js/ui/icons/IconClose/index.js
@@ -1,0 +1,1 @@
+export { IconClose } from './IconClose.js';
