+++ js/ui/icons/IconDeveloper/IconDeveloper.js
@@ -1,0 +1,12 @@
+/* eslint-disable */
+
+/**
+ * @function IconDeveloper
+ * @returns {string} HTML
+ */
+
+export const IconDeveloper = () => `
+  <svg width="130" height="46" viewBox="0 0 130 46" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M53.0156 0L65.3856 29.3943L14.7011 46L0 3.19635L53.0156 0ZM33.2519 15.3286C29.9532 17.0524 28.2551 20.0227 28.5202 24.0077C28.7802 27.9273 30.8366 30.439 34.2565 31.7642C37.0116 32.8309 39.8391 32.9759 42.7057 32.541C43.4813 32.4228 44.2414 32.1983 44.9704 32.0045L45.404 31.8926V27.6555C43.4111 27.8587 41.6007 28.2661 39.8118 28.1747C37.2125 28.0423 35.5026 26.0947 35.3733 23.5925C35.1864 19.9604 37.7584 17.7489 41.5335 18.3816C42.5344 18.5486 43.5086 18.8976 44.5383 19.2112L44.983 19.342V14.5998C40.9256 13.3519 36.9902 13.3747 33.2519 15.3286ZM76.1645 25.7898C76.4969 25.7898 76.8145 25.8387 77.1196 25.9364C77.4224 26.034 77.6898 26.2326 77.9188 26.532L77.2488 27.1835C77.1454 27.0126 76.9992 26.8747 76.8101 26.7707C76.6202 26.6667 76.4053 26.6147 76.1645 26.6147C75.9355 26.6147 75.7449 26.6462 75.5935 26.7108C75.4414 26.7754 75.3195 26.8582 75.2279 26.959C75.1363 27.0598 75.0713 27.1686 75.0344 27.2844C74.9967 27.401 74.979 27.5081 74.979 27.6058C74.979 27.8808 75.0521 28.0919 75.202 28.2385C75.3513 28.385 75.5359 28.5048 75.756 28.5962C75.9769 28.6884 76.2177 28.7695 76.4792 28.8436C76.7399 28.9176 76.9807 29.02 77.2008 29.1508C77.4217 29.2832 77.607 29.4605 77.7548 29.6834C77.9055 29.9064 77.9793 30.2136 77.9793 30.6052C77.9793 30.9172 77.9232 31.1985 77.8124 31.4482C77.6994 31.6996 77.5494 31.9107 77.3604 32.0817C77.1713 32.2534 76.9482 32.385 76.6897 32.4756C76.4311 32.5686 76.1556 32.6143 75.8646 32.6143C75.4731 32.6143 75.1038 32.5402 74.7552 32.3937C74.4043 32.2471 74.1266 32.021 73.9205 31.7153L74.6001 31.0913C74.7493 31.33 74.9398 31.5057 75.1718 31.6184C75.4044 31.7319 75.6438 31.7886 75.8897 31.7886C76.0278 31.7886 76.1674 31.7649 76.3107 31.7192C76.4548 31.6735 76.584 31.605 76.7022 31.5136C76.8197 31.4214 76.9164 31.308 76.9903 31.174C77.0642 31.0393 77.1026 30.8825 77.1026 30.7061C77.1026 30.4555 77.0272 30.2641 76.8788 30.1325C76.7303 30.0017 76.5449 29.8914 76.324 29.8016C76.1025 29.7141 75.8631 29.6283 75.6017 29.5455C75.3409 29.4636 75.0994 29.3486 74.8793 29.202C74.6577 29.0555 74.473 28.8577 74.3253 28.6104C74.1754 28.3622 74.1015 28.0273 74.1015 27.6058C74.1015 27.416 74.1399 27.2143 74.2167 27.0008C74.295 26.7865 74.4161 26.5911 74.5831 26.413C74.7493 26.2357 74.9627 26.0876 75.2235 25.9687C75.4835 25.8489 75.7981 25.7898 76.1645 25.7898ZM120.058 27.9911C120.419 27.9911 120.724 28.0588 120.973 28.1928C121.223 28.3267 121.427 28.5032 121.588 28.7199C121.749 28.9373 121.865 29.1823 121.937 29.4541C122.008 29.726 122.044 30.0033 122.044 30.2846V30.5776H118.776C118.776 30.7667 118.815 30.9401 118.893 31.0961C118.969 31.252 119.072 31.3852 119.197 31.4947C119.325 31.605 119.47 31.6909 119.635 31.7523C119.802 31.813 119.975 31.8437 120.153 31.8437C120.393 31.8437 120.602 31.7839 120.78 31.6649C120.957 31.5451 121.121 31.3883 121.27 31.1922L121.854 31.6696C121.424 32.2629 120.823 32.5591 120.049 32.5591C119.728 32.5591 119.436 32.5008 119.176 32.385C118.915 32.2684 118.694 32.1085 118.514 31.9036C118.333 31.6988 118.194 31.4569 118.096 31.1788C118 30.9007 117.951 30.5997 117.951 30.2751C117.951 29.9513 118.004 29.6503 118.109 29.3714C118.217 29.0933 118.362 28.8514 118.548 28.6466C118.734 28.4417 118.957 28.2818 119.214 28.1652C119.473 28.0494 119.754 27.9911 120.058 27.9911ZM110.977 27.9911C111.338 27.9911 111.643 28.0588 111.892 28.1928C112.142 28.3267 112.347 28.5032 112.508 28.7199C112.668 28.9373 112.784 29.1823 112.855 29.4541C112.927 29.726 112.963 30.0033 112.963 30.2846V30.5776H109.695C109.695 30.7667 109.733 30.9401 109.812 31.0961C109.888 31.252 109.99 31.3852 110.117 31.4947C110.242 31.605 110.388 31.6909 110.555 31.7523C110.722 31.813 110.893 31.8437 111.071 31.8437C111.312 31.8437 111.521 31.7839 111.699 31.6649C111.876 31.5451 112.04 31.3883 112.189 31.1922L112.774 31.6696C112.344 32.2629 111.742 32.5591 110.968 32.5591C110.646 32.5591 110.355 32.5008 110.095 32.385C109.834 32.2684 109.614 32.1085 109.433 31.9036C109.252 31.6988 109.113 31.4569 109.015 31.1788C108.919 30.9007 108.869 30.5997 108.869 30.2751C108.869 29.9513 108.922 29.6503 109.029 29.3714C109.135 29.0933 109.28 28.8514 109.467 28.6466C109.654 28.4417 109.876 28.2818 110.134 28.1652C110.391 28.0494 110.673 27.9911 110.977 27.9911ZM99.3407 27.9911C99.7019 27.9911 100.007 28.0588 100.257 28.1928C100.506 28.3267 100.711 28.5032 100.871 28.7199C101.032 28.9373 101.148 29.1823 101.221 29.4541C101.291 29.726 101.327 30.0033 101.327 30.2846V30.5776H98.0592C98.0592 30.7667 98.0983 30.9401 98.1759 31.0961C98.2527 31.252 98.3546 31.3852 98.4809 31.4947C98.6065 31.605 98.7535 31.6909 98.9189 31.7523C99.0851 31.813 99.2572 31.8437 99.436 31.8437C99.6768 31.8437 99.8858 31.7839 100.063 31.6649C100.24 31.5451 100.403 31.3883 100.554 31.1922L101.137 31.6696C100.708 32.2629 100.107 32.5591 99.3318 32.5591C99.0105 32.5591 98.7202 32.5008 98.4588 32.385C98.198 32.2684 97.9772 32.1085 97.7969 31.9036C97.6167 31.6988 97.4771 31.4569 97.3796 31.1788C97.2829 30.9007 97.2341 30.5997 97.2341 30.2751C97.2341 29.9513 97.2858 29.6503 97.3929 29.3714C97.4993 29.0933 97.6455 28.8514 97.8317 28.6466C98.0178 28.4417 98.2394 28.2818 98.4979 28.1652C98.7564 28.0494 99.0379 27.9911 99.3407 27.9911ZM95.2649 27.9998C95.5626 27.9998 95.8484 28.0565 96.1202 28.1699C96.3935 28.2826 96.6299 28.4551 96.8301 28.6876L96.2451 29.2564C96.1018 29.0972 95.9518 28.9767 95.7945 28.894C95.6357 28.812 95.4481 28.7703 95.2302 28.7703C95.0182 28.7703 94.8335 28.812 94.6762 28.894C94.5174 28.9767 94.3859 29.0909 94.2796 29.2335C94.1739 29.3769 94.0934 29.5424 94.0395 29.7299C93.9841 29.915 93.9583 30.1096 93.9583 30.3113C93.9583 30.513 93.9885 30.7053 94.0513 30.8849C94.1148 31.0661 94.2035 31.2237 94.3187 31.3576C94.4332 31.4916 94.5706 31.5971 94.7309 31.6736C94.8919 31.75 95.0728 31.7878 95.2738 31.7878C95.4909 31.7878 95.6771 31.7468 95.8314 31.6641C95.9865 31.5822 96.1306 31.4608 96.2613 31.3017L96.8131 31.8886C96.6122 32.1274 96.3788 32.2991 96.1114 32.4031C95.8455 32.5063 95.5626 32.5583 95.2649 32.5583C94.9488 32.5583 94.6607 32.504 94.4007 32.3937C94.1392 32.2834 93.9147 32.1289 93.7256 31.9304C93.5365 31.7319 93.3895 31.4931 93.2868 31.215C93.1834 30.9369 93.1317 30.6296 93.1317 30.2932C93.1317 29.9568 93.1834 29.648 93.2868 29.3667C93.3902 29.0854 93.535 28.8436 93.7211 28.6419C93.9073 28.4402 94.1311 28.2826 94.3918 28.1699C94.6526 28.0565 94.9428 27.9998 95.2649 27.9998ZM80.8689 27.9911C81.2293 27.9911 81.5351 28.0588 81.7848 28.1928C82.0344 28.3267 82.239 28.5032 82.3993 28.7199C82.5596 28.9373 82.6763 29.1823 82.748 29.4541C82.8189 29.726 82.8551 30.0033 82.8551 30.2846V30.5776H79.5873C79.5873 30.7667 79.6257 30.9401 79.7033 31.0961C79.7809 31.252 79.8821 31.3852 80.0091 31.4947C80.1347 31.605 80.2809 31.6909 80.4471 31.7523C80.6133 31.813 80.7847 31.8437 80.9642 31.8437C81.2042 31.8437 81.4132 31.7839 81.592 31.6649C81.7693 31.5451 81.9318 31.3883 82.081 31.1922L82.666 31.6696C82.2361 32.2629 81.6341 32.5591 80.86 32.5591C80.538 32.5591 80.2484 32.5008 79.9869 32.385C79.7262 32.2684 79.5053 32.1085 79.3251 31.9036C79.1449 31.6988 79.0046 31.4569 78.9078 31.1788C78.8103 30.9007 78.7615 30.5997 78.7615 30.2751C78.7615 29.9513 78.814 29.6503 78.9211 29.3714C79.0267 29.0933 79.173 28.8514 79.3598 28.6466C79.546 28.4417 79.7676 28.2818 80.0261 28.1652C80.2846 28.0494 80.5653 27.9911 80.8689 27.9911ZM11.6934 14.3832L10.4694 14.3863V32.5024H16.8697V27.1197L18.0294 27.074C19.5251 27.011 20.8163 26.9188 21.8924 26.5674C23.448 26.0585 24.5057 25.4101 25.6787 24.2858C28.3326 21.7394 27.6752 17.3345 24.5235 15.5563C23.496 14.9772 22.2972 14.4848 21.159 14.4383C19.2178 14.3588 17.2738 14.3517 15.2964 14.3611L11.6934 14.3832ZM114.297 28.1006L115.518 31.4301L116.688 28.1006H117.513L115.913 32.4496H115.062L113.402 28.1006H114.297ZM87.3711 28.1006L88.5921 31.4301L89.7621 28.1006H90.5871L88.9872 32.4496H88.1363L86.4766 28.1006H87.3711ZM85.7978 27.9903C85.9234 27.9903 86.0386 28.0084 86.1405 28.0455L86.1065 28.9357C86.0438 28.9168 85.981 28.9019 85.9174 28.8893C85.8547 28.8774 85.7919 28.8711 85.7284 28.8711C85.3502 28.8711 85.0606 28.9846 84.8597 29.2107C84.6603 29.4368 84.5584 29.7882 84.5584 30.2664V32.4488H83.785V28.1006H84.5584V28.7703H84.5753C84.6278 28.6545 84.6965 28.5481 84.7822 28.4536C84.8686 28.359 84.9646 28.2779 85.0702 28.2109C85.1766 28.1431 85.2926 28.0896 85.4189 28.0502C85.5452 28.0108 85.6707 27.9903 85.7978 27.9903ZM123.807 25.5141V32.4488H123.034V25.5141H123.807ZM105.671 25.9545V31.6231H108.466V32.4488H104.845V25.9545H105.671ZM92.1006 28.1006V32.4488H91.3265V28.1006H92.1006ZM129.991 25.9545V32.4488H129.165V26.9637L128.047 28.0644L127.566 27.4498L129.261 25.9545H129.991ZM54.7662 14.2674H48.2669V32.3803H54.7662V14.2674ZM80.8511 28.6513C80.662 28.6513 80.4892 28.6899 80.3304 28.7663C80.1731 28.8428 80.0401 28.942 79.9315 29.0642C79.8222 29.1871 79.7373 29.3226 79.6767 29.4731C79.6176 29.622 79.5881 29.7701 79.5881 29.9174H82.03C82.0241 29.7338 81.9968 29.566 81.9488 29.4132C81.8985 29.2603 81.8261 29.1272 81.7286 29.0137C81.6311 28.9003 81.5093 28.812 81.3638 28.7482C81.216 28.6836 81.0461 28.6513 80.8511 28.6513ZM120.04 28.6513C119.85 28.6513 119.678 28.6899 119.52 28.7663C119.362 28.8428 119.229 28.942 119.121 29.0642C119.011 29.1871 118.926 29.3226 118.867 29.4731C118.806 29.622 118.776 29.7701 118.776 29.9174H121.218C121.212 29.7338 121.185 29.566 121.137 29.4132C121.088 29.2603 121.015 29.1272 120.918 29.0137C120.82 28.9003 120.698 28.812 120.552 28.7482C120.406 28.6836 120.235 28.6513 120.04 28.6513ZM110.959 28.6513C110.77 28.6513 110.596 28.6899 110.439 28.7663C110.281 28.8428 110.148 28.942 110.039 29.0642C109.93 29.1871 109.846 29.3226 109.785 29.4731C109.724 29.622 109.695 29.7701 109.695 29.9174H112.138C112.132 29.7338 112.104 29.566 112.056 29.4132C112.007 29.2603 111.934 29.1272 111.837 29.0137C111.739 28.9003 111.618 28.812 111.471 28.7482C111.324 28.6836 111.154 28.6513 110.959 28.6513ZM99.323 28.6513C99.1339 28.6513 98.961 28.6899 98.803 28.7663C98.6449 28.8428 98.5127 28.942 98.4034 29.0642C98.2948 29.1871 98.2091 29.3226 98.1493 29.4731C98.0894 29.622 98.0592 29.7701 98.0592 29.9174H100.502C100.496 29.7338 100.469 29.566 100.42 29.4132C100.371 29.2603 100.299 29.1272 100.2 29.0137C100.104 28.9003 99.9811 28.812 99.8356 28.7482C99.6879 28.6836 99.518 28.6513 99.323 28.6513ZM91.7135 25.9364C91.8672 25.9364 92.0009 25.9962 92.1131 26.1152C92.2247 26.2342 92.2801 26.376 92.2801 26.5414C92.2801 26.7061 92.2247 26.8487 92.1131 26.9677C92.0009 27.0874 91.8672 27.1473 91.7135 27.1473C91.5584 27.1473 91.4247 27.0874 91.3132 26.9677C91.2009 26.8487 91.1455 26.7061 91.1455 26.5414C91.1455 26.376 91.2009 26.2342 91.3132 26.1152C91.4247 25.9962 91.5584 25.9364 91.7135 25.9364ZM20.4787 19.1552C20.8776 19.743 20.9256 21.013 20.5969 21.6701C19.9646 22.9362 18.7939 23.0142 17.5301 23.0008L16.9185 22.9921V18.213C18.3071 17.9293 19.661 17.9506 20.4787 19.1552ZM87.3186 14.0869C87.651 14.0869 87.9694 14.1358 88.273 14.2335C88.5765 14.3312 88.8439 14.5297 89.0729 14.8291L88.4022 15.4807C88.2988 15.3097 88.1526 15.1718 87.9635 15.0678C87.7744 14.9638 87.5594 14.9118 87.3186 14.9118C87.0897 14.9118 86.8991 14.9441 86.7469 15.0079C86.5948 15.0726 86.4736 15.1553 86.3813 15.2561C86.2897 15.357 86.2262 15.4657 86.1878 15.5815C86.1509 15.6981 86.1317 15.8053 86.1317 15.903C86.1317 16.1787 86.2063 16.3891 86.3547 16.5356C86.5047 16.6821 86.6893 16.8019 86.9102 16.8933C87.1303 16.9855 87.3711 17.0674 87.6326 17.1407C87.8926 17.2147 88.1348 17.3172 88.3542 17.4487C88.5758 17.5803 88.7605 17.7576 88.9089 17.9805C89.0589 18.2035 89.1327 18.5108 89.1327 18.9023C89.1327 19.2143 89.0773 19.4956 88.9651 19.7461C88.8543 19.9967 88.7029 20.2078 88.5145 20.3788C88.3239 20.5505 88.1009 20.6821 87.8431 20.7735C87.5846 20.8657 87.3098 20.9114 87.0173 20.9114C86.628 20.9114 86.258 20.8373 85.9079 20.6908C85.5585 20.5442 85.28 20.3181 85.0739 20.0132L85.7535 19.3884C85.9019 19.6272 86.0932 19.8029 86.3244 19.9163C86.5571 20.029 86.7972 20.0857 87.0431 20.0857C87.1813 20.0857 87.3216 20.0629 87.4642 20.0164C87.6082 19.9707 87.7389 19.9021 87.8556 19.8115C87.9738 19.7186 88.0698 19.6051 88.1437 19.4712C88.219 19.3372 88.256 19.1805 88.256 19.0032C88.256 18.7526 88.1821 18.562 88.0322 18.4296C87.883 18.2988 87.6983 18.1885 87.4782 18.0995C87.2566 18.0113 87.0158 17.9262 86.7543 17.8427C86.4943 17.7607 86.2535 17.6457 86.0327 17.4992C85.8118 17.3526 85.6272 17.1549 85.4787 16.9083C85.3295 16.6593 85.2542 16.3252 85.2542 15.903C85.2542 15.7139 85.2933 15.5114 85.3701 15.2979C85.4477 15.0836 85.5703 14.8882 85.7365 14.7101C85.9019 14.5329 86.1154 14.3848 86.3769 14.2658C86.6376 14.146 86.9515 14.0869 87.3186 14.0869ZM82.5374 14.0869C82.8698 14.0869 83.1874 14.1358 83.491 14.2335C83.7953 14.3312 84.062 14.5297 84.2917 14.8291L83.6218 15.4807C83.5176 15.3097 83.3721 15.1718 83.183 15.0678C82.9924 14.9638 82.7775 14.9118 82.5374 14.9118C82.3085 14.9118 82.1172 14.9441 81.9657 15.0079C81.8136 15.0726 81.6917 15.1553 81.6009 15.2561C81.5085 15.357 81.4443 15.4657 81.4073 15.5815C81.3697 15.6981 81.3512 15.8053 81.3512 15.903C81.3512 16.1787 81.4251 16.3891 81.5743 16.5356C81.7235 16.6821 81.9081 16.8019 82.1283 16.8933C82.3491 16.9855 82.5899 17.0674 82.8514 17.1407C83.1114 17.2147 83.3536 17.3172 83.573 17.4487C83.7946 17.5803 83.9793 17.7576 84.127 17.9805C84.2784 18.2035 84.3508 18.5108 84.3508 18.9023C84.3508 19.2143 84.2961 19.4956 84.1846 19.7461C84.0716 19.9967 83.9217 20.2078 83.7326 20.3788C83.5427 20.5505 83.3197 20.6821 83.0611 20.7735C82.8034 20.8657 82.5286 20.9114 82.2361 20.9114C81.8468 20.9114 81.4768 20.8373 81.1274 20.6908C80.7765 20.5442 80.4996 20.3181 80.2927 20.0132L80.9723 19.3884C81.1215 19.6272 81.3121 19.8029 81.5433 19.9163C81.7767 20.029 82.016 20.0857 82.2627 20.0857C82.4001 20.0857 82.5404 20.0629 82.6837 20.0164C82.827 19.9707 82.957 19.9021 83.0752 19.8115C83.1919 19.7186 83.2879 19.6051 83.3632 19.4712C83.4371 19.3372 83.474 19.1805 83.474 19.0032C83.474 18.7526 83.4002 18.562 83.251 18.4296C83.1018 18.2988 82.9171 18.1885 82.697 18.0995C82.4754 18.0113 82.2346 17.9262 81.9731 17.8427C81.7124 17.7607 81.4723 17.6457 81.2507 17.4992C81.0299 17.3526 80.8452 17.1549 80.6968 16.9083C80.5476 16.6593 80.473 16.3252 80.473 15.903C80.473 15.7139 80.5121 15.5114 80.5889 15.2979C80.6665 15.0836 80.7884 14.8882 80.9553 14.7101C81.1215 14.5329 81.335 14.3848 81.5964 14.2658C81.8564 14.146 82.1703 14.0869 82.5374 14.0869ZM95.5397 14.0869C95.9415 14.0869 96.3352 14.1697 96.7222 14.3351C97.1093 14.4998 97.4261 14.7527 97.6728 15.0962L96.9844 15.6556C96.8131 15.4168 96.5952 15.2341 96.3307 15.1056C96.0671 14.9772 95.7945 14.9126 95.5145 14.9126C95.1696 14.9126 94.8587 14.9819 94.5809 15.119C94.3025 15.2569 94.0646 15.4444 93.8667 15.6831C93.6694 15.9219 93.5165 16.2016 93.4065 16.523C93.2986 16.8437 93.244 17.1872 93.244 17.5543C93.244 17.897 93.2957 18.2208 93.3991 18.5273C93.5025 18.833 93.6502 19.1017 93.8452 19.3341C94.0402 19.5665 94.2781 19.7501 94.5595 19.8848C94.8409 20.0187 95.1578 20.0865 95.5145 20.0865C95.8632 20.0865 96.1705 20.0101 96.4342 19.8572C96.6971 19.7044 96.9217 19.4901 97.1048 19.2151L97.8006 19.7745L97.7297 19.869C97.6861 19.9234 97.63 19.988 97.5606 20.0636C97.4461 20.1889 97.2947 20.3142 97.1048 20.4395C96.915 20.5647 96.6875 20.675 96.4209 20.7696C96.155 20.8641 95.8462 20.9114 95.4976 20.9114C95.0152 20.9114 94.5817 20.8137 94.1946 20.6183C93.8068 20.4229 93.4781 20.1661 93.2048 19.8478C92.9323 19.5295 92.7247 19.1702 92.5814 18.77C92.4381 18.3697 92.3672 17.964 92.3672 17.5543C92.3672 17.0532 92.444 16.59 92.5992 16.1645C92.7528 15.7391 92.9692 15.3727 93.2477 15.0639C93.5261 14.755 93.8607 14.5155 94.25 14.3438C94.6393 14.1728 95.0699 14.0869 95.5397 14.0869ZM107.64 15.1781V16.3993H108.749V17.1139H107.64V19.0859C107.64 19.208 107.643 19.3301 107.649 19.4483C107.654 19.5673 107.675 19.6744 107.713 19.7698C107.749 19.8643 107.808 19.9407 107.885 19.999C107.962 20.0566 108.076 20.0857 108.224 20.0857C108.316 20.0857 108.41 20.0763 108.509 20.0581C108.605 20.04 108.695 20.0061 108.775 19.9573V20.7097C108.683 20.7648 108.564 20.8027 108.418 20.8239C108.272 20.8452 108.158 20.8562 108.079 20.8562C107.78 20.8562 107.549 20.8121 107.386 20.7231C107.223 20.6348 107.102 20.5198 107.025 20.3796C106.948 20.2386 106.901 20.081 106.887 19.9069C106.873 19.7335 106.865 19.5571 106.865 19.379V17.1139H105.971V16.3985H106.865V15.1781H107.64ZM125.604 13.812V20.7459H124.83V20.1235H124.813C124.663 20.3552 124.458 20.5356 124.198 20.6632C123.937 20.7916 123.669 20.8562 123.394 20.8562C123.078 20.8562 122.794 20.7979 122.538 20.6829C122.283 20.5655 122.065 20.404 121.885 20.196C121.704 19.988 121.565 19.7469 121.468 19.4712C121.37 19.1962 121.321 18.8968 121.321 18.573C121.321 18.2484 121.37 17.9475 121.468 17.6686C121.565 17.3904 121.704 17.1486 121.885 16.9445C122.065 16.7397 122.283 16.5789 122.538 16.4623C122.794 16.3465 123.078 16.289 123.394 16.289C123.687 16.289 123.962 16.3544 124.219 16.486C124.477 16.6168 124.674 16.7956 124.813 17.0217H124.83V13.812H125.604ZM118.604 16.2882C118.965 16.2882 119.271 16.356 119.52 16.4899C119.77 16.6246 119.974 16.8003 120.135 17.0178C120.295 17.2344 120.412 17.4795 120.483 17.7513C120.554 18.0231 120.591 18.3004 120.591 18.5817V18.8748H117.323C117.323 19.0646 117.361 19.2372 117.439 19.3932C117.516 19.5492 117.618 19.6823 117.745 19.7918C117.87 19.9021 118.016 19.9888 118.183 20.0495C118.349 20.1101 118.52 20.1409 118.699 20.1409C118.94 20.1409 119.149 20.081 119.327 19.962C119.504 19.8423 119.667 19.6863 119.816 19.4893L120.401 19.9667C119.972 20.56 119.37 20.857 118.595 20.857C118.273 20.857 117.984 20.7979 117.722 20.6821C117.462 20.5655 117.241 20.4064 117.061 20.2015C116.88 19.9967 116.74 19.754 116.643 19.4767C116.546 19.1978 116.497 18.8968 116.497 18.5722C116.497 18.2492 116.549 17.9475 116.657 17.6686C116.762 17.3904 116.908 17.1486 117.095 16.9445C117.281 16.7397 117.503 16.5789 117.762 16.4623C118.019 16.3465 118.3 16.2882 118.604 16.2882ZM100.467 16.2882C100.829 16.2882 101.133 16.356 101.383 16.4899C101.633 16.6246 101.837 16.8003 101.998 17.0178C102.159 17.2344 102.275 17.4795 102.347 17.7513C102.417 18.0231 102.453 18.3004 102.453 18.5817V18.8748H99.1856C99.1856 19.0646 99.2247 19.2372 99.3023 19.3932C99.3791 19.5492 99.481 19.6823 99.6073 19.7918C99.7336 19.9021 99.8806 19.9888 100.045 20.0495C100.212 20.1101 100.384 20.1409 100.562 20.1409C100.803 20.1409 101.013 20.081 101.19 19.962C101.368 19.8423 101.53 19.6863 101.68 19.4893L102.264 19.9667C101.834 20.56 101.233 20.857 100.458 20.857C100.137 20.857 99.8467 20.7979 99.5852 20.6821C99.3244 20.5655 99.1036 20.4064 98.9234 20.2015C98.7431 19.9967 98.6035 19.754 98.506 19.4767C98.41 19.1978 98.3605 18.8968 98.3605 18.5722C98.3605 18.2492 98.413 17.9475 98.5193 17.6686C98.6257 17.3904 98.7719 17.1486 98.9581 16.9445C99.1442 16.7397 99.3658 16.5789 99.6243 16.4623C99.8829 16.3465 100.164 16.2882 100.467 16.2882ZM76.4053 14.2524C76.8352 14.2524 77.2126 14.3044 77.5361 14.4084C77.8611 14.5124 78.1404 14.651 78.3789 14.826C78.6168 14.9993 78.8147 15.2002 78.9721 15.4263C79.1301 15.6532 79.2542 15.8848 79.3465 16.1236C79.4381 16.3615 79.5039 16.6002 79.5445 16.8389C79.5844 17.0777 79.6051 17.2975 79.6051 17.4992C79.6051 17.9152 79.5327 18.317 79.3909 18.7062C79.2461 19.0938 79.0334 19.4389 78.7534 19.7422C78.4727 20.0447 78.1241 20.2882 77.7075 20.4718C77.2924 20.6553 76.8123 20.7467 76.2679 20.7467H74.2817V14.2524H76.4053ZM113.505 13.7017C113.591 13.7017 113.681 13.7048 113.776 13.7103C113.87 13.7166 113.972 13.7348 114.081 13.7655L113.995 14.4998C113.92 14.4691 113.85 14.4478 113.78 14.4352C113.711 14.4226 113.636 14.4171 113.556 14.4171C113.413 14.4171 113.298 14.4407 113.212 14.4856C113.127 14.5329 113.059 14.5975 113.011 14.6833C112.961 14.7684 112.929 14.8709 112.916 14.9906C112.901 15.1096 112.894 15.2459 112.894 15.3987V16.3993H113.815V17.1139H112.886V20.7459H112.112V17.1139H111.235V16.3985H112.112V15.4263C112.112 14.8456 112.217 14.4131 112.426 14.1287C112.635 13.8451 112.994 13.7032 113.504 13.7032L113.505 13.7017ZM115.509 16.3985V20.7467H114.734V16.3985H115.509ZM110.581 16.3985V20.7467H109.807V16.3985H110.581ZM105.395 16.2882C105.521 16.2882 105.636 16.3063 105.739 16.3434L105.705 17.2337C105.642 17.2147 105.579 17.1998 105.516 17.1864C105.453 17.1754 105.39 17.169 105.326 17.169C104.948 17.169 104.658 17.2825 104.457 17.5078C104.258 17.7347 104.157 18.0861 104.157 18.5636V20.7459H103.383V16.3985H104.157V17.0682H104.174C104.225 16.9524 104.294 16.846 104.38 16.7515C104.466 16.6569 104.563 16.5758 104.668 16.5088C104.775 16.4411 104.89 16.3875 105.017 16.3481C105.143 16.3079 105.269 16.2882 105.395 16.2882ZM123.489 17.0587C123.282 17.0587 123.096 17.0981 122.929 17.1738C122.764 17.2502 122.623 17.355 122.508 17.485C122.394 17.6166 122.304 17.7765 122.241 17.9624C122.179 18.1491 122.148 18.3524 122.148 18.5722C122.148 18.7928 122.179 18.9961 122.241 19.1828C122.304 19.3688 122.394 19.5279 122.508 19.6587C122.623 19.7903 122.764 19.895 122.929 19.9707C123.096 20.0479 123.282 20.0857 123.489 20.0857C123.695 20.0857 123.881 20.0479 124.047 19.9707C124.213 19.895 124.355 19.7903 124.468 19.6587C124.584 19.5279 124.672 19.3688 124.735 19.1828C124.799 18.9961 124.83 18.792 124.83 18.5722C124.83 18.3524 124.799 18.1491 124.735 17.9624C124.672 17.7757 124.584 17.6166 124.468 17.485C124.355 17.3542 124.213 17.2502 124.047 17.1738C123.881 17.0974 123.695 17.0587 123.489 17.0587ZM76.3455 15.0781H75.1068V19.921H76.1482C76.4976 19.921 76.8278 19.8706 77.141 19.7698C77.4534 19.6689 77.7267 19.5177 77.9616 19.3152C78.198 19.1143 78.3834 18.8614 78.5215 18.5588C78.6589 18.2563 78.7283 17.9025 78.7283 17.4992C78.7283 17.292 78.6958 17.0485 78.6323 16.7704C78.5688 16.4915 78.4506 16.2268 78.2763 15.977C78.1012 15.7249 77.8597 15.5138 77.5494 15.3388C77.24 15.1647 76.8381 15.0781 76.3455 15.0781ZM118.587 16.9484C118.398 16.9484 118.224 16.9878 118.066 17.0635C117.909 17.1399 117.776 17.2392 117.666 17.3613C117.558 17.4842 117.473 17.6197 117.412 17.771C117.352 17.9191 117.323 18.0672 117.323 18.2145H119.765C119.76 18.031 119.732 17.8632 119.683 17.7103C119.634 17.5575 119.562 17.4243 119.464 17.3109C119.367 17.1974 119.245 17.1092 119.099 17.0454C118.951 16.9807 118.782 16.9484 118.587 16.9484ZM100.449 16.9484C100.26 16.9484 100.087 16.9878 99.9294 17.0635C99.7713 17.1399 99.6391 17.2392 99.5305 17.3613C99.4212 17.4842 99.3355 17.6197 99.2757 17.771C99.2166 17.9191 99.1856 18.0672 99.1856 18.2145H101.628C101.622 18.031 101.595 17.8632 101.547 17.7103C101.498 17.5575 101.425 17.4243 101.327 17.3109C101.231 17.1974 101.108 17.1092 100.962 17.0454C100.815 16.9807 100.644 16.9484 100.449 16.9484ZM110.193 14.2335C110.349 14.2335 110.482 14.2934 110.593 14.4123C110.705 14.5321 110.761 14.6739 110.761 14.8386C110.761 15.004 110.705 15.1466 110.593 15.2656C110.482 15.3845 110.349 15.4444 110.193 15.4444C110.039 15.4444 109.905 15.3845 109.794 15.2656C109.682 15.1466 109.626 15.004 109.626 14.8386C109.626 14.6739 109.682 14.5321 109.794 14.4123C109.905 14.2934 110.039 14.2335 110.193 14.2335ZM115.121 14.2335C115.276 14.2335 115.409 14.2934 115.521 14.4123C115.632 14.5321 115.688 14.6739 115.688 14.8386C115.688 15.004 115.632 15.1466 115.521 15.2656C115.409 15.3845 115.276 15.4444 115.121 15.4444C114.966 15.4444 114.833 15.3845 114.721 15.2656C114.61 15.1466 114.553 15.004 114.553 14.8386C114.553 14.6739 114.61 14.5321 114.721 14.4123C114.833 14.2934 114.966 14.2335 115.121 14.2335ZM51.4586 4.3679C49.0492 4.36159 47.4928 5.9058 47.4759 8.32217C47.4574 10.733 49.0263 12.33 51.415 12.3363C53.8466 12.3418 55.38 10.811 55.3852 8.3718C55.3903 5.93495 53.8599 4.37499 51.4586 4.3679Z" fill="white"/>
+  </svg>
+`;
