+++ js/ui/icons/IconDeveloper/index.js
@@ -1,0 +1,1 @@
+export { IconDeveloper } from './IconDeveloper.js';
