+++ js/ui/icons/IconGoogle/IconGoogle.js
@@ -1,0 +1,21 @@
+/* eslint-disable */
+
+/**
+ * @function IconGoogle
+ * @returns {string} HTML
+ */
+
+export const IconGoogle = () => `
+  <svg width="150" height="44" viewBox="0 0 150 44" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <rect x="0.5" y="0.5" width="149" height="43" rx="4.5" stroke="#A6A6A6"/>
+    <path d="M60.034 26.142C60.31 26.142 60.532 26.226 60.7 26.394C60.868 26.55 60.952 26.76 60.952 27.024V30.894C60.952 31.446 60.88 31.836 60.736 32.064C60.592 32.292 60.298 32.478 59.854 32.622C59.374 32.79 58.816 32.922 58.18 33.018C57.544 33.114 56.932 33.162 56.344 33.162C55.036 33.162 53.902 32.904 52.942 32.388C51.982 31.872 51.244 31.128 50.728 30.156C50.224 29.172 49.972 28.002 49.972 26.646C49.972 25.314 50.224 24.162 50.728 23.19C51.244 22.206 51.97 21.456 52.906 20.94C53.842 20.412 54.946 20.148 56.218 20.148C57.91 20.148 59.296 20.604 60.376 21.516C60.52 21.648 60.622 21.78 60.682 21.912C60.754 22.044 60.79 22.2 60.79 22.38C60.79 22.644 60.718 22.866 60.574 23.046C60.43 23.226 60.256 23.316 60.052 23.316C59.92 23.316 59.8 23.298 59.692 23.262C59.584 23.214 59.458 23.142 59.314 23.046C58.786 22.674 58.294 22.416 57.838 22.272C57.382 22.116 56.866 22.038 56.29 22.038C54.982 22.038 53.992 22.428 53.32 23.208C52.66 23.976 52.33 25.122 52.33 26.646C52.33 29.766 53.698 31.326 56.434 31.326C57.25 31.326 58.066 31.212 58.882 30.984V27.798H56.848C56.56 27.798 56.332 27.726 56.164 27.582C56.008 27.438 55.93 27.234 55.93 26.97C55.93 26.706 56.008 26.502 56.164 26.358C56.332 26.214 56.56 26.142 56.848 26.142H60.034ZM65.7652 33.144C64.8652 33.144 64.0732 32.958 63.3892 32.586C62.7172 32.214 62.1952 31.68 61.8232 30.984C61.4512 30.288 61.2652 29.478 61.2652 28.554C61.2652 27.63 61.4512 26.826 61.8232 26.142C62.1952 25.446 62.7172 24.912 63.3892 24.54C64.0732 24.168 64.8652 23.982 65.7652 23.982C66.6652 23.982 67.4512 24.168 68.1232 24.54C68.8072 24.912 69.3292 25.446 69.6892 26.142C70.0612 26.826 70.2472 27.63 70.2472 28.554C70.2472 29.478 70.0612 30.288 69.6892 30.984C69.3292 31.68 68.8072 32.214 68.1232 32.586C67.4512 32.958 66.6652 33.144 65.7652 33.144ZM65.7472 31.398C66.4792 31.398 67.0372 31.158 67.4212 30.678C67.8052 30.198 67.9972 29.49 67.9972 28.554C67.9972 27.63 67.8052 26.928 67.4212 26.448C67.0372 25.956 66.4852 25.71 65.7652 25.71C65.0452 25.71 64.4872 25.956 64.0912 26.448C63.7072 26.928 63.5152 27.63 63.5152 28.554C63.5152 29.49 63.7072 30.198 64.0912 30.678C64.4752 31.158 65.0272 31.398 65.7472 31.398ZM74.4987 33.144C73.5987 33.144 72.8067 32.958 72.1227 32.586C71.4507 32.214 70.9287 31.68 70.5567 30.984C70.1847 30.288 69.9987 29.478 69.9987 28.554C69.9987 27.63 70.1847 26.826 70.5567 26.142C70.9287 25.446 71.4507 24.912 72.1227 24.54C72.8067 24.168 73.5987 23.982 74.4987 23.982C75.3987 23.982 76.1847 24.168 76.8567 24.54C77.5407 24.912 78.0627 25.446 78.4227 26.142C78.7947 26.826 78.9807 27.63 78.9807 28.554C78.9807 29.478 78.7947 30.288 78.4227 30.984C78.0627 31.68 77.5407 32.214 76.8567 32.586C76.1847 32.958 75.3987 33.144 74.4987 33.144ZM74.4807 31.398C75.2127 31.398 75.7707 31.158 76.1547 30.678C76.5387 30.198 76.7307 29.49 76.7307 28.554C76.7307 27.63 76.5387 26.928 76.1547 26.448C75.7707 25.956 75.2187 25.71 74.4987 25.71C73.7787 25.71 73.2207 25.956 72.8247 26.448C72.4407 26.928 72.2487 27.63 72.2487 28.554C72.2487 29.49 72.4407 30.198 72.8247 30.678C73.2087 31.158 73.7607 31.398 74.4807 31.398ZM86.6702 24.018C87.0062 24.018 87.2762 24.114 87.4802 24.306C87.6842 24.498 87.7862 24.756 87.7862 25.08V32.19C87.7862 33.57 87.4022 34.614 86.6342 35.322C85.8662 36.03 84.7322 36.384 83.2322 36.384C81.9242 36.384 80.8322 36.138 79.9562 35.646C79.5122 35.382 79.2902 35.064 79.2902 34.692C79.2902 34.428 79.3502 34.218 79.4702 34.062C79.6022 33.906 79.7702 33.828 79.9742 33.828C80.0822 33.828 80.2142 33.852 80.3702 33.9C80.5262 33.96 80.6882 34.026 80.8562 34.098C81.2882 34.278 81.6722 34.416 82.0082 34.512C82.3562 34.608 82.7522 34.656 83.1962 34.656C84.7682 34.656 85.5542 33.888 85.5542 32.352V31.2C85.2902 31.692 84.9062 32.076 84.4022 32.352C83.8982 32.616 83.3222 32.748 82.6742 32.748C81.9062 32.748 81.2222 32.568 80.6222 32.208C80.0342 31.836 79.5722 31.32 79.2362 30.66C78.9002 30 78.7322 29.238 78.7322 28.374C78.7322 27.51 78.8942 26.748 79.2182 26.088C79.5542 25.416 80.0222 24.9 80.6222 24.54C81.2222 24.168 81.9062 23.982 82.6742 23.982C83.3222 23.982 83.8982 24.12 84.4022 24.396C84.9062 24.66 85.2902 25.038 85.5542 25.53V25.062C85.5542 24.75 85.6562 24.498 85.8602 24.306C86.0642 24.114 86.3342 24.018 86.6702 24.018ZM83.2682 31.002C83.9882 31.002 84.5462 30.768 84.9422 30.3C85.3502 29.832 85.5542 29.19 85.5542 28.374C85.5542 27.546 85.3502 26.898 84.9422 26.43C84.5462 25.962 83.9882 25.728 83.2682 25.728C82.5602 25.728 82.0022 25.962 81.5942 26.43C81.1862 26.898 80.9822 27.546 80.9822 28.374C80.9822 29.19 81.1862 29.832 81.5942 30.3C82.0022 30.768 82.5602 31.002 83.2682 31.002ZM89.6126 33.108C89.2886 33.108 89.0186 33.018 88.8026 32.838C88.5986 32.646 88.4966 32.376 88.4966 32.028V21.282C88.4966 20.934 88.5986 20.67 88.8026 20.49C89.0186 20.31 89.2886 20.22 89.6126 20.22C89.9366 20.22 90.2066 20.31 90.4226 20.49C90.6386 20.67 90.7466 20.934 90.7466 21.282V32.028C90.7466 32.376 90.6386 32.646 90.4226 32.838C90.2066 33.018 89.9366 33.108 89.6126 33.108ZM98.413 30.588C98.617 30.588 98.779 30.666 98.899 30.822C99.031 30.978 99.097 31.188 99.097 31.452C99.097 31.824 98.875 32.136 98.431 32.388C98.023 32.616 97.561 32.802 97.045 32.946C96.529 33.078 96.037 33.144 95.569 33.144C94.153 33.144 93.031 32.736 92.203 31.92C91.375 31.104 90.961 29.988 90.961 28.572C90.961 27.672 91.141 26.874 91.501 26.178C91.861 25.482 92.365 24.942 93.013 24.558C93.673 24.174 94.417 23.982 95.245 23.982C96.037 23.982 96.727 24.156 97.315 24.504C97.903 24.852 98.359 25.344 98.683 25.98C99.007 26.616 99.169 27.366 99.169 28.23C99.169 28.746 98.941 29.004 98.485 29.004H93.175C93.247 29.832 93.481 30.444 93.877 30.84C94.273 31.224 94.849 31.416 95.605 31.416C95.989 31.416 96.325 31.368 96.613 31.272C96.913 31.176 97.249 31.044 97.621 30.876C97.981 30.684 98.245 30.588 98.413 30.588ZM95.299 25.566C94.687 25.566 94.195 25.758 93.823 26.142C93.463 26.526 93.247 27.078 93.175 27.798H97.243C97.219 27.066 97.039 26.514 96.703 26.142C96.367 25.758 95.899 25.566 95.299 25.566ZM104.006 33.108C103.658 33.108 103.376 33.006 103.16 32.802C102.956 32.586 102.854 32.298 102.854 31.938V21.408C102.854 21.06 102.95 20.79 103.142 20.598C103.346 20.406 103.622 20.31 103.97 20.31H108.398C109.754 20.31 110.81 20.652 111.566 21.336C112.322 22.008 112.7 22.956 112.7 24.18C112.7 25.404 112.322 26.358 111.566 27.042C110.81 27.726 109.754 28.068 108.398 28.068H105.176V31.938C105.176 32.298 105.068 32.586 104.852 32.802C104.648 33.006 104.366 33.108 104.006 33.108ZM108.11 26.286C109.706 26.286 110.504 25.59 110.504 24.198C110.504 22.806 109.706 22.11 108.11 22.11H105.176V26.286H108.11ZM114.035 33.108C113.711 33.108 113.441 33.018 113.225 32.838C113.021 32.646 112.919 32.376 112.919 32.028V21.282C112.919 20.934 113.021 20.67 113.225 20.49C113.441 20.31 113.711 20.22 114.035 20.22C114.359 20.22 114.629 20.31 114.845 20.49C115.061 20.67 115.169 20.934 115.169 21.282V32.028C115.169 32.376 115.061 32.646 114.845 32.838C114.629 33.018 114.359 33.108 114.035 33.108ZM119.685 23.982C120.909 23.982 121.815 24.288 122.403 24.9C123.003 25.5 123.303 26.418 123.303 27.654V32.046C123.303 32.37 123.207 32.628 123.015 32.82C122.823 33 122.559 33.09 122.223 33.09C121.911 33.09 121.653 32.994 121.449 32.802C121.257 32.61 121.161 32.358 121.161 32.046V31.65C120.957 32.118 120.633 32.484 120.189 32.748C119.757 33.012 119.253 33.144 118.677 33.144C118.089 33.144 117.555 33.024 117.075 32.784C116.595 32.544 116.217 32.214 115.941 31.794C115.665 31.374 115.527 30.906 115.527 30.39C115.527 29.742 115.689 29.232 116.013 28.86C116.349 28.488 116.889 28.218 117.633 28.05C118.377 27.882 119.403 27.798 120.711 27.798H121.161V27.384C121.161 26.796 121.035 26.37 120.783 26.106C120.531 25.83 120.123 25.692 119.559 25.692C119.211 25.692 118.857 25.746 118.497 25.854C118.137 25.95 117.711 26.094 117.219 26.286C116.907 26.442 116.679 26.52 116.535 26.52C116.319 26.52 116.139 26.442 115.995 26.286C115.863 26.13 115.797 25.926 115.797 25.674C115.797 25.47 115.845 25.296 115.941 25.152C116.049 24.996 116.223 24.852 116.463 24.72C116.883 24.492 117.381 24.312 117.957 24.18C118.545 24.048 119.121 23.982 119.685 23.982ZM119.127 31.524C119.727 31.524 120.213 31.326 120.585 30.93C120.969 30.522 121.161 30 121.161 29.364V28.986H120.837C120.033 28.986 119.409 29.022 118.965 29.094C118.521 29.166 118.203 29.292 118.011 29.472C117.819 29.652 117.723 29.898 117.723 30.21C117.723 30.594 117.855 30.912 118.119 31.164C118.395 31.404 118.731 31.524 119.127 31.524ZM130.108 24.666C130.204 24.45 130.33 24.294 130.486 24.198C130.654 24.09 130.834 24.036 131.026 24.036C131.302 24.036 131.548 24.132 131.764 24.324C131.992 24.504 132.106 24.732 132.106 25.008C132.106 25.14 132.07 25.278 131.998 25.422L127.084 35.718C126.88 36.126 126.562 36.33 126.13 36.33C125.854 36.33 125.608 36.24 125.392 36.06C125.188 35.892 125.086 35.676 125.086 35.412C125.086 35.28 125.122 35.13 125.194 34.962L126.418 32.388L123.07 25.422C123.01 25.302 122.98 25.164 122.98 25.008C122.98 24.732 123.094 24.498 123.322 24.306C123.562 24.114 123.832 24.018 124.132 24.018C124.336 24.018 124.522 24.072 124.69 24.18C124.858 24.276 124.99 24.432 125.086 24.648L127.606 30.21L130.108 24.666Z" fill="white"/>
+    <path d="M55.13 12.19C55.2833 12.19 55.4067 12.2367 55.5 12.33C55.5933 12.4167 55.64 12.5333 55.64 12.68V14.83C55.64 15.1367 55.6 15.3533 55.52 15.48C55.44 15.6067 55.2767 15.71 55.03 15.79C54.7633 15.8833 54.4533 15.9567 54.1 16.01C53.7467 16.0633 53.4067 16.09 53.08 16.09C52.3533 16.09 51.7233 15.9467 51.19 15.66C50.6567 15.3733 50.2467 14.96 49.96 14.42C49.68 13.8733 49.54 13.2233 49.54 12.47C49.54 11.73 49.68 11.09 49.96 10.55C50.2467 10.0033 50.65 9.58667 51.17 9.3C51.69 9.00667 52.3033 8.86 53.01 8.86C53.95 8.86 54.72 9.11333 55.32 9.62C55.4 9.69333 55.4567 9.76667 55.49 9.84C55.53 9.91333 55.55 10 55.55 10.1C55.55 10.2467 55.51 10.37 55.43 10.47C55.35 10.57 55.2533 10.62 55.14 10.62C55.0667 10.62 55 10.61 54.94 10.59C54.88 10.5633 54.81 10.5233 54.73 10.47C54.4367 10.2633 54.1633 10.12 53.91 10.04C53.6567 9.95333 53.37 9.91 53.05 9.91C52.3233 9.91 51.7733 10.1267 51.4 10.56C51.0333 10.9867 50.85 11.6233 50.85 12.47C50.85 14.2033 51.61 15.07 53.13 15.07C53.5833 15.07 54.0367 15.0067 54.49 14.88V13.11H53.36C53.2 13.11 53.0733 13.07 52.98 12.99C52.8933 12.91 52.85 12.7967 52.85 12.65C52.85 12.5033 52.8933 12.39 52.98 12.31C53.0733 12.23 53.2 12.19 53.36 12.19H55.13ZM57.714 16C57.5207 16 57.3673 15.9467 57.254 15.84C57.1473 15.7333 57.094 15.5833 57.094 15.39V9.56C57.094 9.36667 57.1473 9.21667 57.254 9.11C57.3673 9.00333 57.5207 8.95 57.714 8.95H61.324C61.5173 8.95 61.6673 8.99333 61.774 9.08C61.8807 9.16667 61.934 9.29 61.934 9.45C61.934 9.61667 61.8807 9.74667 61.774 9.84C61.6673 9.92667 61.5173 9.97 61.324 9.97H58.354V11.91H61.124C61.3173 11.91 61.4673 11.9533 61.574 12.04C61.6807 12.1267 61.734 12.2533 61.734 12.42C61.734 12.58 61.6807 12.7033 61.574 12.79C61.4673 12.8767 61.3173 12.92 61.124 12.92H58.354V14.98H61.324C61.5173 14.98 61.6673 15.0267 61.774 15.12C61.8807 15.2067 61.934 15.3333 61.934 15.5C61.934 15.66 61.8807 15.7833 61.774 15.87C61.6673 15.9567 61.5173 16 61.324 16H57.714ZM65.3024 16.05C65.1024 16.05 64.9424 15.9933 64.8224 15.88C64.7091 15.76 64.6524 15.6033 64.6524 15.41V10.01H62.8324C62.4324 10.01 62.2324 9.83333 62.2324 9.48C62.2324 9.12667 62.4324 8.95 62.8324 8.95H67.7624C68.1624 8.95 68.3624 9.12667 68.3624 9.48C68.3624 9.83333 68.1624 10.01 67.7624 10.01H65.9424V15.41C65.9424 15.6033 65.8824 15.76 65.7624 15.88C65.6491 15.9933 65.4958 16.05 65.3024 16.05ZM72.4801 16.06C72.2867 16.06 72.1301 16 72.0101 15.88C71.8967 15.76 71.8401 15.6033 71.8401 15.41V9.54C71.8401 9.34667 71.8967 9.19333 72.0101 9.08C72.1301 8.96 72.2867 8.9 72.4801 8.9C72.6801 8.9 72.8367 8.96 72.9501 9.08C73.0701 9.19333 73.1301 9.34667 73.1301 9.54V15.41C73.1301 15.6033 73.0701 15.76 72.9501 15.88C72.8367 16 72.6801 16.06 72.4801 16.06ZM76.9626 16.05C76.7626 16.05 76.6026 15.9933 76.4826 15.88C76.3692 15.76 76.3126 15.6033 76.3126 15.41V10.01H74.4926C74.0926 10.01 73.8926 9.83333 73.8926 9.48C73.8926 9.12667 74.0926 8.95 74.4926 8.95H79.4226C79.8226 8.95 80.0226 9.12667 80.0226 9.48C80.0226 9.83333 79.8226 10.01 79.4226 10.01H77.6026V15.41C77.6026 15.6033 77.5426 15.76 77.4226 15.88C77.3092 15.9933 77.1559 16.05 76.9626 16.05ZM86.6402 16.09C85.9536 16.09 85.3569 15.9433 84.8502 15.65C84.3502 15.3567 83.9636 14.9367 83.6902 14.39C83.4169 13.8433 83.2802 13.2033 83.2802 12.47C83.2802 11.7367 83.4136 11.1 83.6802 10.56C83.9536 10.0133 84.3436 9.59333 84.8502 9.3C85.3569 9.00667 85.9536 8.86 86.6402 8.86C87.3269 8.86 87.9202 9.00667 88.4202 9.3C88.9269 9.59333 89.3136 10.0133 89.5802 10.56C89.8536 11.1 89.9902 11.7367 89.9902 12.47C89.9902 13.2033 89.8536 13.8433 89.5802 14.39C89.3069 14.9367 88.9169 15.3567 88.4102 15.65C87.9102 15.9433 87.3202 16.09 86.6402 16.09ZM86.6402 15.06C87.2869 15.06 87.7902 14.8367 88.1502 14.39C88.5169 13.9433 88.7002 13.3033 88.7002 12.47C88.7002 11.6367 88.5169 11 88.1502 10.56C87.7902 10.1133 87.2869 9.89 86.6402 9.89C85.9869 9.89 85.4769 10.1133 85.1102 10.56C84.7502 11 84.5702 11.6367 84.5702 12.47C84.5702 13.3033 84.7502 13.9433 85.1102 14.39C85.4769 14.8367 85.9869 15.06 86.6402 15.06ZM96.6332 8.9C96.8199 8.9 96.9665 8.96 97.0732 9.08C97.1865 9.19333 97.2432 9.34667 97.2432 9.54V15.41C97.2432 15.6033 97.1865 15.76 97.0732 15.88C96.9599 16 96.8132 16.06 96.6332 16.06C96.4199 16.06 96.2565 15.9833 96.1432 15.83L92.4932 11.08V15.41C92.4932 15.6033 92.4399 15.76 92.3332 15.88C92.2265 16 92.0799 16.06 91.8932 16.06C91.7065 16.06 91.5599 16 91.4532 15.88C91.3465 15.76 91.2932 15.6033 91.2932 15.41V9.54C91.2932 9.34667 91.3499 9.19333 91.4632 9.08C91.5765 8.96 91.7265 8.9 91.9132 8.9C92.1199 8.9 92.2799 8.97667 92.3932 9.13L96.0332 13.86V9.54C96.0332 9.34667 96.0865 9.19333 96.1932 9.08C96.3065 8.96 96.4532 8.9 96.6332 8.9Z" fill="white"/>
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M33.4722 26.6532L28.9963 22.1554V21.8401L33.4722 17.3423L33.5694 17.4024L38.8638 20.4259C40.3786 21.2837 40.3786 22.7005 38.8638 23.5696L33.5694 26.5931L33.4722 26.6532Z" fill="#FFE000"/>
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M33.5703 26.5932L28.9955 21.9978L15.4941 35.5679C15.9884 36.0991 16.8182 36.1601 17.7442 35.6411L33.5694 26.5932" fill="#FF3A44"/>
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M33.5702 17.4025L17.7433 8.36851C16.8172 7.83644 15.9848 7.90959 15.4932 8.44166L28.9963 21.9979L33.5693 17.4025H33.5702Z" fill="#06E775"/>
+    <path opacity="0.2" d="M33.4723 26.4956L17.7555 35.4704C16.8771 35.9781 16.0941 35.9424 15.5886 35.4808L15.5054 35.5653L15.5886 35.6384C16.0941 36.0974 16.8771 36.1348 17.7555 35.6271L33.5798 26.5931L33.4723 26.4956Z" fill="black"/>
+    <path opacity="0.12" d="M38.8639 23.412L33.4585 26.4956L33.5565 26.5932L38.85 23.5696C39.6087 23.1351 39.9816 22.5647 39.9816 21.9978C39.9356 22.5194 39.5498 23.0132 38.863 23.412H38.8639Z" fill="black"/>
+    <path opacity="0.25" fill-rule="evenodd" clip-rule="evenodd" d="M17.7443 8.52605L38.8639 20.5836C39.5489 20.9719 39.9357 21.4796 39.9955 21.9978C39.9955 21.4309 39.6226 20.8605 38.8639 20.4259L17.7443 8.36931C16.2295 7.49761 15 8.22475 15 9.96467V10.1223C15 8.37976 16.2295 7.66655 17.7443 8.52605Z" fill="white"/>
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M15.4942 8.42778C15.1812 8.7674 15 9.28555 15 9.96392V34.0337C15 34.7129 15.1812 35.231 15.5055 35.5576L15.5888 35.6308L29.0078 22.1442V21.8429L15.5783 8.35376L15.4942 8.42778Z" fill="#00D2FF"/>
+  </svg>
+`;
