+++ js/ui/icons/IconGoogle/index.js
@@ -1,0 +1,1 @@
+export { IconGoogle } from './IconGoogle.js';
