+++ js/ui/icons/IconIn/IconIn.js
@@ -1,0 +1,20 @@
+/* eslint-disable */
+
+/**
+ * @function IconIn
+ * @returns {string}
+ */
+
+export const IconIn = () => `
+  <svg width="40" height="40" viewBox="0 0 40 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <rect x="1" y="1" width="38" height="38" stroke="white" stroke-width="2"/>
+    <g clip-path="url(#clip0_941_951)">
+      <path d="M15.0299 27.875H11.2973V17.3598H15.0299V27.875ZM13.1616 15.9254C11.9683 15.9254 11 15.0605 11 14.0164C11 12.9723 11.9683 12.125 13.1616 12.125C14.3549 12.125 15.3232 12.9723 15.3232 14.0164C15.3232 15.0605 14.3549 15.9254 13.1616 15.9254ZM29 27.875H25.2754V22.7562C25.2754 21.5363 25.2473 19.9719 23.3348 19.9719C21.3942 19.9719 21.0969 21.2973 21.0969 22.6684V27.875H17.3683V17.3598H20.9482V18.7941H21.0004C21.4986 17.968 22.7161 17.0961 24.5321 17.0961C28.3089 17.0961 29.004 19.2723 29.004 22.0988V27.875H29Z" fill="white"/>
+    </g>
+    <defs>
+      <clipPath id="clip0_941_951">
+        <rect width="18" height="18" fill="white" transform="translate(11 11)"/>
+      </clipPath>
+    </defs>
+  </svg>
+`;
