+++ js/ui/icons/IconIn/index.js
@@ -1,0 +1,1 @@
+export { IconIn } from './IconIn.js';
