+++ js/ui/icons/IconLogo/IconLogo.js
@@ -1,0 +1,13 @@
+/* eslint-disable */
+
+/**
+ * @function IconLogo
+ * @returns {string} HTML
+ */
+
+export const IconLogo = () => `
+  <svg viewBox="0 0 102 40" xmlns="http://www.w3.org/2000/svg">
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M68.5369 25.1423C68.8238 25.3894 68.9807 25.753 68.9647 26.1307C68.9647 27.195 68.3408 27.7262 67.1002 27.7262H56.4159C55.596 27.7757 54.8865 27.1562 54.8384 26.3425C54.8348 26.2931 54.8348 26.2419 54.8366 26.1907C54.8366 25.6436 55.0648 25.1247 55.4658 24.7522L64.2856 15.3554H57.4587C57.013 15.3854 56.5692 15.2584 56.2109 14.9936C55.9186 14.7377 55.7617 14.3635 55.7831 13.977C55.7831 12.9339 56.4034 12.4114 57.6476 12.4114H67.8577C68.2891 12.3902 68.7115 12.5561 69.011 12.865C69.2891 13.1562 69.4424 13.5445 69.437 13.947C69.4263 14.4976 69.1982 15.0254 68.806 15.4172L60.0183 24.7822H67.2873C67.733 24.7522 68.1786 24.8793 68.5369 25.1423ZM74.1785 18.6736H81.3851C81.4795 17.6411 81.1872 16.6086 80.5633 15.779C79.9573 15.103 79.0714 14.7412 78.1605 14.7924C75.9271 14.7924 74.6009 16.0861 74.1785 18.6754V18.6736ZM84.3743 20.4121C84.1248 20.6539 83.7861 20.7863 83.4403 20.7722H74.0537C74.0269 21.93 74.4494 23.0508 75.2355 23.905C76.0786 24.6569 77.1891 25.0399 78.3192 24.9711C78.9965 24.9799 79.6739 24.8899 80.3263 24.7064C80.9537 24.514 81.5544 24.2475 82.1123 23.905C82.353 23.7674 82.6239 23.6932 82.902 23.6862C83.2282 23.6721 83.5437 23.8186 83.7397 24.078C83.9447 24.3551 84.0517 24.6905 84.0392 25.0329C84.0428 25.633 83.7469 26.196 83.2496 26.5366C82.5633 27.0097 81.7986 27.3609 80.9894 27.5709C80.0375 27.8427 79.0554 27.9804 78.0661 27.9769C76.6312 28.0104 75.2105 27.6927 73.9253 27.0538C72.7738 26.4678 71.8238 25.5553 71.191 24.4381C70.5279 23.2291 70.1946 21.8665 70.2284 20.4898C70.2017 18.979 70.5386 17.4822 71.207 16.1232C71.8184 14.8824 72.7721 13.8411 73.9574 13.1157C75.1767 12.3849 76.5777 12.009 78.0037 12.0337C79.8646 11.9208 81.6827 12.6179 82.9822 13.9452C84.2388 15.4772 84.8663 17.4222 84.7362 19.3955C84.7611 19.7697 84.6256 20.1368 84.3725 20.4121H84.3743ZM100.64 13.4281C101.611 14.5753 102.092 16.0544 101.985 17.5475C101.985 17.9429 101.973 18.2465 101.95 18.4565L101.319 26.0354C101.304 26.5614 101.076 27.0573 100.686 27.4121C100.285 27.7563 99.7626 27.9345 99.2332 27.9151C98.7109 27.9469 98.1976 27.7651 97.809 27.4138C97.4614 27.0432 97.2974 26.5366 97.3687 26.0354L98.0015 18.6136C98.169 16.2944 97.2421 15.1366 95.2208 15.1366C94.2048 15.0801 93.2155 15.456 92.5007 16.1691C91.8002 16.9722 91.3991 17.9888 91.3635 19.0513L90.7628 26.0354C90.7485 26.5614 90.5203 27.0573 90.13 27.4121C89.7289 27.7563 89.2067 27.9345 88.6773 27.9151C88.1603 27.9486 87.647 27.7668 87.2691 27.4138C86.9304 27.0379 86.7629 26.5366 86.811 26.0354L87.8234 13.8817C87.8252 13.3786 88.0587 12.9003 88.4544 12.5826C88.8769 12.2526 89.4045 12.0814 89.941 12.0973C90.4312 12.069 90.9089 12.2385 91.269 12.5667C91.5916 12.9021 91.7324 13.3663 91.6487 13.8199L91.6148 14.4147C92.1531 13.6557 92.8822 13.0503 93.7342 12.6603C94.6629 12.2349 95.6789 12.0196 96.7056 12.0355C98.1565 11.9313 99.5843 12.4379 100.64 13.4298V13.4281ZM38.4448 11.6031C36.8334 9.14265 33.6178 8.26369 30.9601 9.55566C28.1224 10.9853 26.9905 14.4217 28.429 17.2334C29.2507 18.836 30.7854 19.9638 32.5715 20.2762C32.354 27.0538 26.6322 32.377 19.7892 32.1617C13.1441 31.9534 7.84832 26.5861 7.78771 19.9974C7.77702 14.9212 10.9606 10.3781 15.7662 8.61669C16.6895 8.28663 17.2242 7.32824 17.0103 6.37691L16.9961 6.31867C16.7608 5.27732 17.4203 4.24657 18.4684 4.01183C18.554 3.99418 18.6413 3.98006 18.7287 3.973C19.2099 3.92887 19.6966 3.90593 20.1832 3.90593C20.7857 3.90593 21.3899 3.94123 21.9871 4.00477C23.0708 4.12302 24.0476 3.34642 24.1635 2.27507C24.1688 2.23977 24.1724 2.20094 24.1742 2.16564C24.2027 1.10135 23.3863 0.199444 22.3168 0.111194C21.6074 0.0352998 20.8944 0 20.1832 0C9.03724 0 0 8.95557 0 20.0009C0 31.048 9.03724 40 20.1832 40C31.3309 40 40.3663 31.048 40.3663 20.0009C40.3663 18.0453 39.9688 13.8305 38.4448 11.6031Z" fill="currentColor" />
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M26.0962 4C27.2542 4 28.1927 4.9278 28.1927 6.07193C28.1927 7.21606 27.2542 8.14386 26.0962 8.14386C24.9386 8.14386 24 7.21606 24 6.07193C24 4.9278 24.9386 4 26.0962 4Z" fill="currentColor" />
+  </svg>
+`;
