+++ js/ui/icons/IconLogo/index.js
@@ -1,0 +1,1 @@
+export { IconLogo } from './IconLogo.js';
