+++ js/ui/icons/IconLogo2/index.js
@@ -1,0 +1,1 @@
+export { IconLogo2 } from './IconLogo2.js';
