+++ js/ui/icons/IconMoon/IconMoon.js
@@ -1,0 +1,16 @@
+/* eslint-disable */
+
+/**
+ * @function IconMoon
+ * @returns {string} HTML
+ */
+
+export const IconMoon = () => `
+  <svg width="24" height="24" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+    <path d="M23.9999 19.314C8.10542 22.1253 2.68539 9.23676 12.2615 0C-1.80277 0.644425 -1.87046 14.9797 2.68539 19.9371C11.4083 29.4287 23.9999 19.314 23.9999 19.314Z" fill="currentColor"/>
+    <path d="M15.1999 5.60003C14.9959 5.60003 14.8939 5.49485 14.8939 5.28449V3.46777H13.1059C12.9019 3.46777 12.7999 3.36896 12.7999 3.17135C12.7999 2.97374 12.9019 2.87494 13.1059 2.87494H14.8939V1.11559C14.8939 0.905228 14.9959 0.800049 15.1999 0.800049C15.4039 0.800049 15.5059 0.905228 15.5059 1.11559V2.87494H17.2939C17.4979 2.87494 17.5999 2.97374 17.5999 3.17135C17.5999 3.36896 17.4979 3.46777 17.2939 3.46777H15.5059V5.28449C15.5059 5.49485 15.4039 5.60003 15.1999 5.60003Z" fill="currentColor"/>
+    <path d="M21.1997 8.80015C21.0297 8.80015 20.9447 8.71251 20.9447 8.53721V7.02327H19.4547C19.2847 7.02327 19.1997 6.94093 19.1997 6.77626C19.1997 6.61158 19.2847 6.52925 19.4547 6.52925H20.9447V5.06312C20.9447 4.88782 21.0297 4.80017 21.1997 4.80017C21.3697 4.80017 21.4547 4.88782 21.4547 5.06312V6.52925H22.9447C23.1147 6.52925 23.1997 6.61158 23.1997 6.77626C23.1997 6.94093 23.1147 7.02327 22.9447 7.02327H21.4547V8.53721C21.4547 8.71251 21.3697 8.80015 21.1997 8.80015Z" fill="currentColor"/>
+    <path d="M13.1999 12.7998C13.03 12.7998 12.945 12.7121 12.945 12.5368V11.0229H11.4549C11.2849 11.0229 11.2 10.9406 11.2 10.7759C11.2 10.6112 11.2849 10.5289 11.4549 10.5289H12.945V9.06275C12.945 8.88745 13.03 8.7998 13.1999 8.7998C13.3699 8.7998 13.4549 8.88745 13.4549 9.06275V10.5289H14.945C15.1149 10.5289 15.1999 10.6112 15.1999 10.7759C15.1999 10.9406 15.1149 11.0229 14.945 11.0229H13.4549V12.5368C13.4549 12.7121 13.3699 12.7998 13.1999 12.7998Z" fill="currentColor"/>
+    <path d="M21.1997 16.7999C21.0297 16.7999 20.9447 16.7123 20.9447 16.537V15.023H19.4547C19.2847 15.023 19.1997 14.9407 19.1997 14.776C19.1997 14.6113 19.2847 14.529 19.4547 14.529H20.9447V13.0629C20.9447 12.8876 21.0297 12.7999 21.1997 12.7999C21.3697 12.7999 21.4547 12.8876 21.4547 13.0629V14.529H22.9447C23.1147 14.529 23.1997 14.6113 23.1997 14.776C23.1997 14.9407 23.1147 15.023 22.9447 15.023H21.4547V16.537C21.4547 16.7123 21.3697 16.7999 21.1997 16.7999Z" fill="currentColor"/>
+  </svg>
+`;
