+++ js/ui/icons/IconMoon/index.js
@@ -1,0 +1,1 @@
+export { IconMoon } from './IconMoon.js';
