+++ js/ui/icons/IconSun/IconSun.js
@@ -1,0 +1,12 @@
+/* eslint-disable */
+
+/**
+ * @function IconSun
+ * @returns {string} HTML
+ */
+
+export const IconSun = () => `
+  <svg viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+    <path d="M11.9856 0.00257827C11.9612 0.0070376 11.9367 0.0137257 11.9145 0.0204144C11.6499 0.0806116 11.4631 0.31917 11.4698 0.591171V4.01571C11.4676 4.22083 11.5743 4.41257 11.7522 4.51735C11.93 4.61991 12.1479 4.61991 12.3258 4.51735C12.5037 4.41257 12.6104 4.22083 12.6082 4.01571V0.591171C12.6149 0.426187 12.5482 0.267891 12.4303 0.156415C12.3102 0.0427098 12.1479 -0.0130285 11.9856 0.00257827ZM3.8745 3.35577C3.66105 3.39368 3.48985 3.54974 3.42981 3.75709C3.36978 3.96666 3.43204 4.18961 3.5899 4.33676L6.00901 6.76248C6.14686 6.93192 6.36698 7.00996 6.58043 6.96091C6.79166 6.91186 6.95842 6.74464 7.00733 6.53284C7.05625 6.3188 6.97843 6.09808 6.80945 5.95985L4.39034 3.53414C4.2725 3.40705 4.10129 3.3424 3.92787 3.35577C3.91008 3.35577 3.89229 3.35577 3.8745 3.35577ZM20.0256 3.35577C19.8966 3.37361 19.7766 3.43827 19.6876 3.53414L17.2685 5.95985C17.0995 6.09808 17.0217 6.3188 17.0706 6.53284C17.1196 6.74464 17.2863 6.91186 17.4975 6.96091C17.711 7.00996 17.9311 6.93192 18.069 6.76248L20.4881 4.33676C20.6659 4.16732 20.7171 3.90201 20.6126 3.67905C20.5059 3.45387 20.2702 3.32456 20.0256 3.35577ZM11.9856 6.29874C11.9678 6.3032 11.9501 6.30989 11.9323 6.31657C11.8967 6.3188 11.8611 6.32549 11.8255 6.33441C11.8189 6.3411 11.8144 6.34556 11.8077 6.35225C8.78831 6.48156 6.34697 8.94741 6.34697 12.0063C6.34697 15.1477 8.90616 17.7139 12.039 17.7139C15.1718 17.7139 17.731 15.1477 17.731 12.0063C17.731 8.95855 15.3097 6.50163 12.3058 6.35225C12.2858 6.35225 12.2724 6.33441 12.2524 6.33441C12.1946 6.31212 12.1346 6.30097 12.0746 6.29874C12.0634 6.29874 12.0501 6.29874 12.039 6.29874C12.0212 6.29874 12.0034 6.29874 11.9856 6.29874ZM12.0034 7.44025C12.0145 7.44025 12.0279 7.44025 12.039 7.44025C12.0568 7.44025 12.0746 7.44025 12.0924 7.44025C14.5826 7.46924 16.5926 9.50256 16.5926 12.0063C16.5926 14.5301 14.5559 16.5724 12.039 16.5724C9.52427 16.5724 7.48537 14.5301 7.48537 12.0063C7.48537 9.49587 9.50426 7.46032 12.0034 7.44025ZM0.494869 11.4355C0.181363 11.4801 -0.0387576 11.7722 0.0057113 12.0866C0.0501802 12.4009 0.341451 12.6217 0.654957 12.5771H4.07017C4.27472 12.5793 4.46594 12.4723 4.57044 12.2939C4.67272 12.1156 4.67272 11.8971 4.57044 11.7187C4.46594 11.5403 4.27472 11.4333 4.07017 11.4355H0.654957C0.637169 11.4355 0.619382 11.4355 0.601594 11.4355C0.583807 11.4355 0.566019 11.4355 0.548231 11.4355C0.530444 11.4355 0.512656 11.4355 0.494869 11.4355ZM19.8477 11.4355C19.5342 11.4801 19.3141 11.7722 19.3586 12.0866C19.403 12.4009 19.6943 12.6217 20.0078 12.5771H23.423C23.6276 12.5793 23.8188 12.4723 23.9233 12.2939C24.0256 12.1156 24.0256 11.8971 23.9233 11.7187C23.8188 11.5403 23.6276 11.4333 23.423 11.4355H20.0078C19.99 11.4355 19.9722 11.4355 19.9544 11.4355C19.9367 11.4355 19.9189 11.4355 19.9011 11.4355C19.8833 11.4355 19.8655 11.4355 19.8477 11.4355ZM6.34697 17.0718C6.21801 17.0896 6.09795 17.1543 6.00901 17.2501L3.5899 19.6758C3.42092 19.8141 3.3431 20.0348 3.39202 20.2488C3.44093 20.4606 3.60769 20.6279 3.81892 20.6769C4.03237 20.726 4.25249 20.6479 4.39034 20.4785L6.80945 18.0528C6.97843 17.89 7.02957 17.6381 6.93841 17.4218C6.84947 17.2055 6.6338 17.0673 6.40033 17.0718C6.38255 17.0718 6.36476 17.0718 6.34697 17.0718ZM17.5531 17.0718C17.3397 17.1097 17.1685 17.2657 17.1084 17.4731C17.0484 17.6827 17.1107 17.9056 17.2685 18.0528L19.6876 20.4785C19.8255 20.6479 20.0456 20.726 20.2591 20.6769C20.4703 20.6279 20.637 20.4606 20.686 20.2488C20.7349 20.0348 20.6571 19.8141 20.4881 19.6758L18.069 17.2501C17.9622 17.1364 17.8155 17.074 17.6599 17.0718C17.6421 17.0718 17.6243 17.0718 17.6065 17.0718C17.5887 17.0718 17.5709 17.0718 17.5531 17.0718ZM11.9856 19.4083C11.9612 19.4128 11.9367 19.4195 11.9145 19.4261C11.6499 19.4863 11.4631 19.7249 11.4698 19.9969V23.4214C11.4676 23.6266 11.5743 23.8183 11.7522 23.9231C11.93 24.0256 12.1479 24.0256 12.3258 23.9231C12.5037 23.8183 12.6104 23.6266 12.6082 23.4214V19.9969C12.6149 19.8319 12.5482 19.6736 12.4303 19.5621C12.3102 19.4484 12.1479 19.3927 11.9856 19.4083Z" fill="currentColor"/>
+  </svg>
+`;
