+++ js/ui/icons/IconSun/index.js
@@ -1,0 +1,1 @@
+export { IconSun } from './IconSun.js';
