+++ js/ui/icons/index.js
@@ -1,138 +1,10 @@
-/* eslint-disable */
-
-/**
- * @function IconApple
- * @returns {string} HTML
- */
-
-export const IconApple = () => `
-  <svg width="150" height="44" viewBox="0 0 150 44" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <rect x="0.5" y="0.5" width="149" height="43" rx="4.5" stroke="#A6A6A6"/>
-    <path d="M62.87 32.614C62.942 32.77 62.978 32.926 62.978 33.082C62.978 33.358 62.864 33.592 62.636 33.784C62.42 33.976 62.168 34.072 61.88 34.072C61.688 34.072 61.508 34.024 61.34 33.928C61.172 33.82 61.04 33.658 60.944 33.442L59.864 31.03H53.528L52.448 33.442C52.352 33.658 52.22 33.82 52.052 33.928C51.884 34.024 51.698 34.072 51.494 34.072C51.218 34.072 50.966 33.976 50.738 33.784C50.51 33.592 50.396 33.358 50.396 33.082C50.396 32.926 50.432 32.77 50.504 32.614L55.472 21.94C55.58 21.7 55.742 21.52 55.958 21.4C56.186 21.268 56.426 21.202 56.678 21.202C56.93 21.202 57.164 21.268 57.38 21.4C57.608 21.52 57.776 21.7 57.884 21.94L62.87 32.614ZM54.32 29.23H59.072L56.696 23.884L54.32 29.23ZM69.5874 24.982C70.3434 24.982 71.0154 25.174 71.6034 25.558C72.2034 25.942 72.6654 26.488 72.9894 27.196C73.3254 27.892 73.4934 28.696 73.4934 29.608C73.4934 30.52 73.3254 31.318 72.9894 32.002C72.6654 32.686 72.2094 33.214 71.6214 33.586C71.0334 33.958 70.3554 34.144 69.5874 34.144C68.9634 34.144 68.4054 34.012 67.9134 33.748C67.4214 33.484 67.0434 33.118 66.7794 32.65V36.322C66.7794 36.634 66.6774 36.88 66.4734 37.06C66.2694 37.24 65.9994 37.33 65.6634 37.33C65.3274 37.33 65.0514 37.234 64.8354 37.042C64.6314 36.85 64.5294 36.598 64.5294 36.286V26.08C64.5294 25.756 64.6254 25.498 64.8174 25.306C65.0214 25.114 65.2914 25.018 65.6274 25.018C65.9634 25.018 66.2334 25.114 66.4374 25.306C66.6414 25.498 66.7434 25.756 66.7434 26.08V26.53C67.0074 26.038 67.3854 25.66 67.8774 25.396C68.3814 25.12 68.9514 24.982 69.5874 24.982ZM68.9934 32.398C69.7014 32.398 70.2534 32.158 70.6494 31.678C71.0454 31.186 71.2434 30.496 71.2434 29.608C71.2434 28.696 71.0454 27.988 70.6494 27.484C70.2534 26.98 69.7014 26.728 68.9934 26.728C68.2734 26.728 67.7154 26.974 67.3194 27.466C66.9354 27.958 66.7434 28.66 66.7434 29.572C66.7434 30.472 66.9354 31.168 67.3194 31.66C67.7154 32.152 68.2734 32.398 68.9934 32.398ZM80.3979 24.982C81.1539 24.982 81.8259 25.174 82.4139 25.558C83.0139 25.942 83.4759 26.488 83.7999 27.196C84.1359 27.892 84.3039 28.696 84.3039 29.608C84.3039 30.52 84.1359 31.318 83.7999 32.002C83.4759 32.686 83.0199 33.214 82.4319 33.586C81.8439 33.958 81.1659 34.144 80.3979 34.144C79.7739 34.144 79.2159 34.012 78.7239 33.748C78.2319 33.484 77.8539 33.118 77.5899 32.65V36.322C77.5899 36.634 77.4879 36.88 77.2839 37.06C77.0799 37.24 76.8099 37.33 76.4739 37.33C76.1379 37.33 75.8619 37.234 75.6459 37.042C75.4419 36.85 75.3399 36.598 75.3399 36.286V26.08C75.3399 25.756 75.4359 25.498 75.6279 25.306C75.8319 25.114 76.1019 25.018 76.4379 25.018C76.7739 25.018 77.0439 25.114 77.2479 25.306C77.4519 25.498 77.5539 25.756 77.5539 26.08V26.53C77.8179 26.038 78.1959 25.66 78.6879 25.396C79.1919 25.12 79.7619 24.982 80.3979 24.982ZM79.8039 32.398C80.5119 32.398 81.0639 32.158 81.4599 31.678C81.8559 31.186 82.0539 30.496 82.0539 29.608C82.0539 28.696 81.8559 27.988 81.4599 27.484C81.0639 26.98 80.5119 26.728 79.8039 26.728C79.0839 26.728 78.5259 26.974 78.1299 27.466C77.7459 27.958 77.5539 28.66 77.5539 29.572C77.5539 30.472 77.7459 31.168 78.1299 31.66C78.5259 32.152 79.0839 32.398 79.8039 32.398ZM95.6272 34.162C94.7032 34.162 93.8272 34.042 92.9992 33.802C92.1832 33.55 91.5172 33.214 91.0012 32.794C90.7252 32.59 90.5872 32.302 90.5872 31.93C90.5872 31.678 90.6592 31.462 90.8032 31.282C90.9592 31.09 91.1392 30.994 91.3432 30.994C91.5592 30.994 91.8052 31.078 92.0812 31.246C92.5972 31.618 93.1372 31.888 93.7012 32.056C94.2652 32.224 94.8892 32.308 95.5732 32.308C96.4492 32.308 97.1212 32.158 97.5892 31.858C98.0572 31.546 98.2912 31.096 98.2912 30.508C98.2912 30.052 98.0692 29.704 97.6252 29.464C97.1932 29.212 96.4672 28.972 95.4472 28.744C94.3912 28.516 93.5332 28.252 92.8732 27.952C92.2132 27.652 91.7152 27.268 91.3792 26.8C91.0432 26.332 90.8752 25.738 90.8752 25.018C90.8752 24.274 91.0852 23.608 91.5052 23.02C91.9252 22.432 92.5072 21.976 93.2512 21.652C94.0072 21.316 94.8532 21.148 95.7892 21.148C97.5052 21.148 98.9032 21.604 99.9832 22.516C100.139 22.648 100.247 22.78 100.307 22.912C100.379 23.032 100.415 23.188 100.415 23.38C100.415 23.632 100.337 23.854 100.181 24.046C100.037 24.226 99.8632 24.316 99.6592 24.316C99.5392 24.316 99.4252 24.298 99.3172 24.262C99.2212 24.226 99.0892 24.16 98.9212 24.064C98.4292 23.704 97.9492 23.44 97.4812 23.272C97.0252 23.092 96.4612 23.002 95.7892 23.002C94.9732 23.002 94.3252 23.17 93.8452 23.506C93.3772 23.83 93.1432 24.286 93.1432 24.874C93.1432 25.366 93.3532 25.744 93.7732 26.008C94.1932 26.272 94.9072 26.518 95.9152 26.746C96.9832 26.986 97.8472 27.256 98.5072 27.556C99.1792 27.844 99.6892 28.216 100.037 28.672C100.397 29.128 100.577 29.698 100.577 30.382C100.577 31.126 100.367 31.786 99.9472 32.362C99.5392 32.926 98.9572 33.37 98.2012 33.694C97.4572 34.006 96.5992 34.162 95.6272 34.162ZM106.965 32.434C107.565 32.47 107.865 32.74 107.865 33.244C107.865 33.532 107.745 33.754 107.505 33.91C107.277 34.054 106.947 34.114 106.515 34.09L106.029 34.054C104.013 33.91 103.005 32.83 103.005 30.814V26.89H102.105C101.781 26.89 101.529 26.818 101.349 26.674C101.181 26.53 101.097 26.32 101.097 26.044C101.097 25.768 101.181 25.558 101.349 25.414C101.529 25.27 101.781 25.198 102.105 25.198H103.005V23.542C103.005 23.218 103.107 22.96 103.311 22.768C103.515 22.576 103.791 22.48 104.139 22.48C104.475 22.48 104.745 22.576 104.949 22.768C105.153 22.96 105.255 23.218 105.255 23.542V25.198H106.785C107.109 25.198 107.355 25.27 107.523 25.414C107.703 25.558 107.793 25.768 107.793 26.044C107.793 26.32 107.703 26.53 107.523 26.674C107.355 26.818 107.109 26.89 106.785 26.89H105.255V30.976C105.255 31.864 105.663 32.338 106.479 32.398L106.965 32.434ZM113.051 34.144C112.151 34.144 111.359 33.958 110.675 33.586C110.003 33.214 109.481 32.68 109.109 31.984C108.737 31.288 108.551 30.478 108.551 29.554C108.551 28.63 108.737 27.826 109.109 27.142C109.481 26.446 110.003 25.912 110.675 25.54C111.359 25.168 112.151 24.982 113.051 24.982C113.951 24.982 114.737 25.168 115.409 25.54C116.093 25.912 116.615 26.446 116.975 27.142C117.347 27.826 117.533 28.63 117.533 29.554C117.533 30.478 117.347 31.288 116.975 31.984C116.615 32.68 116.093 33.214 115.409 33.586C114.737 33.958 113.951 34.144 113.051 34.144ZM113.033 32.398C113.765 32.398 114.323 32.158 114.707 31.678C115.091 31.198 115.283 30.49 115.283 29.554C115.283 28.63 115.091 27.928 114.707 27.448C114.323 26.956 113.771 26.71 113.051 26.71C112.331 26.71 111.773 26.956 111.377 27.448C110.993 27.928 110.801 28.63 110.801 29.554C110.801 30.49 110.993 31.198 111.377 31.678C111.761 32.158 112.313 32.398 113.033 32.398ZM124.305 25.018C124.617 24.994 124.863 25.06 125.043 25.216C125.223 25.372 125.313 25.606 125.313 25.918C125.313 26.242 125.235 26.482 125.079 26.638C124.923 26.794 124.641 26.896 124.233 26.944L123.693 26.998C122.985 27.07 122.463 27.31 122.127 27.718C121.803 28.126 121.641 28.636 121.641 29.248V33.028C121.641 33.376 121.533 33.646 121.317 33.838C121.101 34.018 120.831 34.108 120.507 34.108C120.183 34.108 119.913 34.018 119.697 33.838C119.493 33.646 119.391 33.376 119.391 33.028V26.062C119.391 25.726 119.493 25.468 119.697 25.288C119.913 25.108 120.177 25.018 120.489 25.018C120.801 25.018 121.053 25.108 121.245 25.288C121.437 25.456 121.533 25.702 121.533 26.026V26.746C121.761 26.218 122.097 25.81 122.541 25.522C122.997 25.234 123.501 25.072 124.053 25.036L124.305 25.018ZM133.072 31.588C133.276 31.588 133.438 31.666 133.558 31.822C133.69 31.978 133.756 32.188 133.756 32.452C133.756 32.824 133.534 33.136 133.09 33.388C132.682 33.616 132.22 33.802 131.704 33.946C131.188 34.078 130.696 34.144 130.228 34.144C128.812 34.144 127.69 33.736 126.862 32.92C126.034 32.104 125.62 30.988 125.62 29.572C125.62 28.672 125.8 27.874 126.16 27.178C126.52 26.482 127.024 25.942 127.672 25.558C128.332 25.174 129.076 24.982 129.904 24.982C130.696 24.982 131.386 25.156 131.974 25.504C132.562 25.852 133.018 26.344 133.342 26.98C133.666 27.616 133.828 28.366 133.828 29.23C133.828 29.746 133.6 30.004 133.144 30.004H127.834C127.906 30.832 128.14 31.444 128.536 31.84C128.932 32.224 129.508 32.416 130.264 32.416C130.648 32.416 130.984 32.368 131.272 32.272C131.572 32.176 131.908 32.044 132.28 31.876C132.64 31.684 132.904 31.588 133.072 31.588ZM129.958 26.566C129.346 26.566 128.854 26.758 128.482 27.142C128.122 27.526 127.906 28.078 127.834 28.798H131.902C131.878 28.066 131.698 27.514 131.362 27.142C131.026 26.758 130.558 26.566 129.958 26.566Z" fill="white"/>
-    <path d="M51.45 16C51.23 16 51.06 15.9433 50.94 15.83C50.8267 15.71 50.77 15.54 50.77 15.32V9.63C50.77 9.41 50.8267 9.24333 50.94 9.13C51.06 9.01 51.23 8.95 51.45 8.95H53.39C54.57 8.95 55.48 9.25333 56.12 9.86C56.7667 10.4667 57.09 11.3367 57.09 12.47C57.09 13.0367 57.0067 13.54 56.84 13.98C56.68 14.4133 56.44 14.78 56.12 15.08C55.8067 15.38 55.42 15.61 54.96 15.77C54.5067 15.9233 53.9833 16 53.39 16H51.45ZM52.05 14.94H53.31C53.7233 14.94 54.08 14.89 54.38 14.79C54.6867 14.6833 54.94 14.53 55.14 14.33C55.3467 14.1233 55.5 13.8667 55.6 13.56C55.7 13.2467 55.75 12.8833 55.75 12.47C55.75 11.6433 55.5467 11.0267 55.14 10.62C54.7333 10.2133 54.1233 10.01 53.31 10.01H52.05V14.94ZM60.4972 16.1C59.9905 16.1 59.5505 15.9967 59.1772 15.79C58.8039 15.5833 58.5139 15.29 58.3072 14.91C58.1005 14.5233 57.9972 14.07 57.9972 13.55C57.9972 13.1567 58.0539 12.8067 58.1672 12.5C58.2872 12.1867 58.4572 11.92 58.6772 11.7C58.8972 11.4733 59.1605 11.3033 59.4672 11.19C59.7739 11.07 60.1172 11.01 60.4972 11.01C61.0039 11.01 61.4439 11.1133 61.8172 11.32C62.1905 11.5267 62.4805 11.82 62.6872 12.2C62.8939 12.58 62.9972 13.03 62.9972 13.55C62.9972 13.9433 62.9372 14.2967 62.8172 14.61C62.7039 14.9233 62.5372 15.1933 62.3172 15.42C62.0972 15.64 61.8339 15.81 61.5272 15.93C61.2205 16.0433 60.8772 16.1 60.4972 16.1ZM60.4972 15.15C60.7439 15.15 60.9605 15.09 61.1472 14.97C61.3339 14.85 61.4772 14.6733 61.5772 14.44C61.6839 14.2 61.7372 13.9033 61.7372 13.55C61.7372 13.0167 61.6239 12.62 61.3972 12.36C61.1705 12.0933 60.8705 11.96 60.4972 11.96C60.2505 11.96 60.0339 12.02 59.8472 12.14C59.6605 12.2533 59.5139 12.43 59.4072 12.67C59.3072 12.9033 59.2572 13.1967 59.2572 13.55C59.2572 14.0767 59.3705 14.4767 59.5972 14.75C59.8239 15.0167 60.1239 15.15 60.4972 15.15ZM65.8213 16.08C65.6546 16.08 65.5113 16.04 65.3913 15.96C65.2713 15.8733 65.1746 15.74 65.1013 15.56L63.6513 11.86C63.5913 11.7 63.5713 11.56 63.5913 11.44C63.6179 11.3133 63.6813 11.2133 63.7813 11.14C63.8813 11.0667 64.0146 11.03 64.1813 11.03C64.3279 11.03 64.4479 11.0667 64.5413 11.14C64.6346 11.2067 64.7146 11.3367 64.7813 11.53L65.9913 14.86H65.7613L67.0113 11.45C67.0646 11.3033 67.1346 11.1967 67.2213 11.13C67.3146 11.0633 67.4346 11.03 67.5813 11.03C67.7279 11.03 67.8479 11.0667 67.9413 11.14C68.0346 11.2067 68.1046 11.31 68.1513 11.45L69.3813 14.86H69.1713L70.3913 11.5C70.4579 11.32 70.5413 11.1967 70.6413 11.13C70.7479 11.0633 70.8646 11.03 70.9913 11.03C71.1513 11.03 71.2746 11.07 71.3613 11.15C71.4479 11.23 71.4979 11.3333 71.5113 11.46C71.5246 11.58 71.5013 11.7133 71.4413 11.86L70.0013 15.56C69.9346 15.7333 69.8379 15.8633 69.7113 15.95C69.5913 16.0367 69.4479 16.08 69.2813 16.08C69.1146 16.08 68.9679 16.0367 68.8413 15.95C68.7213 15.8633 68.6279 15.7333 68.5612 15.56L67.2513 12.09H67.8213L66.5413 15.55C66.4746 15.73 66.3813 15.8633 66.2613 15.95C66.1413 16.0367 65.9946 16.08 65.8213 16.08ZM73.0766 16.08C72.8766 16.08 72.7233 16.0267 72.6166 15.92C72.51 15.8067 72.4566 15.6467 72.4566 15.44V11.66C72.4566 11.4533 72.51 11.2967 72.6166 11.19C72.7233 11.0833 72.8733 11.03 73.0666 11.03C73.26 11.03 73.41 11.0833 73.5166 11.19C73.6233 11.2967 73.6766 11.4533 73.6766 11.66V12.34L73.5666 12.09C73.7133 11.7367 73.94 11.47 74.2466 11.29C74.56 11.1033 74.9133 11.01 75.3066 11.01C75.7 11.01 76.0233 11.0833 76.2766 11.23C76.53 11.3767 76.72 11.6 76.8466 11.9C76.9733 12.1933 77.0366 12.5667 77.0366 13.02V15.44C77.0366 15.6467 76.9833 15.8067 76.8766 15.92C76.77 16.0267 76.6166 16.08 76.4166 16.08C76.2166 16.08 76.06 16.0267 75.9466 15.92C75.84 15.8067 75.7866 15.6467 75.7866 15.44V13.08C75.7866 12.7 75.7133 12.4233 75.5666 12.25C75.4266 12.0767 75.2066 11.99 74.9066 11.99C74.54 11.99 74.2466 12.1067 74.0266 12.34C73.8133 12.5667 73.7066 12.87 73.7066 13.25V15.44C73.7066 15.8667 73.4966 16.08 73.0766 16.08ZM79.9763 16.1C79.4229 16.1 79.0063 15.9433 78.7263 15.63C78.4463 15.31 78.3063 14.8433 78.3063 14.23V9.5C78.3063 9.29333 78.3596 9.13667 78.4663 9.03C78.5729 8.92333 78.7263 8.87 78.9263 8.87C79.1263 8.87 79.2796 8.92333 79.3863 9.03C79.4996 9.13667 79.5563 9.29333 79.5563 9.5V14.17C79.5563 14.4767 79.6196 14.7033 79.7463 14.85C79.8796 14.9967 80.0663 15.07 80.3063 15.07C80.3596 15.07 80.4096 15.07 80.4563 15.07C80.5029 15.0633 80.5496 15.0567 80.5962 15.05C80.6896 15.0367 80.7529 15.0633 80.7863 15.13C80.8196 15.19 80.8363 15.3167 80.8363 15.51C80.8363 15.6767 80.8029 15.8067 80.7363 15.9C80.6696 15.9933 80.5596 16.05 80.4063 16.07C80.3396 16.0767 80.2696 16.0833 80.1963 16.09C80.1229 16.0967 80.0496 16.1 79.9763 16.1ZM83.6808 16.1C83.1741 16.1 82.7341 15.9967 82.3608 15.79C81.9874 15.5833 81.6974 15.29 81.4908 14.91C81.2841 14.5233 81.1808 14.07 81.1808 13.55C81.1808 13.1567 81.2374 12.8067 81.3508 12.5C81.4708 12.1867 81.6408 11.92 81.8608 11.7C82.0808 11.4733 82.3441 11.3033 82.6508 11.19C82.9574 11.07 83.3008 11.01 83.6808 11.01C84.1874 11.01 84.6274 11.1133 85.0008 11.32C85.3741 11.5267 85.6641 11.82 85.8708 12.2C86.0774 12.58 86.1808 13.03 86.1808 13.55C86.1808 13.9433 86.1208 14.2967 86.0008 14.61C85.8874 14.9233 85.7208 15.1933 85.5008 15.42C85.2808 15.64 85.0174 15.81 84.7108 15.93C84.4041 16.0433 84.0608 16.1 83.6808 16.1ZM83.6808 15.15C83.9274 15.15 84.1441 15.09 84.3308 14.97C84.5174 14.85 84.6608 14.6733 84.7608 14.44C84.8674 14.2 84.9208 13.9033 84.9208 13.55C84.9208 13.0167 84.8074 12.62 84.5808 12.36C84.3541 12.0933 84.0541 11.96 83.6808 11.96C83.4341 11.96 83.2174 12.02 83.0308 12.14C82.8441 12.2533 82.6974 12.43 82.5908 12.67C82.4908 12.9033 82.4408 13.1967 82.4408 13.55C82.4408 14.0767 82.5541 14.4767 82.7808 14.75C83.0074 15.0167 83.3074 15.15 83.6808 15.15ZM88.8225 16.1C88.4692 16.1 88.1525 16.0333 87.8725 15.9C87.5992 15.76 87.3825 15.5733 87.2225 15.34C87.0692 15.1067 86.9925 14.8433 86.9925 14.55C86.9925 14.19 87.0858 13.9067 87.2725 13.7C87.4592 13.4867 87.7625 13.3333 88.1825 13.24C88.6025 13.1467 89.1658 13.1 89.8725 13.1H90.3725V13.82H89.8825C89.4692 13.82 89.1392 13.84 88.8925 13.88C88.6458 13.92 88.4692 13.99 88.3625 14.09C88.2625 14.1833 88.2125 14.3167 88.2125 14.49C88.2125 14.71 88.2892 14.89 88.4425 15.03C88.5958 15.17 88.8092 15.24 89.0825 15.24C89.3025 15.24 89.4958 15.19 89.6625 15.09C89.8358 14.9833 89.9725 14.84 90.0725 14.66C90.1725 14.48 90.2225 14.2733 90.2225 14.04V12.89C90.2225 12.5567 90.1492 12.3167 90.0025 12.17C89.8558 12.0233 89.6092 11.95 89.2625 11.95C89.0692 11.95 88.8592 11.9733 88.6325 12.02C88.4125 12.0667 88.1792 12.1467 87.9325 12.26C87.8058 12.32 87.6925 12.3367 87.5925 12.31C87.4992 12.2833 87.4258 12.23 87.3725 12.15C87.3192 12.0633 87.2925 11.97 87.2925 11.87C87.2925 11.77 87.3192 11.6733 87.3725 11.58C87.4258 11.48 87.5158 11.4067 87.6425 11.36C87.9492 11.2333 88.2425 11.1433 88.5225 11.09C88.8092 11.0367 89.0692 11.01 89.3025 11.01C89.7825 11.01 90.1758 11.0833 90.4825 11.23C90.7958 11.3767 91.0292 11.6 91.1825 11.9C91.3358 12.1933 91.4125 12.5733 91.4125 13.04V15.44C91.4125 15.6467 91.3625 15.8067 91.2625 15.92C91.1625 16.0267 91.0192 16.08 90.8325 16.08C90.6458 16.08 90.4992 16.0267 90.3925 15.92C90.2925 15.8067 90.2425 15.6467 90.2425 15.44V14.96H90.3225C90.2758 15.1933 90.1825 15.3967 90.0425 15.57C89.9092 15.7367 89.7392 15.8667 89.5325 15.96C89.3258 16.0533 89.0892 16.1 88.8225 16.1ZM94.5613 16.1C94.1346 16.1 93.7579 15.9967 93.4313 15.79C93.1113 15.5833 92.8613 15.29 92.6813 14.91C92.5013 14.5233 92.4113 14.07 92.4113 13.55C92.4113 13.0233 92.5013 12.5733 92.6813 12.2C92.8613 11.82 93.1113 11.5267 93.4313 11.32C93.7579 11.1133 94.1346 11.01 94.5613 11.01C94.9679 11.01 95.3213 11.11 95.6213 11.31C95.9279 11.51 96.1346 11.7733 96.2413 12.1H96.1313V9.5C96.1313 9.29333 96.1846 9.13667 96.2913 9.03C96.3979 8.92333 96.5513 8.87 96.7513 8.87C96.9513 8.87 97.1046 8.92333 97.2113 9.03C97.3246 9.13667 97.3813 9.29333 97.3813 9.5V15.44C97.3813 15.6467 97.3279 15.8067 97.2213 15.92C97.1146 16.0267 96.9613 16.08 96.7613 16.08C96.5613 16.08 96.4079 16.0267 96.3013 15.92C96.1946 15.8067 96.1413 15.6467 96.1413 15.44V14.66L96.2513 14.96C96.1513 15.3067 95.9479 15.5833 95.6413 15.79C95.3346 15.9967 94.9746 16.1 94.5613 16.1ZM94.9113 15.15C95.1579 15.15 95.3746 15.09 95.5613 14.97C95.7479 14.85 95.8913 14.6733 95.9913 14.44C96.0979 14.2 96.1513 13.9033 96.1513 13.55C96.1513 13.0167 96.0379 12.62 95.8113 12.36C95.5846 12.0933 95.2846 11.96 94.9113 11.96C94.6646 11.96 94.4479 12.02 94.2613 12.14C94.0746 12.2533 93.9279 12.43 93.8213 12.67C93.7213 12.9033 93.6713 13.1967 93.6713 13.55C93.6713 14.0767 93.7846 14.4767 94.0113 14.75C94.2379 15.0167 94.5379 15.15 94.9113 15.15ZM103.622 16.1C103.116 16.1 102.676 15.9967 102.302 15.79C101.929 15.5833 101.639 15.29 101.432 14.91C101.226 14.5233 101.122 14.07 101.122 13.55C101.122 13.1567 101.179 12.8067 101.292 12.5C101.412 12.1867 101.582 11.92 101.802 11.7C102.022 11.4733 102.286 11.3033 102.592 11.19C102.899 11.07 103.242 11.01 103.622 11.01C104.129 11.01 104.569 11.1133 104.942 11.32C105.316 11.5267 105.606 11.82 105.812 12.2C106.019 12.58 106.122 13.03 106.122 13.55C106.122 13.9433 106.062 14.2967 105.942 14.61C105.829 14.9233 105.662 15.1933 105.442 15.42C105.222 15.64 104.959 15.81 104.652 15.93C104.346 16.0433 104.002 16.1 103.622 16.1ZM103.622 15.15C103.869 15.15 104.086 15.09 104.272 14.97C104.459 14.85 104.602 14.6733 104.702 14.44C104.809 14.2 104.862 13.9033 104.862 13.55C104.862 13.0167 104.749 12.62 104.522 12.36C104.296 12.0933 103.996 11.96 103.622 11.96C103.376 11.96 103.159 12.02 102.972 12.14C102.786 12.2533 102.639 12.43 102.532 12.67C102.432 12.9033 102.382 13.1967 102.382 13.55C102.382 14.0767 102.496 14.4767 102.722 14.75C102.949 15.0167 103.249 15.15 103.622 15.15ZM107.774 16.08C107.574 16.08 107.421 16.0267 107.314 15.92C107.207 15.8067 107.154 15.6467 107.154 15.44V11.66C107.154 11.4533 107.207 11.2967 107.314 11.19C107.421 11.0833 107.571 11.03 107.764 11.03C107.957 11.03 108.107 11.0833 108.214 11.19C108.321 11.2967 108.374 11.4533 108.374 11.66V12.34L108.264 12.09C108.411 11.7367 108.637 11.47 108.944 11.29C109.257 11.1033 109.611 11.01 110.004 11.01C110.397 11.01 110.721 11.0833 110.974 11.23C111.227 11.3767 111.417 11.6 111.544 11.9C111.671 12.1933 111.734 12.5667 111.734 13.02V15.44C111.734 15.6467 111.681 15.8067 111.574 15.92C111.467 16.0267 111.314 16.08 111.114 16.08C110.914 16.08 110.757 16.0267 110.644 15.92C110.537 15.8067 110.484 15.6467 110.484 15.44V13.08C110.484 12.7 110.411 12.4233 110.264 12.25C110.124 12.0767 109.904 11.99 109.604 11.99C109.237 11.99 108.944 12.1067 108.724 12.34C108.511 12.5667 108.404 12.87 108.404 13.25V15.44C108.404 15.8667 108.194 16.08 107.774 16.08ZM117.958 16.1C117.525 16.1 117.162 16.0267 116.868 15.88C116.582 15.7333 116.368 15.52 116.228 15.24C116.088 14.9533 116.018 14.6 116.018 14.18V12.05H115.498C115.338 12.05 115.215 12.01 115.128 11.93C115.042 11.8433 114.998 11.7267 114.998 11.58C114.998 11.4267 115.042 11.31 115.128 11.23C115.215 11.15 115.338 11.11 115.498 11.11H116.018V10.2C116.018 9.99333 116.072 9.83667 116.178 9.73C116.292 9.62333 116.448 9.57 116.648 9.57C116.848 9.57 117.002 9.62333 117.108 9.73C117.215 9.83667 117.268 9.99333 117.268 10.2V11.11H118.328C118.488 11.11 118.612 11.15 118.698 11.23C118.785 11.31 118.828 11.4267 118.828 11.58C118.828 11.7267 118.785 11.8433 118.698 11.93C118.612 12.01 118.488 12.05 118.328 12.05H117.268V14.11C117.268 14.43 117.338 14.67 117.478 14.83C117.618 14.99 117.845 15.07 118.158 15.07C118.272 15.07 118.372 15.06 118.458 15.04C118.545 15.02 118.622 15.0067 118.688 15C118.768 14.9933 118.835 15.02 118.888 15.08C118.942 15.1333 118.968 15.2467 118.968 15.42C118.968 15.5533 118.945 15.6733 118.898 15.78C118.858 15.88 118.782 15.95 118.668 15.99C118.582 16.0167 118.468 16.04 118.328 16.06C118.188 16.0867 118.065 16.1 117.958 16.1ZM120.303 16.08C120.103 16.08 119.95 16.0267 119.843 15.92C119.737 15.8067 119.683 15.6467 119.683 15.44V9.5C119.683 9.29333 119.737 9.13667 119.843 9.03C119.95 8.92333 120.103 8.87 120.303 8.87C120.503 8.87 120.657 8.92333 120.763 9.03C120.877 9.13667 120.933 9.29333 120.933 9.5V12.09H120.793C120.94 11.7367 121.167 11.47 121.473 11.29C121.787 11.1033 122.14 11.01 122.533 11.01C122.927 11.01 123.25 11.0833 123.503 11.23C123.757 11.3767 123.947 11.6 124.073 11.9C124.2 12.1933 124.263 12.5667 124.263 13.02V15.44C124.263 15.6467 124.21 15.8067 124.103 15.92C123.997 16.0267 123.843 16.08 123.643 16.08C123.443 16.08 123.287 16.0267 123.173 15.92C123.067 15.8067 123.013 15.6467 123.013 15.44V13.08C123.013 12.7 122.94 12.4233 122.793 12.25C122.653 12.0767 122.433 11.99 122.133 11.99C121.767 11.99 121.473 12.1067 121.253 12.34C121.04 12.5667 120.933 12.87 120.933 13.25V15.44C120.933 15.8667 120.723 16.08 120.303 16.08ZM127.943 16.1C127.389 16.1 126.913 15.9967 126.513 15.79C126.113 15.5833 125.803 15.29 125.583 14.91C125.369 14.53 125.263 14.08 125.263 13.56C125.263 13.0533 125.366 12.61 125.573 12.23C125.786 11.85 126.076 11.5533 126.443 11.34C126.816 11.12 127.239 11.01 127.713 11.01C128.059 11.01 128.369 11.0667 128.643 11.18C128.923 11.2933 129.159 11.4567 129.353 11.67C129.553 11.8833 129.703 12.1433 129.803 12.45C129.909 12.75 129.963 13.09 129.963 13.47C129.963 13.59 129.919 13.6833 129.833 13.75C129.753 13.81 129.636 13.84 129.483 13.84H126.273V13.12H129.083L128.923 13.27C128.923 12.9633 128.876 12.7067 128.783 12.5C128.696 12.2933 128.566 12.1367 128.393 12.03C128.226 11.9167 128.016 11.86 127.763 11.86C127.483 11.86 127.243 11.9267 127.043 12.06C126.849 12.1867 126.699 12.37 126.593 12.61C126.493 12.8433 126.443 13.1233 126.443 13.45V13.52C126.443 14.0667 126.569 14.4767 126.823 14.75C127.083 15.0167 127.463 15.15 127.963 15.15C128.136 15.15 128.329 15.13 128.543 15.09C128.763 15.0433 128.969 14.9667 129.163 14.86C129.303 14.78 129.426 14.7467 129.533 14.76C129.639 14.7667 129.723 14.8033 129.783 14.87C129.849 14.9367 129.889 15.02 129.903 15.12C129.916 15.2133 129.896 15.31 129.843 15.41C129.796 15.51 129.713 15.5967 129.593 15.67C129.359 15.8167 129.089 15.9267 128.783 16C128.483 16.0667 128.203 16.1 127.943 16.1Z" fill="white"/>
-    <path d="M35.0457 22.8903C35.0123 19.3528 38.0969 17.6319 38.2377 17.552C36.4907 15.1339 33.7822 14.8038 32.8309 14.7774C30.5568 14.55 28.3504 16.069 27.1925 16.069C26.0106 16.069 24.2265 14.7994 22.3036 14.8363C19.8285 14.8731 17.5146 16.2323 16.2447 18.344C13.6242 22.6498 15.5777 28.9768 18.088 32.4572C19.344 34.1615 20.8113 36.065 22.7324 35.9983C24.611 35.9245 25.3131 34.8613 27.5807 34.8613C29.8269 34.8613 30.4864 35.9983 32.4455 35.9553C34.4621 35.9245 35.7311 34.2431 36.9427 32.5231C38.3942 30.5695 38.9768 28.6458 39 28.5465C38.9528 28.5316 35.0836 27.1303 35.0447 22.8903H35.0457ZM31.3451 12.4884C32.3557 11.2882 33.0476 9.65682 32.855 8C31.3933 8.06146 29.5638 8.96055 28.5106 10.1336C27.5779 11.1688 26.7461 12.8634 26.96 14.4561C28.6033 14.572 30.2901 13.6694 31.3442 12.4884H31.3451Z" fill="white"/>
-  </svg>
-`;
-
-/**
- * @function IconArrowDown
- * @returns {string} HTML
- */
-
-export const IconArrowDown = () => `
-  <svg viewBox="0 0 9 6" xmlns="http://www.w3.org/2000/svg">
-    <path d="M4.5 6L0 1.49973L1.50071 0L4.5 3.00053L7.49929 0L9 1.49973L4.5 6Z" fill="currentColor" />
-  </svg>
-`;
-
-/**
- * @function IconClose
- * @returns {string} HTML
- */
-
-export const IconClose = () => `
-  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <rect width="31.8192" height="2.12128" transform="matrix(0.707107 -0.707106 0.707107 0.707106 0.000488281 22.4996)" fill="white"/>
-    <rect width="31.8192" height="2.12128" transform="matrix(0.707107 0.707106 -0.707107 0.707106 1.5 0.000488281)" fill="white"/>
-  </svg>
-`;
-
-/**
- * @function IconDeveloper
- * @returns {string} HTML
- */
-
-export const IconDeveloper = () => `
-  <svg width="130" height="46" viewBox="0 0 130 46" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M53.0156 0L65.3856 29.3943L14.7011 46L0 3.19635L53.0156 0ZM33.2519 15.3286C29.9532 17.0524 28.2551 20.0227 28.5202 24.0077C28.7802 27.9273 30.8366 30.439 34.2565 31.7642C37.0116 32.8309 39.8391 32.9759 42.7057 32.541C43.4813 32.4228 44.2414 32.1983 44.9704 32.0045L45.404 31.8926V27.6555C43.4111 27.8587 41.6007 28.2661 39.8118 28.1747C37.2125 28.0423 35.5026 26.0947 35.3733 23.5925C35.1864 19.9604 37.7584 17.7489 41.5335 18.3816C42.5344 18.5486 43.5086 18.8976 44.5383 19.2112L44.983 19.342V14.5998C40.9256 13.3519 36.9902 13.3747 33.2519 15.3286ZM76.1645 25.7898C76.4969 25.7898 76.8145 25.8387 77.1196 25.9364C77.4224 26.034 77.6898 26.2326 77.9188 26.532L77.2488 27.1835C77.1454 27.0126 76.9992 26.8747 76.8101 26.7707C76.6202 26.6667 76.4053 26.6147 76.1645 26.6147C75.9355 26.6147 75.7449 26.6462 75.5935 26.7108C75.4414 26.7754 75.3195 26.8582 75.2279 26.959C75.1363 27.0598 75.0713 27.1686 75.0344 27.2844C74.9967 27.401 74.979 27.5081 74.979 27.6058C74.979 27.8808 75.0521 28.0919 75.202 28.2385C75.3513 28.385 75.5359 28.5048 75.756 28.5962C75.9769 28.6884 76.2177 28.7695 76.4792 28.8436C76.7399 28.9176 76.9807 29.02 77.2008 29.1508C77.4217 29.2832 77.607 29.4605 77.7548 29.6834C77.9055 29.9064 77.9793 30.2136 77.9793 30.6052C77.9793 30.9172 77.9232 31.1985 77.8124 31.4482C77.6994 31.6996 77.5494 31.9107 77.3604 32.0817C77.1713 32.2534 76.9482 32.385 76.6897 32.4756C76.4311 32.5686 76.1556 32.6143 75.8646 32.6143C75.4731 32.6143 75.1038 32.5402 74.7552 32.3937C74.4043 32.2471 74.1266 32.021 73.9205 31.7153L74.6001 31.0913C74.7493 31.33 74.9398 31.5057 75.1718 31.6184C75.4044 31.7319 75.6438 31.7886 75.8897 31.7886C76.0278 31.7886 76.1674 31.7649 76.3107 31.7192C76.4548 31.6735 76.584 31.605 76.7022 31.5136C76.8197 31.4214 76.9164 31.308 76.9903 31.174C77.0642 31.0393 77.1026 30.8825 77.1026 30.7061C77.1026 30.4555 77.0272 30.2641 76.8788 30.1325C76.7303 30.0017 76.5449 29.8914 76.324 29.8016C76.1025 29.7141 75.8631 29.6283 75.6017 29.5455C75.3409 29.4636 75.0994 29.3486 74.8793 29.202C74.6577 29.0555 74.473 28.8577 74.3253 28.6104C74.1754 28.3622 74.1015 28.0273 74.1015 27.6058C74.1015 27.416 74.1399 27.2143 74.2167 27.0008C74.295 26.7865 74.4161 26.5911 74.5831 26.413C74.7493 26.2357 74.9627 26.0876 75.2235 25.9687C75.4835 25.8489 75.7981 25.7898 76.1645 25.7898ZM120.058 27.9911C120.419 27.9911 120.724 28.0588 120.973 28.1928C121.223 28.3267 121.427 28.5032 121.588 28.7199C121.749 28.9373 121.865 29.1823 121.937 29.4541C122.008 29.726 122.044 30.0033 122.044 30.2846V30.5776H118.776C118.776 30.7667 118.815 30.9401 118.893 31.0961C118.969 31.252 119.072 31.3852 119.197 31.4947C119.325 31.605 119.47 31.6909 119.635 31.7523C119.802 31.813 119.975 31.8437 120.153 31.8437C120.393 31.8437 120.602 31.7839 120.78 31.6649C120.957 31.5451 121.121 31.3883 121.27 31.1922L121.854 31.6696C121.424 32.2629 120.823 32.5591 120.049 32.5591C119.728 32.5591 119.436 32.5008 119.176 32.385C118.915 32.2684 118.694 32.1085 118.514 31.9036C118.333 31.6988 118.194 31.4569 118.096 31.1788C118 30.9007 117.951 30.5997 117.951 30.2751C117.951 29.9513 118.004 29.6503 118.109 29.3714C118.217 29.0933 118.362 28.8514 118.548 28.6466C118.734 28.4417 118.957 28.2818 119.214 28.1652C119.473 28.0494 119.754 27.9911 120.058 27.9911ZM110.977 27.9911C111.338 27.9911 111.643 28.0588 111.892 28.1928C112.142 28.3267 112.347 28.5032 112.508 28.7199C112.668 28.9373 112.784 29.1823 112.855 29.4541C112.927 29.726 112.963 30.0033 112.963 30.2846V30.5776H109.695C109.695 30.7667 109.733 30.9401 109.812 31.0961C109.888 31.252 109.99 31.3852 110.117 31.4947C110.242 31.605 110.388 31.6909 110.555 31.7523C110.722 31.813 110.893 31.8437 111.071 31.8437C111.312 31.8437 111.521 31.7839 111.699 31.6649C111.876 31.5451 112.04 31.3883 112.189 31.1922L112.774 31.6696C112.344 32.2629 111.742 32.5591 110.968 32.5591C110.646 32.5591 110.355 32.5008 110.095 32.385C109.834 32.2684 109.614 32.1085 109.433 31.9036C109.252 31.6988 109.113 31.4569 109.015 31.1788C108.919 30.9007 108.869 30.5997 108.869 30.2751C108.869 29.9513 108.922 29.6503 109.029 29.3714C109.135 29.0933 109.28 28.8514 109.467 28.6466C109.654 28.4417 109.876 28.2818 110.134 28.1652C110.391 28.0494 110.673 27.9911 110.977 27.9911ZM99.3407 27.9911C99.7019 27.9911 100.007 28.0588 100.257 28.1928C100.506 28.3267 100.711 28.5032 100.871 28.7199C101.032 28.9373 101.148 29.1823 101.221 29.4541C101.291 29.726 101.327 30.0033 101.327 30.2846V30.5776H98.0592C98.0592 30.7667 98.0983 30.9401 98.1759 31.0961C98.2527 31.252 98.3546 31.3852 98.4809 31.4947C98.6065 31.605 98.7535 31.6909 98.9189 31.7523C99.0851 31.813 99.2572 31.8437 99.436 31.8437C99.6768 31.8437 99.8858 31.7839 100.063 31.6649C100.24 31.5451 100.403 31.3883 100.554 31.1922L101.137 31.6696C100.708 32.2629 100.107 32.5591 99.3318 32.5591C99.0105 32.5591 98.7202 32.5008 98.4588 32.385C98.198 32.2684 97.9772 32.1085 97.7969 31.9036C97.6167 31.6988 97.4771 31.4569 97.3796 31.1788C97.2829 30.9007 97.2341 30.5997 97.2341 30.2751C97.2341 29.9513 97.2858 29.6503 97.3929 29.3714C97.4993 29.0933 97.6455 28.8514 97.8317 28.6466C98.0178 28.4417 98.2394 28.2818 98.4979 28.1652C98.7564 28.0494 99.0379 27.9911 99.3407 27.9911ZM95.2649 27.9998C95.5626 27.9998 95.8484 28.0565 96.1202 28.1699C96.3935 28.2826 96.6299 28.4551 96.8301 28.6876L96.2451 29.2564C96.1018 29.0972 95.9518 28.9767 95.7945 28.894C95.6357 28.812 95.4481 28.7703 95.2302 28.7703C95.0182 28.7703 94.8335 28.812 94.6762 28.894C94.5174 28.9767 94.3859 29.0909 94.2796 29.2335C94.1739 29.3769 94.0934 29.5424 94.0395 29.7299C93.9841 29.915 93.9583 30.1096 93.9583 30.3113C93.9583 30.513 93.9885 30.7053 94.0513 30.8849C94.1148 31.0661 94.2035 31.2237 94.3187 31.3576C94.4332 31.4916 94.5706 31.5971 94.7309 31.6736C94.8919 31.75 95.0728 31.7878 95.2738 31.7878C95.4909 31.7878 95.6771 31.7468 95.8314 31.6641C95.9865 31.5822 96.1306 31.4608 96.2613 31.3017L96.8131 31.8886C96.6122 32.1274 96.3788 32.2991 96.1114 32.4031C95.8455 32.5063 95.5626 32.5583 95.2649 32.5583C94.9488 32.5583 94.6607 32.504 94.4007 32.3937C94.1392 32.2834 93.9147 32.1289 93.7256 31.9304C93.5365 31.7319 93.3895 31.4931 93.2868 31.215C93.1834 30.9369 93.1317 30.6296 93.1317 30.2932C93.1317 29.9568 93.1834 29.648 93.2868 29.3667C93.3902 29.0854 93.535 28.8436 93.7211 28.6419C93.9073 28.4402 94.1311 28.2826 94.3918 28.1699C94.6526 28.0565 94.9428 27.9998 95.2649 27.9998ZM80.8689 27.9911C81.2293 27.9911 81.5351 28.0588 81.7848 28.1928C82.0344 28.3267 82.239 28.5032 82.3993 28.7199C82.5596 28.9373 82.6763 29.1823 82.748 29.4541C82.8189 29.726 82.8551 30.0033 82.8551 30.2846V30.5776H79.5873C79.5873 30.7667 79.6257 30.9401 79.7033 31.0961C79.7809 31.252 79.8821 31.3852 80.0091 31.4947C80.1347 31.605 80.2809 31.6909 80.4471 31.7523C80.6133 31.813 80.7847 31.8437 80.9642 31.8437C81.2042 31.8437 81.4132 31.7839 81.592 31.6649C81.7693 31.5451 81.9318 31.3883 82.081 31.1922L82.666 31.6696C82.2361 32.2629 81.6341 32.5591 80.86 32.5591C80.538 32.5591 80.2484 32.5008 79.9869 32.385C79.7262 32.2684 79.5053 32.1085 79.3251 31.9036C79.1449 31.6988 79.0046 31.4569 78.9078 31.1788C78.8103 30.9007 78.7615 30.5997 78.7615 30.2751C78.7615 29.9513 78.814 29.6503 78.9211 29.3714C79.0267 29.0933 79.173 28.8514 79.3598 28.6466C79.546 28.4417 79.7676 28.2818 80.0261 28.1652C80.2846 28.0494 80.5653 27.9911 80.8689 27.9911ZM11.6934 14.3832L10.4694 14.3863V32.5024H16.8697V27.1197L18.0294 27.074C19.5251 27.011 20.8163 26.9188 21.8924 26.5674C23.448 26.0585 24.5057 25.4101 25.6787 24.2858C28.3326 21.7394 27.6752 17.3345 24.5235 15.5563C23.496 14.9772 22.2972 14.4848 21.159 14.4383C19.2178 14.3588 17.2738 14.3517 15.2964 14.3611L11.6934 14.3832ZM114.297 28.1006L115.518 31.4301L116.688 28.1006H117.513L115.913 32.4496H115.062L113.402 28.1006H114.297ZM87.3711 28.1006L88.5921 31.4301L89.7621 28.1006H90.5871L88.9872 32.4496H88.1363L86.4766 28.1006H87.3711ZM85.7978 27.9903C85.9234 27.9903 86.0386 28.0084 86.1405 28.0455L86.1065 28.9357C86.0438 28.9168 85.981 28.9019 85.9174 28.8893C85.8547 28.8774 85.7919 28.8711 85.7284 28.8711C85.3502 28.8711 85.0606 28.9846 84.8597 29.2107C84.6603 29.4368 84.5584 29.7882 84.5584 30.2664V32.4488H83.785V28.1006H84.5584V28.7703H84.5753C84.6278 28.6545 84.6965 28.5481 84.7822 28.4536C84.8686 28.359 84.9646 28.2779 85.0702 28.2109C85.1766 28.1431 85.2926 28.0896 85.4189 28.0502C85.5452 28.0108 85.6707 27.9903 85.7978 27.9903ZM123.807 25.5141V32.4488H123.034V25.5141H123.807ZM105.671 25.9545V31.6231H108.466V32.4488H104.845V25.9545H105.671ZM92.1006 28.1006V32.4488H91.3265V28.1006H92.1006ZM129.991 25.9545V32.4488H129.165V26.9637L128.047 28.0644L127.566 27.4498L129.261 25.9545H129.991ZM54.7662 14.2674H48.2669V32.3803H54.7662V14.2674ZM80.8511 28.6513C80.662 28.6513 80.4892 28.6899 80.3304 28.7663C80.1731 28.8428 80.0401 28.942 79.9315 29.0642C79.8222 29.1871 79.7373 29.3226 79.6767 29.4731C79.6176 29.622 79.5881 29.7701 79.5881 29.9174H82.03C82.0241 29.7338 81.9968 29.566 81.9488 29.4132C81.8985 29.2603 81.8261 29.1272 81.7286 29.0137C81.6311 28.9003 81.5093 28.812 81.3638 28.7482C81.216 28.6836 81.0461 28.6513 80.8511 28.6513ZM120.04 28.6513C119.85 28.6513 119.678 28.6899 119.52 28.7663C119.362 28.8428 119.229 28.942 119.121 29.0642C119.011 29.1871 118.926 29.3226 118.867 29.4731C118.806 29.622 118.776 29.7701 118.776 29.9174H121.218C121.212 29.7338 121.185 29.566 121.137 29.4132C121.088 29.2603 121.015 29.1272 120.918 29.0137C120.82 28.9003 120.698 28.812 120.552 28.7482C120.406 28.6836 120.235 28.6513 120.04 28.6513ZM110.959 28.6513C110.77 28.6513 110.596 28.6899 110.439 28.7663C110.281 28.8428 110.148 28.942 110.039 29.0642C109.93 29.1871 109.846 29.3226 109.785 29.4731C109.724 29.622 109.695 29.7701 109.695 29.9174H112.138C112.132 29.7338 112.104 29.566 112.056 29.4132C112.007 29.2603 111.934 29.1272 111.837 29.0137C111.739 28.9003 111.618 28.812 111.471 28.7482C111.324 28.6836 111.154 28.6513 110.959 28.6513ZM99.323 28.6513C99.1339 28.6513 98.961 28.6899 98.803 28.7663C98.6449 28.8428 98.5127 28.942 98.4034 29.0642C98.2948 29.1871 98.2091 29.3226 98.1493 29.4731C98.0894 29.622 98.0592 29.7701 98.0592 29.9174H100.502C100.496 29.7338 100.469 29.566 100.42 29.4132C100.371 29.2603 100.299 29.1272 100.2 29.0137C100.104 28.9003 99.9811 28.812 99.8356 28.7482C99.6879 28.6836 99.518 28.6513 99.323 28.6513ZM91.7135 25.9364C91.8672 25.9364 92.0009 25.9962 92.1131 26.1152C92.2247 26.2342 92.2801 26.376 92.2801 26.5414C92.2801 26.7061 92.2247 26.8487 92.1131 26.9677C92.0009 27.0874 91.8672 27.1473 91.7135 27.1473C91.5584 27.1473 91.4247 27.0874 91.3132 26.9677C91.2009 26.8487 91.1455 26.7061 91.1455 26.5414C91.1455 26.376 91.2009 26.2342 91.3132 26.1152C91.4247 25.9962 91.5584 25.9364 91.7135 25.9364ZM20.4787 19.1552C20.8776 19.743 20.9256 21.013 20.5969 21.6701C19.9646 22.9362 18.7939 23.0142 17.5301 23.0008L16.9185 22.9921V18.213C18.3071 17.9293 19.661 17.9506 20.4787 19.1552ZM87.3186 14.0869C87.651 14.0869 87.9694 14.1358 88.273 14.2335C88.5765 14.3312 88.8439 14.5297 89.0729 14.8291L88.4022 15.4807C88.2988 15.3097 88.1526 15.1718 87.9635 15.0678C87.7744 14.9638 87.5594 14.9118 87.3186 14.9118C87.0897 14.9118 86.8991 14.9441 86.7469 15.0079C86.5948 15.0726 86.4736 15.1553 86.3813 15.2561C86.2897 15.357 86.2262 15.4657 86.1878 15.5815C86.1509 15.6981 86.1317 15.8053 86.1317 15.903C86.1317 16.1787 86.2063 16.3891 86.3547 16.5356C86.5047 16.6821 86.6893 16.8019 86.9102 16.8933C87.1303 16.9855 87.3711 17.0674 87.6326 17.1407C87.8926 17.2147 88.1348 17.3172 88.3542 17.4487C88.5758 17.5803 88.7605 17.7576 88.9089 17.9805C89.0589 18.2035 89.1327 18.5108 89.1327 18.9023C89.1327 19.2143 89.0773 19.4956 88.9651 19.7461C88.8543 19.9967 88.7029 20.2078 88.5145 20.3788C88.3239 20.5505 88.1009 20.6821 87.8431 20.7735C87.5846 20.8657 87.3098 20.9114 87.0173 20.9114C86.628 20.9114 86.258 20.8373 85.9079 20.6908C85.5585 20.5442 85.28 20.3181 85.0739 20.0132L85.7535 19.3884C85.9019 19.6272 86.0932 19.8029 86.3244 19.9163C86.5571 20.029 86.7972 20.0857 87.0431 20.0857C87.1813 20.0857 87.3216 20.0629 87.4642 20.0164C87.6082 19.9707 87.7389 19.9021 87.8556 19.8115C87.9738 19.7186 88.0698 19.6051 88.1437 19.4712C88.219 19.3372 88.256 19.1805 88.256 19.0032C88.256 18.7526 88.1821 18.562 88.0322 18.4296C87.883 18.2988 87.6983 18.1885 87.4782 18.0995C87.2566 18.0113 87.0158 17.9262 86.7543 17.8427C86.4943 17.7607 86.2535 17.6457 86.0327 17.4992C85.8118 17.3526 85.6272 17.1549 85.4787 16.9083C85.3295 16.6593 85.2542 16.3252 85.2542 15.903C85.2542 15.7139 85.2933 15.5114 85.3701 15.2979C85.4477 15.0836 85.5703 14.8882 85.7365 14.7101C85.9019 14.5329 86.1154 14.3848 86.3769 14.2658C86.6376 14.146 86.9515 14.0869 87.3186 14.0869ZM82.5374 14.0869C82.8698 14.0869 83.1874 14.1358 83.491 14.2335C83.7953 14.3312 84.062 14.5297 84.2917 14.8291L83.6218 15.4807C83.5176 15.3097 83.3721 15.1718 83.183 15.0678C82.9924 14.9638 82.7775 14.9118 82.5374 14.9118C82.3085 14.9118 82.1172 14.9441 81.9657 15.0079C81.8136 15.0726 81.6917 15.1553 81.6009 15.2561C81.5085 15.357 81.4443 15.4657 81.4073 15.5815C81.3697 15.6981 81.3512 15.8053 81.3512 15.903C81.3512 16.1787 81.4251 16.3891 81.5743 16.5356C81.7235 16.6821 81.9081 16.8019 82.1283 16.8933C82.3491 16.9855 82.5899 17.0674 82.8514 17.1407C83.1114 17.2147 83.3536 17.3172 83.573 17.4487C83.7946 17.5803 83.9793 17.7576 84.127 17.9805C84.2784 18.2035 84.3508 18.5108 84.3508 18.9023C84.3508 19.2143 84.2961 19.4956 84.1846 19.7461C84.0716 19.9967 83.9217 20.2078 83.7326 20.3788C83.5427 20.5505 83.3197 20.6821 83.0611 20.7735C82.8034 20.8657 82.5286 20.9114 82.2361 20.9114C81.8468 20.9114 81.4768 20.8373 81.1274 20.6908C80.7765 20.5442 80.4996 20.3181 80.2927 20.0132L80.9723 19.3884C81.1215 19.6272 81.3121 19.8029 81.5433 19.9163C81.7767 20.029 82.016 20.0857 82.2627 20.0857C82.4001 20.0857 82.5404 20.0629 82.6837 20.0164C82.827 19.9707 82.957 19.9021 83.0752 19.8115C83.1919 19.7186 83.2879 19.6051 83.3632 19.4712C83.4371 19.3372 83.474 19.1805 83.474 19.0032C83.474 18.7526 83.4002 18.562 83.251 18.4296C83.1018 18.2988 82.9171 18.1885 82.697 18.0995C82.4754 18.0113 82.2346 17.9262 81.9731 17.8427C81.7124 17.7607 81.4723 17.6457 81.2507 17.4992C81.0299 17.3526 80.8452 17.1549 80.6968 16.9083C80.5476 16.6593 80.473 16.3252 80.473 15.903C80.473 15.7139 80.5121 15.5114 80.5889 15.2979C80.6665 15.0836 80.7884 14.8882 80.9553 14.7101C81.1215 14.5329 81.335 14.3848 81.5964 14.2658C81.8564 14.146 82.1703 14.0869 82.5374 14.0869ZM95.5397 14.0869C95.9415 14.0869 96.3352 14.1697 96.7222 14.3351C97.1093 14.4998 97.4261 14.7527 97.6728 15.0962L96.9844 15.6556C96.8131 15.4168 96.5952 15.2341 96.3307 15.1056C96.0671 14.9772 95.7945 14.9126 95.5145 14.9126C95.1696 14.9126 94.8587 14.9819 94.5809 15.119C94.3025 15.2569 94.0646 15.4444 93.8667 15.6831C93.6694 15.9219 93.5165 16.2016 93.4065 16.523C93.2986 16.8437 93.244 17.1872 93.244 17.5543C93.244 17.897 93.2957 18.2208 93.3991 18.5273C93.5025 18.833 93.6502 19.1017 93.8452 19.3341C94.0402 19.5665 94.2781 19.7501 94.5595 19.8848C94.8409 20.0187 95.1578 20.0865 95.5145 20.0865C95.8632 20.0865 96.1705 20.0101 96.4342 19.8572C96.6971 19.7044 96.9217 19.4901 97.1048 19.2151L97.8006 19.7745L97.7297 19.869C97.6861 19.9234 97.63 19.988 97.5606 20.0636C97.4461 20.1889 97.2947 20.3142 97.1048 20.4395C96.915 20.5647 96.6875 20.675 96.4209 20.7696C96.155 20.8641 95.8462 20.9114 95.4976 20.9114C95.0152 20.9114 94.5817 20.8137 94.1946 20.6183C93.8068 20.4229 93.4781 20.1661 93.2048 19.8478C92.9323 19.5295 92.7247 19.1702 92.5814 18.77C92.4381 18.3697 92.3672 17.964 92.3672 17.5543C92.3672 17.0532 92.444 16.59 92.5992 16.1645C92.7528 15.7391 92.9692 15.3727 93.2477 15.0639C93.5261 14.755 93.8607 14.5155 94.25 14.3438C94.6393 14.1728 95.0699 14.0869 95.5397 14.0869ZM107.64 15.1781V16.3993H108.749V17.1139H107.64V19.0859C107.64 19.208 107.643 19.3301 107.649 19.4483C107.654 19.5673 107.675 19.6744 107.713 19.7698C107.749 19.8643 107.808 19.9407 107.885 19.999C107.962 20.0566 108.076 20.0857 108.224 20.0857C108.316 20.0857 108.41 20.0763 108.509 20.0581C108.605 20.04 108.695 20.0061 108.775 19.9573V20.7097C108.683 20.7648 108.564 20.8027 108.418 20.8239C108.272 20.8452 108.158 20.8562 108.079 20.8562C107.78 20.8562 107.549 20.8121 107.386 20.7231C107.223 20.6348 107.102 20.5198 107.025 20.3796C106.948 20.2386 106.901 20.081 106.887 19.9069C106.873 19.7335 106.865 19.5571 106.865 19.379V17.1139H105.971V16.3985H106.865V15.1781H107.64ZM125.604 13.812V20.7459H124.83V20.1235H124.813C124.663 20.3552 124.458 20.5356 124.198 20.6632C123.937 20.7916 123.669 20.8562 123.394 20.8562C123.078 20.8562 122.794 20.7979 122.538 20.6829C122.283 20.5655 122.065 20.404 121.885 20.196C121.704 19.988 121.565 19.7469 121.468 19.4712C121.37 19.1962 121.321 18.8968 121.321 18.573C121.321 18.2484 121.37 17.9475 121.468 17.6686C121.565 17.3904 121.704 17.1486 121.885 16.9445C122.065 16.7397 122.283 16.5789 122.538 16.4623C122.794 16.3465 123.078 16.289 123.394 16.289C123.687 16.289 123.962 16.3544 124.219 16.486C124.477 16.6168 124.674 16.7956 124.813 17.0217H124.83V13.812H125.604ZM118.604 16.2882C118.965 16.2882 119.271 16.356 119.52 16.4899C119.77 16.6246 119.974 16.8003 120.135 17.0178C120.295 17.2344 120.412 17.4795 120.483 17.7513C120.554 18.0231 120.591 18.3004 120.591 18.5817V18.8748H117.323C117.323 19.0646 117.361 19.2372 117.439 19.3932C117.516 19.5492 117.618 19.6823 117.745 19.7918C117.87 19.9021 118.016 19.9888 118.183 20.0495C118.349 20.1101 118.52 20.1409 118.699 20.1409C118.94 20.1409 119.149 20.081 119.327 19.962C119.504 19.8423 119.667 19.6863 119.816 19.4893L120.401 19.9667C119.972 20.56 119.37 20.857 118.595 20.857C118.273 20.857 117.984 20.7979 117.722 20.6821C117.462 20.5655 117.241 20.4064 117.061 20.2015C116.88 19.9967 116.74 19.754 116.643 19.4767C116.546 19.1978 116.497 18.8968 116.497 18.5722C116.497 18.2492 116.549 17.9475 116.657 17.6686C116.762 17.3904 116.908 17.1486 117.095 16.9445C117.281 16.7397 117.503 16.5789 117.762 16.4623C118.019 16.3465 118.3 16.2882 118.604 16.2882ZM100.467 16.2882C100.829 16.2882 101.133 16.356 101.383 16.4899C101.633 16.6246 101.837 16.8003 101.998 17.0178C102.159 17.2344 102.275 17.4795 102.347 17.7513C102.417 18.0231 102.453 18.3004 102.453 18.5817V18.8748H99.1856C99.1856 19.0646 99.2247 19.2372 99.3023 19.3932C99.3791 19.5492 99.481 19.6823 99.6073 19.7918C99.7336 19.9021 99.8806 19.9888 100.045 20.0495C100.212 20.1101 100.384 20.1409 100.562 20.1409C100.803 20.1409 101.013 20.081 101.19 19.962C101.368 19.8423 101.53 19.6863 101.68 19.4893L102.264 19.9667C101.834 20.56 101.233 20.857 100.458 20.857C100.137 20.857 99.8467 20.7979 99.5852 20.6821C99.3244 20.5655 99.1036 20.4064 98.9234 20.2015C98.7431 19.9967 98.6035 19.754 98.506 19.4767C98.41 19.1978 98.3605 18.8968 98.3605 18.5722C98.3605 18.2492 98.413 17.9475 98.5193 17.6686C98.6257 17.3904 98.7719 17.1486 98.9581 16.9445C99.1442 16.7397 99.3658 16.5789 99.6243 16.4623C99.8829 16.3465 100.164 16.2882 100.467 16.2882ZM76.4053 14.2524C76.8352 14.2524 77.2126 14.3044 77.5361 14.4084C77.8611 14.5124 78.1404 14.651 78.3789 14.826C78.6168 14.9993 78.8147 15.2002 78.9721 15.4263C79.1301 15.6532 79.2542 15.8848 79.3465 16.1236C79.4381 16.3615 79.5039 16.6002 79.5445 16.8389C79.5844 17.0777 79.6051 17.2975 79.6051 17.4992C79.6051 17.9152 79.5327 18.317 79.3909 18.7062C79.2461 19.0938 79.0334 19.4389 78.7534 19.7422C78.4727 20.0447 78.1241 20.2882 77.7075 20.4718C77.2924 20.6553 76.8123 20.7467 76.2679 20.7467H74.2817V14.2524H76.4053ZM113.505 13.7017C113.591 13.7017 113.681 13.7048 113.776 13.7103C113.87 13.7166 113.972 13.7348 114.081 13.7655L113.995 14.4998C113.92 14.4691 113.85 14.4478 113.78 14.4352C113.711 14.4226 113.636 14.4171 113.556 14.4171C113.413 14.4171 113.298 14.4407 113.212 14.4856C113.127 14.5329 113.059 14.5975 113.011 14.6833C112.961 14.7684 112.929 14.8709 112.916 14.9906C112.901 15.1096 112.894 15.2459 112.894 15.3987V16.3993H113.815V17.1139H112.886V20.7459H112.112V17.1139H111.235V16.3985H112.112V15.4263C112.112 14.8456 112.217 14.4131 112.426 14.1287C112.635 13.8451 112.994 13.7032 113.504 13.7032L113.505 13.7017ZM115.509 16.3985V20.7467H114.734V16.3985H115.509ZM110.581 16.3985V20.7467H109.807V16.3985H110.581ZM105.395 16.2882C105.521 16.2882 105.636 16.3063 105.739 16.3434L105.705 17.2337C105.642 17.2147 105.579 17.1998 105.516 17.1864C105.453 17.1754 105.39 17.169 105.326 17.169C104.948 17.169 104.658 17.2825 104.457 17.5078C104.258 17.7347 104.157 18.0861 104.157 18.5636V20.7459H103.383V16.3985H104.157V17.0682H104.174C104.225 16.9524 104.294 16.846 104.38 16.7515C104.466 16.6569 104.563 16.5758 104.668 16.5088C104.775 16.4411 104.89 16.3875 105.017 16.3481C105.143 16.3079 105.269 16.2882 105.395 16.2882ZM123.489 17.0587C123.282 17.0587 123.096 17.0981 122.929 17.1738C122.764 17.2502 122.623 17.355 122.508 17.485C122.394 17.6166 122.304 17.7765 122.241 17.9624C122.179 18.1491 122.148 18.3524 122.148 18.5722C122.148 18.7928 122.179 18.9961 122.241 19.1828C122.304 19.3688 122.394 19.5279 122.508 19.6587C122.623 19.7903 122.764 19.895 122.929 19.9707C123.096 20.0479 123.282 20.0857 123.489 20.0857C123.695 20.0857 123.881 20.0479 124.047 19.9707C124.213 19.895 124.355 19.7903 124.468 19.6587C124.584 19.5279 124.672 19.3688 124.735 19.1828C124.799 18.9961 124.83 18.792 124.83 18.5722C124.83 18.3524 124.799 18.1491 124.735 17.9624C124.672 17.7757 124.584 17.6166 124.468 17.485C124.355 17.3542 124.213 17.2502 124.047 17.1738C123.881 17.0974 123.695 17.0587 123.489 17.0587ZM76.3455 15.0781H75.1068V19.921H76.1482C76.4976 19.921 76.8278 19.8706 77.141 19.7698C77.4534 19.6689 77.7267 19.5177 77.9616 19.3152C78.198 19.1143 78.3834 18.8614 78.5215 18.5588C78.6589 18.2563 78.7283 17.9025 78.7283 17.4992C78.7283 17.292 78.6958 17.0485 78.6323 16.7704C78.5688 16.4915 78.4506 16.2268 78.2763 15.977C78.1012 15.7249 77.8597 15.5138 77.5494 15.3388C77.24 15.1647 76.8381 15.0781 76.3455 15.0781ZM118.587 16.9484C118.398 16.9484 118.224 16.9878 118.066 17.0635C117.909 17.1399 117.776 17.2392 117.666 17.3613C117.558 17.4842 117.473 17.6197 117.412 17.771C117.352 17.9191 117.323 18.0672 117.323 18.2145H119.765C119.76 18.031 119.732 17.8632 119.683 17.7103C119.634 17.5575 119.562 17.4243 119.464 17.3109C119.367 17.1974 119.245 17.1092 119.099 17.0454C118.951 16.9807 118.782 16.9484 118.587 16.9484ZM100.449 16.9484C100.26 16.9484 100.087 16.9878 99.9294 17.0635C99.7713 17.1399 99.6391 17.2392 99.5305 17.3613C99.4212 17.4842 99.3355 17.6197 99.2757 17.771C99.2166 17.9191 99.1856 18.0672 99.1856 18.2145H101.628C101.622 18.031 101.595 17.8632 101.547 17.7103C101.498 17.5575 101.425 17.4243 101.327 17.3109C101.231 17.1974 101.108 17.1092 100.962 17.0454C100.815 16.9807 100.644 16.9484 100.449 16.9484ZM110.193 14.2335C110.349 14.2335 110.482 14.2934 110.593 14.4123C110.705 14.5321 110.761 14.6739 110.761 14.8386C110.761 15.004 110.705 15.1466 110.593 15.2656C110.482 15.3845 110.349 15.4444 110.193 15.4444C110.039 15.4444 109.905 15.3845 109.794 15.2656C109.682 15.1466 109.626 15.004 109.626 14.8386C109.626 14.6739 109.682 14.5321 109.794 14.4123C109.905 14.2934 110.039 14.2335 110.193 14.2335ZM115.121 14.2335C115.276 14.2335 115.409 14.2934 115.521 14.4123C115.632 14.5321 115.688 14.6739 115.688 14.8386C115.688 15.004 115.632 15.1466 115.521 15.2656C115.409 15.3845 115.276 15.4444 115.121 15.4444C114.966 15.4444 114.833 15.3845 114.721 15.2656C114.61 15.1466 114.553 15.004 114.553 14.8386C114.553 14.6739 114.61 14.5321 114.721 14.4123C114.833 14.2934 114.966 14.2335 115.121 14.2335ZM51.4586 4.3679C49.0492 4.36159 47.4928 5.9058 47.4759 8.32217C47.4574 10.733 49.0263 12.33 51.415 12.3363C53.8466 12.3418 55.38 10.811 55.3852 8.3718C55.3903 5.93495 53.8599 4.37499 51.4586 4.3679Z" fill="white"/>
-  </svg>
-`;
-
-/**
- * @function IconGoogle
- * @returns {string} HTML
- */
-
-export const IconGoogle = () => `
-  <svg width="150" height="44" viewBox="0 0 150 44" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <rect x="0.5" y="0.5" width="149" height="43" rx="4.5" stroke="#A6A6A6"/>
-    <path d="M60.034 26.142C60.31 26.142 60.532 26.226 60.7 26.394C60.868 26.55 60.952 26.76 60.952 27.024V30.894C60.952 31.446 60.88 31.836 60.736 32.064C60.592 32.292 60.298 32.478 59.854 32.622C59.374 32.79 58.816 32.922 58.18 33.018C57.544 33.114 56.932 33.162 56.344 33.162C55.036 33.162 53.902 32.904 52.942 32.388C51.982 31.872 51.244 31.128 50.728 30.156C50.224 29.172 49.972 28.002 49.972 26.646C49.972 25.314 50.224 24.162 50.728 23.19C51.244 22.206 51.97 21.456 52.906 20.94C53.842 20.412 54.946 20.148 56.218 20.148C57.91 20.148 59.296 20.604 60.376 21.516C60.52 21.648 60.622 21.78 60.682 21.912C60.754 22.044 60.79 22.2 60.79 22.38C60.79 22.644 60.718 22.866 60.574 23.046C60.43 23.226 60.256 23.316 60.052 23.316C59.92 23.316 59.8 23.298 59.692 23.262C59.584 23.214 59.458 23.142 59.314 23.046C58.786 22.674 58.294 22.416 57.838 22.272C57.382 22.116 56.866 22.038 56.29 22.038C54.982 22.038 53.992 22.428 53.32 23.208C52.66 23.976 52.33 25.122 52.33 26.646C52.33 29.766 53.698 31.326 56.434 31.326C57.25 31.326 58.066 31.212 58.882 30.984V27.798H56.848C56.56 27.798 56.332 27.726 56.164 27.582C56.008 27.438 55.93 27.234 55.93 26.97C55.93 26.706 56.008 26.502 56.164 26.358C56.332 26.214 56.56 26.142 56.848 26.142H60.034ZM65.7652 33.144C64.8652 33.144 64.0732 32.958 63.3892 32.586C62.7172 32.214 62.1952 31.68 61.8232 30.984C61.4512 30.288 61.2652 29.478 61.2652 28.554C61.2652 27.63 61.4512 26.826 61.8232 26.142C62.1952 25.446 62.7172 24.912 63.3892 24.54C64.0732 24.168 64.8652 23.982 65.7652 23.982C66.6652 23.982 67.4512 24.168 68.1232 24.54C68.8072 24.912 69.3292 25.446 69.6892 26.142C70.0612 26.826 70.2472 27.63 70.2472 28.554C70.2472 29.478 70.0612 30.288 69.6892 30.984C69.3292 31.68 68.8072 32.214 68.1232 32.586C67.4512 32.958 66.6652 33.144 65.7652 33.144ZM65.7472 31.398C66.4792 31.398 67.0372 31.158 67.4212 30.678C67.8052 30.198 67.9972 29.49 67.9972 28.554C67.9972 27.63 67.8052 26.928 67.4212 26.448C67.0372 25.956 66.4852 25.71 65.7652 25.71C65.0452 25.71 64.4872 25.956 64.0912 26.448C63.7072 26.928 63.5152 27.63 63.5152 28.554C63.5152 29.49 63.7072 30.198 64.0912 30.678C64.4752 31.158 65.0272 31.398 65.7472 31.398ZM74.4987 33.144C73.5987 33.144 72.8067 32.958 72.1227 32.586C71.4507 32.214 70.9287 31.68 70.5567 30.984C70.1847 30.288 69.9987 29.478 69.9987 28.554C69.9987 27.63 70.1847 26.826 70.5567 26.142C70.9287 25.446 71.4507 24.912 72.1227 24.54C72.8067 24.168 73.5987 23.982 74.4987 23.982C75.3987 23.982 76.1847 24.168 76.8567 24.54C77.5407 24.912 78.0627 25.446 78.4227 26.142C78.7947 26.826 78.9807 27.63 78.9807 28.554C78.9807 29.478 78.7947 30.288 78.4227 30.984C78.0627 31.68 77.5407 32.214 76.8567 32.586C76.1847 32.958 75.3987 33.144 74.4987 33.144ZM74.4807 31.398C75.2127 31.398 75.7707 31.158 76.1547 30.678C76.5387 30.198 76.7307 29.49 76.7307 28.554C76.7307 27.63 76.5387 26.928 76.1547 26.448C75.7707 25.956 75.2187 25.71 74.4987 25.71C73.7787 25.71 73.2207 25.956 72.8247 26.448C72.4407 26.928 72.2487 27.63 72.2487 28.554C72.2487 29.49 72.4407 30.198 72.8247 30.678C73.2087 31.158 73.7607 31.398 74.4807 31.398ZM86.6702 24.018C87.0062 24.018 87.2762 24.114 87.4802 24.306C87.6842 24.498 87.7862 24.756 87.7862 25.08V32.19C87.7862 33.57 87.4022 34.614 86.6342 35.322C85.8662 36.03 84.7322 36.384 83.2322 36.384C81.9242 36.384 80.8322 36.138 79.9562 35.646C79.5122 35.382 79.2902 35.064 79.2902 34.692C79.2902 34.428 79.3502 34.218 79.4702 34.062C79.6022 33.906 79.7702 33.828 79.9742 33.828C80.0822 33.828 80.2142 33.852 80.3702 33.9C80.5262 33.96 80.6882 34.026 80.8562 34.098C81.2882 34.278 81.6722 34.416 82.0082 34.512C82.3562 34.608 82.7522 34.656 83.1962 34.656C84.7682 34.656 85.5542 33.888 85.5542 32.352V31.2C85.2902 31.692 84.9062 32.076 84.4022 32.352C83.8982 32.616 83.3222 32.748 82.6742 32.748C81.9062 32.748 81.2222 32.568 80.6222 32.208C80.0342 31.836 79.5722 31.32 79.2362 30.66C78.9002 30 78.7322 29.238 78.7322 28.374C78.7322 27.51 78.8942 26.748 79.2182 26.088C79.5542 25.416 80.0222 24.9 80.6222 24.54C81.2222 24.168 81.9062 23.982 82.6742 23.982C83.3222 23.982 83.8982 24.12 84.4022 24.396C84.9062 24.66 85.2902 25.038 85.5542 25.53V25.062C85.5542 24.75 85.6562 24.498 85.8602 24.306C86.0642 24.114 86.3342 24.018 86.6702 24.018ZM83.2682 31.002C83.9882 31.002 84.5462 30.768 84.9422 30.3C85.3502 29.832 85.5542 29.19 85.5542 28.374C85.5542 27.546 85.3502 26.898 84.9422 26.43C84.5462 25.962 83.9882 25.728 83.2682 25.728C82.5602 25.728 82.0022 25.962 81.5942 26.43C81.1862 26.898 80.9822 27.546 80.9822 28.374C80.9822 29.19 81.1862 29.832 81.5942 30.3C82.0022 30.768 82.5602 31.002 83.2682 31.002ZM89.6126 33.108C89.2886 33.108 89.0186 33.018 88.8026 32.838C88.5986 32.646 88.4966 32.376 88.4966 32.028V21.282C88.4966 20.934 88.5986 20.67 88.8026 20.49C89.0186 20.31 89.2886 20.22 89.6126 20.22C89.9366 20.22 90.2066 20.31 90.4226 20.49C90.6386 20.67 90.7466 20.934 90.7466 21.282V32.028C90.7466 32.376 90.6386 32.646 90.4226 32.838C90.2066 33.018 89.9366 33.108 89.6126 33.108ZM98.413 30.588C98.617 30.588 98.779 30.666 98.899 30.822C99.031 30.978 99.097 31.188 99.097 31.452C99.097 31.824 98.875 32.136 98.431 32.388C98.023 32.616 97.561 32.802 97.045 32.946C96.529 33.078 96.037 33.144 95.569 33.144C94.153 33.144 93.031 32.736 92.203 31.92C91.375 31.104 90.961 29.988 90.961 28.572C90.961 27.672 91.141 26.874 91.501 26.178C91.861 25.482 92.365 24.942 93.013 24.558C93.673 24.174 94.417 23.982 95.245 23.982C96.037 23.982 96.727 24.156 97.315 24.504C97.903 24.852 98.359 25.344 98.683 25.98C99.007 26.616 99.169 27.366 99.169 28.23C99.169 28.746 98.941 29.004 98.485 29.004H93.175C93.247 29.832 93.481 30.444 93.877 30.84C94.273 31.224 94.849 31.416 95.605 31.416C95.989 31.416 96.325 31.368 96.613 31.272C96.913 31.176 97.249 31.044 97.621 30.876C97.981 30.684 98.245 30.588 98.413 30.588ZM95.299 25.566C94.687 25.566 94.195 25.758 93.823 26.142C93.463 26.526 93.247 27.078 93.175 27.798H97.243C97.219 27.066 97.039 26.514 96.703 26.142C96.367 25.758 95.899 25.566 95.299 25.566ZM104.006 33.108C103.658 33.108 103.376 33.006 103.16 32.802C102.956 32.586 102.854 32.298 102.854 31.938V21.408C102.854 21.06 102.95 20.79 103.142 20.598C103.346 20.406 103.622 20.31 103.97 20.31H108.398C109.754 20.31 110.81 20.652 111.566 21.336C112.322 22.008 112.7 22.956 112.7 24.18C112.7 25.404 112.322 26.358 111.566 27.042C110.81 27.726 109.754 28.068 108.398 28.068H105.176V31.938C105.176 32.298 105.068 32.586 104.852 32.802C104.648 33.006 104.366 33.108 104.006 33.108ZM108.11 26.286C109.706 26.286 110.504 25.59 110.504 24.198C110.504 22.806 109.706 22.11 108.11 22.11H105.176V26.286H108.11ZM114.035 33.108C113.711 33.108 113.441 33.018 113.225 32.838C113.021 32.646 112.919 32.376 112.919 32.028V21.282C112.919 20.934 113.021 20.67 113.225 20.49C113.441 20.31 113.711 20.22 114.035 20.22C114.359 20.22 114.629 20.31 114.845 20.49C115.061 20.67 115.169 20.934 115.169 21.282V32.028C115.169 32.376 115.061 32.646 114.845 32.838C114.629 33.018 114.359 33.108 114.035 33.108ZM119.685 23.982C120.909 23.982 121.815 24.288 122.403 24.9C123.003 25.5 123.303 26.418 123.303 27.654V32.046C123.303 32.37 123.207 32.628 123.015 32.82C122.823 33 122.559 33.09 122.223 33.09C121.911 33.09 121.653 32.994 121.449 32.802C121.257 32.61 121.161 32.358 121.161 32.046V31.65C120.957 32.118 120.633 32.484 120.189 32.748C119.757 33.012 119.253 33.144 118.677 33.144C118.089 33.144 117.555 33.024 117.075 32.784C116.595 32.544 116.217 32.214 115.941 31.794C115.665 31.374 115.527 30.906 115.527 30.39C115.527 29.742 115.689 29.232 116.013 28.86C116.349 28.488 116.889 28.218 117.633 28.05C118.377 27.882 119.403 27.798 120.711 27.798H121.161V27.384C121.161 26.796 121.035 26.37 120.783 26.106C120.531 25.83 120.123 25.692 119.559 25.692C119.211 25.692 118.857 25.746 118.497 25.854C118.137 25.95 117.711 26.094 117.219 26.286C116.907 26.442 116.679 26.52 116.535 26.52C116.319 26.52 116.139 26.442 115.995 26.286C115.863 26.13 115.797 25.926 115.797 25.674C115.797 25.47 115.845 25.296 115.941 25.152C116.049 24.996 116.223 24.852 116.463 24.72C116.883 24.492 117.381 24.312 117.957 24.18C118.545 24.048 119.121 23.982 119.685 23.982ZM119.127 31.524C119.727 31.524 120.213 31.326 120.585 30.93C120.969 30.522 121.161 30 121.161 29.364V28.986H120.837C120.033 28.986 119.409 29.022 118.965 29.094C118.521 29.166 118.203 29.292 118.011 29.472C117.819 29.652 117.723 29.898 117.723 30.21C117.723 30.594 117.855 30.912 118.119 31.164C118.395 31.404 118.731 31.524 119.127 31.524ZM130.108 24.666C130.204 24.45 130.33 24.294 130.486 24.198C130.654 24.09 130.834 24.036 131.026 24.036C131.302 24.036 131.548 24.132 131.764 24.324C131.992 24.504 132.106 24.732 132.106 25.008C132.106 25.14 132.07 25.278 131.998 25.422L127.084 35.718C126.88 36.126 126.562 36.33 126.13 36.33C125.854 36.33 125.608 36.24 125.392 36.06C125.188 35.892 125.086 35.676 125.086 35.412C125.086 35.28 125.122 35.13 125.194 34.962L126.418 32.388L123.07 25.422C123.01 25.302 122.98 25.164 122.98 25.008C122.98 24.732 123.094 24.498 123.322 24.306C123.562 24.114 123.832 24.018 124.132 24.018C124.336 24.018 124.522 24.072 124.69 24.18C124.858 24.276 124.99 24.432 125.086 24.648L127.606 30.21L130.108 24.666Z" fill="white"/>
-    <path d="M55.13 12.19C55.2833 12.19 55.4067 12.2367 55.5 12.33C55.5933 12.4167 55.64 12.5333 55.64 12.68V14.83C55.64 15.1367 55.6 15.3533 55.52 15.48C55.44 15.6067 55.2767 15.71 55.03 15.79C54.7633 15.8833 54.4533 15.9567 54.1 16.01C53.7467 16.0633 53.4067 16.09 53.08 16.09C52.3533 16.09 51.7233 15.9467 51.19 15.66C50.6567 15.3733 50.2467 14.96 49.96 14.42C49.68 13.8733 49.54 13.2233 49.54 12.47C49.54 11.73 49.68 11.09 49.96 10.55C50.2467 10.0033 50.65 9.58667 51.17 9.3C51.69 9.00667 52.3033 8.86 53.01 8.86C53.95 8.86 54.72 9.11333 55.32 9.62C55.4 9.69333 55.4567 9.76667 55.49 9.84C55.53 9.91333 55.55 10 55.55 10.1C55.55 10.2467 55.51 10.37 55.43 10.47C55.35 10.57 55.2533 10.62 55.14 10.62C55.0667 10.62 55 10.61 54.94 10.59C54.88 10.5633 54.81 10.5233 54.73 10.47C54.4367 10.2633 54.1633 10.12 53.91 10.04C53.6567 9.95333 53.37 9.91 53.05 9.91C52.3233 9.91 51.7733 10.1267 51.4 10.56C51.0333 10.9867 50.85 11.6233 50.85 12.47C50.85 14.2033 51.61 15.07 53.13 15.07C53.5833 15.07 54.0367 15.0067 54.49 14.88V13.11H53.36C53.2 13.11 53.0733 13.07 52.98 12.99C52.8933 12.91 52.85 12.7967 52.85 12.65C52.85 12.5033 52.8933 12.39 52.98 12.31C53.0733 12.23 53.2 12.19 53.36 12.19H55.13ZM57.714 16C57.5207 16 57.3673 15.9467 57.254 15.84C57.1473 15.7333 57.094 15.5833 57.094 15.39V9.56C57.094 9.36667 57.1473 9.21667 57.254 9.11C57.3673 9.00333 57.5207 8.95 57.714 8.95H61.324C61.5173 8.95 61.6673 8.99333 61.774 9.08C61.8807 9.16667 61.934 9.29 61.934 9.45C61.934 9.61667 61.8807 9.74667 61.774 9.84C61.6673 9.92667 61.5173 9.97 61.324 9.97H58.354V11.91H61.124C61.3173 11.91 61.4673 11.9533 61.574 12.04C61.6807 12.1267 61.734 12.2533 61.734 12.42C61.734 12.58 61.6807 12.7033 61.574 12.79C61.4673 12.8767 61.3173 12.92 61.124 12.92H58.354V14.98H61.324C61.5173 14.98 61.6673 15.0267 61.774 15.12C61.8807 15.2067 61.934 15.3333 61.934 15.5C61.934 15.66 61.8807 15.7833 61.774 15.87C61.6673 15.9567 61.5173 16 61.324 16H57.714ZM65.3024 16.05C65.1024 16.05 64.9424 15.9933 64.8224 15.88C64.7091 15.76 64.6524 15.6033 64.6524 15.41V10.01H62.8324C62.4324 10.01 62.2324 9.83333 62.2324 9.48C62.2324 9.12667 62.4324 8.95 62.8324 8.95H67.7624C68.1624 8.95 68.3624 9.12667 68.3624 9.48C68.3624 9.83333 68.1624 10.01 67.7624 10.01H65.9424V15.41C65.9424 15.6033 65.8824 15.76 65.7624 15.88C65.6491 15.9933 65.4958 16.05 65.3024 16.05ZM72.4801 16.06C72.2867 16.06 72.1301 16 72.0101 15.88C71.8967 15.76 71.8401 15.6033 71.8401 15.41V9.54C71.8401 9.34667 71.8967 9.19333 72.0101 9.08C72.1301 8.96 72.2867 8.9 72.4801 8.9C72.6801 8.9 72.8367 8.96 72.9501 9.08C73.0701 9.19333 73.1301 9.34667 73.1301 9.54V15.41C73.1301 15.6033 73.0701 15.76 72.9501 15.88C72.8367 16 72.6801 16.06 72.4801 16.06ZM76.9626 16.05C76.7626 16.05 76.6026 15.9933 76.4826 15.88C76.3692 15.76 76.3126 15.6033 76.3126 15.41V10.01H74.4926C74.0926 10.01 73.8926 9.83333 73.8926 9.48C73.8926 9.12667 74.0926 8.95 74.4926 8.95H79.4226C79.8226 8.95 80.0226 9.12667 80.0226 9.48C80.0226 9.83333 79.8226 10.01 79.4226 10.01H77.6026V15.41C77.6026 15.6033 77.5426 15.76 77.4226 15.88C77.3092 15.9933 77.1559 16.05 76.9626 16.05ZM86.6402 16.09C85.9536 16.09 85.3569 15.9433 84.8502 15.65C84.3502 15.3567 83.9636 14.9367 83.6902 14.39C83.4169 13.8433 83.2802 13.2033 83.2802 12.47C83.2802 11.7367 83.4136 11.1 83.6802 10.56C83.9536 10.0133 84.3436 9.59333 84.8502 9.3C85.3569 9.00667 85.9536 8.86 86.6402 8.86C87.3269 8.86 87.9202 9.00667 88.4202 9.3C88.9269 9.59333 89.3136 10.0133 89.5802 10.56C89.8536 11.1 89.9902 11.7367 89.9902 12.47C89.9902 13.2033 89.8536 13.8433 89.5802 14.39C89.3069 14.9367 88.9169 15.3567 88.4102 15.65C87.9102 15.9433 87.3202 16.09 86.6402 16.09ZM86.6402 15.06C87.2869 15.06 87.7902 14.8367 88.1502 14.39C88.5169 13.9433 88.7002 13.3033 88.7002 12.47C88.7002 11.6367 88.5169 11 88.1502 10.56C87.7902 10.1133 87.2869 9.89 86.6402 9.89C85.9869 9.89 85.4769 10.1133 85.1102 10.56C84.7502 11 84.5702 11.6367 84.5702 12.47C84.5702 13.3033 84.7502 13.9433 85.1102 14.39C85.4769 14.8367 85.9869 15.06 86.6402 15.06ZM96.6332 8.9C96.8199 8.9 96.9665 8.96 97.0732 9.08C97.1865 9.19333 97.2432 9.34667 97.2432 9.54V15.41C97.2432 15.6033 97.1865 15.76 97.0732 15.88C96.9599 16 96.8132 16.06 96.6332 16.06C96.4199 16.06 96.2565 15.9833 96.1432 15.83L92.4932 11.08V15.41C92.4932 15.6033 92.4399 15.76 92.3332 15.88C92.2265 16 92.0799 16.06 91.8932 16.06C91.7065 16.06 91.5599 16 91.4532 15.88C91.3465 15.76 91.2932 15.6033 91.2932 15.41V9.54C91.2932 9.34667 91.3499 9.19333 91.4632 9.08C91.5765 8.96 91.7265 8.9 91.9132 8.9C92.1199 8.9 92.2799 8.97667 92.3932 9.13L96.0332 13.86V9.54C96.0332 9.34667 96.0865 9.19333 96.1932 9.08C96.3065 8.96 96.4532 8.9 96.6332 8.9Z" fill="white"/>
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M33.4722 26.6532L28.9963 22.1554V21.8401L33.4722 17.3423L33.5694 17.4024L38.8638 20.4259C40.3786 21.2837 40.3786 22.7005 38.8638 23.5696L33.5694 26.5931L33.4722 26.6532Z" fill="#FFE000"/>
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M33.5703 26.5932L28.9955 21.9978L15.4941 35.5679C15.9884 36.0991 16.8182 36.1601 17.7442 35.6411L33.5694 26.5932" fill="#FF3A44"/>
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M33.5702 17.4025L17.7433 8.36851C16.8172 7.83644 15.9848 7.90959 15.4932 8.44166L28.9963 21.9979L33.5693 17.4025H33.5702Z" fill="#06E775"/>
-    <path opacity="0.2" d="M33.4723 26.4956L17.7555 35.4704C16.8771 35.9781 16.0941 35.9424 15.5886 35.4808L15.5054 35.5653L15.5886 35.6384C16.0941 36.0974 16.8771 36.1348 17.7555 35.6271L33.5798 26.5931L33.4723 26.4956Z" fill="black"/>
-    <path opacity="0.12" d="M38.8639 23.412L33.4585 26.4956L33.5565 26.5932L38.85 23.5696C39.6087 23.1351 39.9816 22.5647 39.9816 21.9978C39.9356 22.5194 39.5498 23.0132 38.863 23.412H38.8639Z" fill="black"/>
-    <path opacity="0.25" fill-rule="evenodd" clip-rule="evenodd" d="M17.7443 8.52605L38.8639 20.5836C39.5489 20.9719 39.9357 21.4796 39.9955 21.9978C39.9955 21.4309 39.6226 20.8605 38.8639 20.4259L17.7443 8.36931C16.2295 7.49761 15 8.22475 15 9.96467V10.1223C15 8.37976 16.2295 7.66655 17.7443 8.52605Z" fill="white"/>
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M15.4942 8.42778C15.1812 8.7674 15 9.28555 15 9.96392V34.0337C15 34.7129 15.1812 35.231 15.5055 35.5576L15.5888 35.6308L29.0078 22.1442V21.8429L15.5783 8.35376L15.4942 8.42778Z" fill="#00D2FF"/>
-  </svg>
-`;
-
-/**
- * @function IconIn
- * @returns {string}
- */
-
-export const IconIn = () => `
-  <svg width="40" height="40" viewBox="0 0 40 40" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <rect x="1" y="1" width="38" height="38" stroke="white" stroke-width="2"/>
-    <g clip-path="url(#clip0_941_951)">
-      <path d="M15.0299 27.875H11.2973V17.3598H15.0299V27.875ZM13.1616 15.9254C11.9683 15.9254 11 15.0605 11 14.0164C11 12.9723 11.9683 12.125 13.1616 12.125C14.3549 12.125 15.3232 12.9723 15.3232 14.0164C15.3232 15.0605 14.3549 15.9254 13.1616 15.9254ZM29 27.875H25.2754V22.7562C25.2754 21.5363 25.2473 19.9719 23.3348 19.9719C21.3942 19.9719 21.0969 21.2973 21.0969 22.6684V27.875H17.3683V17.3598H20.9482V18.7941H21.0004C21.4986 17.968 22.7161 17.0961 24.5321 17.0961C28.3089 17.0961 29.004 19.2723 29.004 22.0988V27.875H29Z" fill="white"/>
-    </g>
-    <defs>
-      <clipPath id="clip0_941_951">
-        <rect width="18" height="18" fill="white" transform="translate(11 11)"/>
-      </clipPath>
-    </defs>
-  </svg>
-`;
-
-/**
- * @function IconLogo
- * @returns {string} HTML
- */
-
-export const IconLogo = () => `
-  <svg viewBox="0 0 102 40" xmlns="http://www.w3.org/2000/svg">
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M68.5369 25.1423C68.8238 25.3894 68.9807 25.753 68.9647 26.1307C68.9647 27.195 68.3408 27.7262 67.1002 27.7262H56.4159C55.596 27.7757 54.8865 27.1562 54.8384 26.3425C54.8348 26.2931 54.8348 26.2419 54.8366 26.1907C54.8366 25.6436 55.0648 25.1247 55.4658 24.7522L64.2856 15.3554H57.4587C57.013 15.3854 56.5692 15.2584 56.2109 14.9936C55.9186 14.7377 55.7617 14.3635 55.7831 13.977C55.7831 12.9339 56.4034 12.4114 57.6476 12.4114H67.8577C68.2891 12.3902 68.7115 12.5561 69.011 12.865C69.2891 13.1562 69.4424 13.5445 69.437 13.947C69.4263 14.4976 69.1982 15.0254 68.806 15.4172L60.0183 24.7822H67.2873C67.733 24.7522 68.1786 24.8793 68.5369 25.1423ZM74.1785 18.6736H81.3851C81.4795 17.6411 81.1872 16.6086 80.5633 15.779C79.9573 15.103 79.0714 14.7412 78.1605 14.7924C75.9271 14.7924 74.6009 16.0861 74.1785 18.6754V18.6736ZM84.3743 20.4121C84.1248 20.6539 83.7861 20.7863 83.4403 20.7722H74.0537C74.0269 21.93 74.4494 23.0508 75.2355 23.905C76.0786 24.6569 77.1891 25.0399 78.3192 24.9711C78.9965 24.9799 79.6739 24.8899 80.3263 24.7064C80.9537 24.514 81.5544 24.2475 82.1123 23.905C82.353 23.7674 82.6239 23.6932 82.902 23.6862C83.2282 23.6721 83.5437 23.8186 83.7397 24.078C83.9447 24.3551 84.0517 24.6905 84.0392 25.0329C84.0428 25.633 83.7469 26.196 83.2496 26.5366C82.5633 27.0097 81.7986 27.3609 80.9894 27.5709C80.0375 27.8427 79.0554 27.9804 78.0661 27.9769C76.6312 28.0104 75.2105 27.6927 73.9253 27.0538C72.7738 26.4678 71.8238 25.5553 71.191 24.4381C70.5279 23.2291 70.1946 21.8665 70.2284 20.4898C70.2017 18.979 70.5386 17.4822 71.207 16.1232C71.8184 14.8824 72.7721 13.8411 73.9574 13.1157C75.1767 12.3849 76.5777 12.009 78.0037 12.0337C79.8646 11.9208 81.6827 12.6179 82.9822 13.9452C84.2388 15.4772 84.8663 17.4222 84.7362 19.3955C84.7611 19.7697 84.6256 20.1368 84.3725 20.4121H84.3743ZM100.64 13.4281C101.611 14.5753 102.092 16.0544 101.985 17.5475C101.985 17.9429 101.973 18.2465 101.95 18.4565L101.319 26.0354C101.304 26.5614 101.076 27.0573 100.686 27.4121C100.285 27.7563 99.7626 27.9345 99.2332 27.9151C98.7109 27.9469 98.1976 27.7651 97.809 27.4138C97.4614 27.0432 97.2974 26.5366 97.3687 26.0354L98.0015 18.6136C98.169 16.2944 97.2421 15.1366 95.2208 15.1366C94.2048 15.0801 93.2155 15.456 92.5007 16.1691C91.8002 16.9722 91.3991 17.9888 91.3635 19.0513L90.7628 26.0354C90.7485 26.5614 90.5203 27.0573 90.13 27.4121C89.7289 27.7563 89.2067 27.9345 88.6773 27.9151C88.1603 27.9486 87.647 27.7668 87.2691 27.4138C86.9304 27.0379 86.7629 26.5366 86.811 26.0354L87.8234 13.8817C87.8252 13.3786 88.0587 12.9003 88.4544 12.5826C88.8769 12.2526 89.4045 12.0814 89.941 12.0973C90.4312 12.069 90.9089 12.2385 91.269 12.5667C91.5916 12.9021 91.7324 13.3663 91.6487 13.8199L91.6148 14.4147C92.1531 13.6557 92.8822 13.0503 93.7342 12.6603C94.6629 12.2349 95.6789 12.0196 96.7056 12.0355C98.1565 11.9313 99.5843 12.4379 100.64 13.4298V13.4281ZM38.4448 11.6031C36.8334 9.14265 33.6178 8.26369 30.9601 9.55566C28.1224 10.9853 26.9905 14.4217 28.429 17.2334C29.2507 18.836 30.7854 19.9638 32.5715 20.2762C32.354 27.0538 26.6322 32.377 19.7892 32.1617C13.1441 31.9534 7.84832 26.5861 7.78771 19.9974C7.77702 14.9212 10.9606 10.3781 15.7662 8.61669C16.6895 8.28663 17.2242 7.32824 17.0103 6.37691L16.9961 6.31867C16.7608 5.27732 17.4203 4.24657 18.4684 4.01183C18.554 3.99418 18.6413 3.98006 18.7287 3.973C19.2099 3.92887 19.6966 3.90593 20.1832 3.90593C20.7857 3.90593 21.3899 3.94123 21.9871 4.00477C23.0708 4.12302 24.0476 3.34642 24.1635 2.27507C24.1688 2.23977 24.1724 2.20094 24.1742 2.16564C24.2027 1.10135 23.3863 0.199444 22.3168 0.111194C21.6074 0.0352998 20.8944 0 20.1832 0C9.03724 0 0 8.95557 0 20.0009C0 31.048 9.03724 40 20.1832 40C31.3309 40 40.3663 31.048 40.3663 20.0009C40.3663 18.0453 39.9688 13.8305 38.4448 11.6031Z" fill="currentColor" />
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M26.0962 4C27.2542 4 28.1927 4.9278 28.1927 6.07193C28.1927 7.21606 27.2542 8.14386 26.0962 8.14386C24.9386 8.14386 24 7.21606 24 6.07193C24 4.9278 24.9386 4 26.0962 4Z" fill="currentColor" />
-  </svg>
-`;
-
-/**
- * @function IconLogo2
- * @returns {string} HTML
- */
-
-export const IconLogo2 = () => `
-  <svg width="216" height="23" viewBox="0 0 216 23" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M35.5079 14.4568C35.6565 14.5989 35.7378 14.808 35.7295 15.0252C35.7295 15.6371 35.4063 15.9426 34.7635 15.9426H29.2282C28.8034 15.971 28.4358 15.6148 28.4109 15.1469C28.4091 15.1185 28.4091 15.0891 28.41 15.0597C28.41 14.745 28.5282 14.4467 28.736 14.2325L33.3053 8.82937H29.7684C29.5375 8.84662 29.3076 8.77355 29.122 8.62132C28.9705 8.47416 28.8893 8.25901 28.9003 8.03676C28.9003 7.43697 29.2217 7.13657 29.8663 7.13657H35.156C35.3795 7.12439 35.5984 7.21979 35.7535 7.39739C35.8976 7.56484 35.977 7.78811 35.9742 8.0195C35.9687 8.33614 35.8505 8.63959 35.6473 8.86489L31.0945 14.2498H34.8605C35.0914 14.2325 35.3222 14.3056 35.5079 14.4568ZM38.4307 10.7373H42.1643C42.2133 10.1436 42.0618 9.54993 41.7386 9.07294C41.4246 8.68424 40.9656 8.47619 40.4937 8.50563C39.3366 8.50563 38.6495 9.24953 38.4307 10.7383V10.7373ZM43.713 11.737C43.5837 11.876 43.4082 11.9521 43.2291 11.944H38.366C38.3522 12.6098 38.571 13.2542 38.9783 13.7454C39.4151 14.1777 39.9904 14.398 40.5759 14.3584C40.9268 14.3635 41.2778 14.3117 41.6158 14.2062C41.9408 14.0955 42.252 13.9423 42.5411 13.7454C42.6658 13.6662 42.8061 13.6236 42.9502 13.6196C43.1192 13.6114 43.2826 13.6957 43.3842 13.8449C43.4904 14.0042 43.5458 14.197 43.5394 14.3939C43.5412 14.739 43.3879 15.0627 43.1303 15.2586C42.7747 15.5306 42.3786 15.7325 41.9593 15.8533C41.4662 16.0096 40.9573 16.0887 40.4448 16.0867C39.7014 16.106 38.9654 15.9233 38.2995 15.5559C37.703 15.219 37.2108 14.6943 36.8829 14.0519C36.5394 13.3567 36.3667 12.5732 36.3842 11.7816C36.3704 10.9129 36.5449 10.0523 36.8912 9.27084C37.208 8.55738 37.702 7.95861 38.3162 7.5415C38.9478 7.12134 39.6737 6.90518 40.4125 6.91938C41.3766 6.85443 42.3185 7.25531 42.9917 8.01849C43.6428 8.89939 43.9679 10.0178 43.9004 11.1524C43.9134 11.3676 43.8432 11.5787 43.7121 11.737H43.713ZM52.1397 7.72113C52.643 8.3808 52.8924 9.23126 52.837 10.0898C52.837 10.3172 52.8305 10.4917 52.8185 10.6125L52.4916 14.9703C52.4842 15.2728 52.366 15.558 52.1638 15.7619C51.956 15.9598 51.6854 16.0623 51.4111 16.0512C51.1405 16.0695 50.8746 15.9649 50.6733 15.763C50.4932 15.5498 50.4082 15.2586 50.4452 14.9703L50.773 10.7028C50.8598 9.36928 50.3796 8.70353 49.3324 8.70353C48.806 8.67105 48.2935 8.88722 47.9231 9.29722C47.5602 9.75899 47.3524 10.3436 47.334 10.9545L47.0227 14.9703C47.0154 15.2728 46.8971 15.558 46.6949 15.7619C46.4871 15.9598 46.2165 16.0623 45.9423 16.0512C45.6745 16.0705 45.4085 15.9659 45.2127 15.763C45.0373 15.5468 44.9504 15.2586 44.9754 14.9703L45.4999 7.98195C45.5008 7.69271 45.6218 7.41769 45.8268 7.23501C46.0457 7.04523 46.319 6.94679 46.597 6.95592C46.851 6.93968 47.0985 7.03711 47.285 7.22587C47.4522 7.4187 47.5251 7.68561 47.4817 7.94643L47.4642 8.28844C47.7431 7.85205 48.1208 7.50395 48.5622 7.27966C49.0433 7.03508 49.5697 6.91126 50.1016 6.9204C50.8533 6.86052 51.593 7.15179 52.1397 7.72215V7.72113ZM19.9176 6.67176C19.0828 5.25703 17.4169 4.75162 16.0399 5.49451C14.5698 6.31655 13.9834 8.2925 14.7286 9.90919C15.1543 10.8307 15.9494 11.4792 16.8748 11.6588C16.7621 15.5559 13.7977 18.6168 10.2525 18.493C6.80975 18.3732 4.06609 15.287 4.03469 11.4985C4.02915 8.57971 5.67849 5.96744 8.16819 4.9546C8.64655 4.76481 8.9236 4.21374 8.81278 3.66673L8.80539 3.63323C8.68349 3.03446 9.02518 2.44178 9.56819 2.3068C9.61252 2.29665 9.65777 2.28853 9.70302 2.28447C9.95236 2.2591 10.2045 2.24591 10.4566 2.24591C10.7687 2.24591 11.0818 2.2662 11.3911 2.30274C11.9526 2.37074 12.4587 1.92419 12.5187 1.30817C12.5215 1.28787 12.5233 1.26554 12.5243 1.24525C12.539 0.633279 12.1161 0.11468 11.562 0.0639368C11.1944 0.0202974 10.825 0 10.4566 0C4.68205 0 0 5.14945 0 11.5005C0 17.8526 4.68205 23 10.4566 23C16.232 23 20.9132 17.8526 20.9132 11.5005C20.9132 10.376 20.7072 7.95252 19.9176 6.67176Z" fill="white"/>
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M14.4007 0.872681C14.9982 0.872681 15.484 1.4065 15.484 2.06414C15.484 2.72177 14.9982 3.2556 14.4007 3.2556C13.8023 3.2556 13.3175 2.72177 13.3175 2.06414C13.3175 1.4065 13.8023 0.872681 14.4007 0.872681ZM61.031 10.4114C61.6285 10.4114 62.1143 10.9443 62.1143 11.6029C62.1143 12.2605 61.6285 12.7944 61.031 12.7944C60.4326 12.7944 59.9468 12.2605 59.9468 11.6029C59.9468 10.9432 60.4326 10.4114 61.031 10.4114ZM74.7115 7.61142C74.8186 7.74031 74.8731 7.90066 74.8731 8.09145C74.8731 8.34009 74.79 8.55829 74.6265 8.7501L69.9934 14.3349H73.9524C74.1574 14.3349 74.3153 14.3918 74.428 14.5044C74.5397 14.6171 74.597 14.7683 74.597 14.9601C74.597 15.1844 74.5351 15.3559 74.4123 15.4746C74.2904 15.5924 74.1103 15.6512 73.8757 15.6512H68.5371C68.3404 15.6512 68.1898 15.5903 68.0836 15.4655C67.9765 15.3437 67.923 15.1844 67.923 14.9936C67.923 14.7693 68.0042 14.5612 68.1677 14.3694L72.8174 8.73386H69.1189C68.9139 8.73386 68.7559 8.67703 68.6433 8.56438C68.5325 8.45173 68.4752 8.30051 68.4752 8.1087C68.4752 7.64694 68.7107 7.41758 69.1807 7.41758H74.2599C74.4538 7.41758 74.6043 7.48151 74.7124 7.6104L74.7115 7.61142ZM77.9132 9.0708C77.5438 9.47675 77.3083 10.0654 77.2067 10.8428H81.1648C81.186 10.0664 81.0392 9.47573 80.7279 9.0708C80.4158 8.66485 79.9735 8.46391 79.4009 8.46391C78.7766 8.46391 78.2807 8.66587 77.9132 9.0708ZM82.3551 11.6435C82.2462 11.7622 82.1021 11.8211 81.9165 11.8211H77.1448C77.1559 12.6533 77.3499 13.2977 77.7285 13.7544C78.1071 14.2091 78.6538 14.4374 79.3695 14.4374C80.0852 14.4374 80.7501 14.2121 81.3642 13.7615C81.4769 13.6824 81.5896 13.6438 81.7022 13.6438C81.8454 13.6438 81.9626 13.7057 82.0541 13.8295C82.1464 13.9533 82.1926 14.0985 82.1926 14.268C82.1926 14.5278 82.091 14.746 81.8869 14.9266C81.6099 15.1732 81.2294 15.3762 80.7436 15.5335C80.257 15.6918 79.7694 15.77 79.2781 15.77C78.5411 15.77 77.8966 15.6066 77.3452 15.2808C76.7911 14.954 76.3663 14.4872 76.0634 13.8803C75.7615 13.2713 75.6109 12.5701 75.6109 11.7714C75.6109 10.8813 75.7661 10.0918 76.0791 9.39962C76.3904 8.70748 76.8272 8.17163 77.3905 7.78902C77.9529 7.40641 78.603 7.21359 79.339 7.21359C80.3623 7.21359 81.1472 7.5485 81.6939 8.21831C82.2406 8.88812 82.5158 9.85834 82.5158 11.129C82.5158 11.3543 82.4613 11.5258 82.3542 11.6435H82.3551ZM88.9202 7.38307C89.0384 7.49572 89.0975 7.65912 89.0975 7.87224C89.0975 8.38982 88.8417 8.67094 88.33 8.7156L87.9163 8.76634C87.2099 8.8455 86.6862 9.0708 86.3427 9.44123C86.001 9.81267 85.8043 10.3597 85.7517 11.0792L85.4608 14.9256C85.4405 15.2067 85.3537 15.4127 85.2004 15.5416C85.0462 15.6715 84.8568 15.7365 84.6324 15.7365C84.4071 15.7365 84.2261 15.6685 84.0876 15.5335C83.95 15.3985 83.8909 15.1966 83.9121 14.9256L84.4173 8.02447C84.4274 7.77684 84.5151 7.58605 84.6786 7.45005C84.8421 7.31609 85.0314 7.24911 85.2456 7.24911C85.4506 7.24911 85.6132 7.31305 85.7369 7.44193C85.8597 7.57184 85.9161 7.75959 85.9059 8.00722L85.8597 8.70037C86.3704 7.83367 87.1794 7.35567 88.2839 7.26535L88.4538 7.24911C88.6477 7.22678 88.8038 7.27144 88.9202 7.38307ZM94.6162 13.5423C94.9948 12.9578 95.186 12.1712 95.186 11.1797C95.186 10.3363 95.0077 9.6919 94.6559 9.24739C94.3031 8.80389 93.798 8.58163 93.1451 8.58163C92.4275 8.58163 91.8799 8.87392 91.5013 9.45848C91.1245 10.0441 90.9352 10.8316 90.9352 11.8211C90.9352 12.6766 91.1116 13.3231 91.4634 13.7625C91.8171 14.2 92.325 14.4202 92.9899 14.4202C93.6973 14.4202 94.2385 14.1269 94.6162 13.5423ZM91.0719 15.2808C90.5307 14.955 90.1114 14.4882 89.815 13.8803C89.5176 13.2713 89.3699 12.5701 89.3699 11.7704C89.3699 10.8468 89.5259 10.0451 89.8381 9.35801C90.1493 8.67196 90.5898 8.14321 91.1568 7.77177C91.7248 7.40032 92.3823 7.2146 93.1285 7.2146C93.8645 7.2146 94.5044 7.38003 95.0456 7.71189C95.5886 8.04477 96.0079 8.51161 96.3043 9.11241C96.6017 9.71423 96.7494 10.4216 96.7494 11.2304C96.7494 12.153 96.5943 12.9557 96.2821 13.6357C95.9682 14.3156 95.5295 14.8424 94.9616 15.2128C94.3945 15.5842 93.7333 15.77 92.9761 15.77C92.2493 15.77 91.6149 15.6066 91.0719 15.2808ZM104.375 9.0708C104.007 9.47675 103.772 10.0654 103.669 10.8428H107.628C107.648 10.0664 107.502 9.47573 107.19 9.0708C106.879 8.66485 106.436 8.46391 105.863 8.46391C105.239 8.46391 104.743 8.66587 104.375 9.0708ZM108.817 11.6435C108.709 11.7622 108.564 11.8211 108.379 11.8211H103.608C103.618 12.6533 103.812 13.2977 104.192 13.7544C104.569 14.2091 105.117 14.4374 105.834 14.4374C106.549 14.4374 107.212 14.2121 107.827 13.7615C107.939 13.6824 108.053 13.6438 108.165 13.6438C108.308 13.6438 108.425 13.7057 108.518 13.8295C108.61 13.9533 108.655 14.0985 108.655 14.268C108.655 14.5278 108.553 14.746 108.348 14.9266C108.072 15.1732 107.692 15.3762 107.206 15.5335C106.721 15.6918 106.232 15.77 105.74 15.77C105.004 15.77 104.359 15.6066 103.808 15.2808C103.255 14.954 102.827 14.4872 102.527 13.8803C102.225 13.2713 102.074 12.5701 102.074 11.7714C102.074 10.8813 102.229 10.0918 102.542 9.39962C102.855 8.70748 103.291 8.17163 103.853 7.78902C104.416 7.40641 105.065 7.21359 105.801 7.21359C106.824 7.21359 107.61 7.5485 108.157 8.21831C108.704 8.88812 108.977 9.85834 108.977 11.129C108.977 11.3543 108.925 11.5258 108.817 11.6435ZM113.409 5.59487C113.148 5.90948 112.992 6.37632 112.941 6.99539L112.909 7.41758H114.429C114.634 7.41758 114.79 7.4734 114.897 7.58605C115.004 7.69768 115.058 7.84991 115.058 8.04071C115.058 8.50247 114.823 8.73285 114.353 8.73285H112.818L112.373 14.9256C112.363 15.1966 112.279 15.3985 112.121 15.5345C111.961 15.6685 111.775 15.7375 111.56 15.7375C111.345 15.7375 111.167 15.6685 111.023 15.5345C110.88 15.3985 110.813 15.1966 110.824 14.9256L111.268 8.73285H110.363C110.159 8.73285 110 8.67703 109.888 8.56438C109.774 8.45173 109.719 8.30051 109.719 8.1087C109.719 7.88442 109.78 7.7129 109.903 7.59416C110.025 7.47542 110.205 7.41758 110.439 7.41758H111.362L111.376 7.19735C111.54 5.03872 112.598 3.8909 114.551 3.75593L114.875 3.73867C115.416 3.70518 115.687 3.90207 115.687 4.32933C115.687 4.54245 115.632 4.70686 115.518 4.8185C115.406 4.93115 115.242 4.99813 115.029 5.02045L114.613 5.05395C114.071 5.09961 113.669 5.28026 113.409 5.59487ZM119.333 5.59487C119.072 5.90948 118.916 6.37632 118.864 6.99539L118.834 7.41758H120.353C120.559 7.41758 120.714 7.4734 120.821 7.58605C120.928 7.69768 120.982 7.84991 120.982 8.04071C120.982 8.50247 120.746 8.73285 120.277 8.73285H118.742L118.297 14.9256C118.287 15.1966 118.203 15.3985 118.044 15.5345C117.885 15.6685 117.699 15.7375 117.483 15.7375C117.269 15.7375 117.09 15.6685 116.947 15.5345C116.804 15.3985 116.737 15.1966 116.747 14.9256L117.193 8.73285H116.288C116.083 8.73285 115.924 8.67703 115.812 8.56438C115.698 8.45173 115.642 8.30051 115.642 8.1087C115.642 7.88442 115.704 7.7129 115.827 7.59416C115.949 7.47542 116.129 7.41758 116.364 7.41758H117.285L117.301 7.19735C117.464 5.03872 118.522 3.8909 120.475 3.75593L120.798 3.73867C121.34 3.70518 121.61 3.90207 121.61 4.32933C121.61 4.54245 121.555 4.70686 121.442 4.8185C121.329 4.93115 121.165 4.99813 120.952 5.02045L120.536 5.05395C119.995 5.09961 119.594 5.28026 119.333 5.59487ZM126.582 13.5423C126.96 12.9578 127.15 12.1712 127.15 11.1797C127.15 10.3363 126.974 9.6919 126.62 9.24739C126.268 8.80389 125.763 8.58163 125.11 8.58163C124.392 8.58163 123.846 8.87392 123.467 9.45848C123.088 10.0441 122.901 10.8316 122.901 11.8211C122.901 12.6766 123.076 13.3231 123.429 13.7625C123.782 14.2 124.29 14.4202 124.955 14.4202C125.662 14.4202 126.202 14.1269 126.582 13.5423ZM123.037 15.2808C122.495 14.955 122.077 14.4882 121.779 13.8803C121.484 13.2713 121.334 12.5701 121.334 11.7704C121.334 10.8468 121.491 10.0451 121.803 9.35801C122.114 8.67196 122.554 8.14321 123.123 7.77177C123.689 7.40032 124.348 7.2146 125.093 7.2146C125.83 7.2146 126.469 7.38003 127.011 7.71189C127.554 8.04477 127.971 8.51161 128.27 9.11241C128.566 9.71423 128.715 10.4216 128.715 11.2304C128.715 12.153 128.558 12.9557 128.247 13.6357C127.934 14.3156 127.494 14.8424 126.926 15.2128C126.36 15.5842 125.698 15.77 124.941 15.77C124.214 15.77 123.58 15.6066 123.038 15.2808H123.037ZM135.121 7.38307C135.239 7.49572 135.298 7.65912 135.298 7.87224C135.298 8.38982 135.042 8.67094 134.532 8.7156L134.116 8.76634C133.412 8.8455 132.887 9.0708 132.544 9.44123C132.202 9.81267 132.005 10.3597 131.953 11.0792L131.662 14.9256C131.641 15.2067 131.555 15.4127 131.402 15.5416C131.247 15.6715 131.059 15.7365 130.834 15.7365C130.608 15.7365 130.428 15.6685 130.287 15.5335C130.151 15.3985 130.092 15.1966 130.113 14.9256L130.619 8.02447C130.629 7.77684 130.716 7.58605 130.88 7.45005C131.043 7.31609 131.232 7.24911 131.446 7.24911C131.651 7.24911 131.815 7.31305 131.939 7.44193C132.061 7.57184 132.118 7.75959 132.106 8.00722L132.061 8.70037C132.571 7.83367 133.379 7.35567 134.486 7.26535L134.654 7.24911C134.848 7.22678 135.005 7.27144 135.121 7.38307ZM138.651 12.5975C138.641 12.6645 138.637 12.7659 138.637 12.9019C138.637 13.4073 138.741 13.792 138.952 14.0569C139.16 14.3217 139.455 14.4527 139.834 14.4527C139.956 14.4527 140.078 14.4445 140.2 14.4293C140.324 14.411 140.422 14.4039 140.493 14.4039C140.615 14.4039 140.714 14.4567 140.785 14.5633C140.856 14.6709 140.892 14.819 140.892 15.0098C140.892 15.292 140.793 15.4888 140.593 15.6015C140.393 15.7131 140.042 15.77 139.542 15.77C138.764 15.77 138.16 15.5335 137.731 15.0616C137.301 14.5876 137.086 13.9249 137.086 13.0694C137.086 12.8898 137.092 12.7497 137.103 12.6472L137.379 8.73386H136.473C136.268 8.73386 136.11 8.67703 135.997 8.56438C135.885 8.45274 135.828 8.29443 135.828 8.09145C135.828 7.87833 135.89 7.7129 136.013 7.59416C136.136 7.47644 136.314 7.41656 136.549 7.41656H137.47L137.592 5.66185C137.613 5.4264 137.7 5.23561 137.854 5.08744C138.008 4.94231 138.196 4.86924 138.422 4.86924C138.657 4.86924 138.84 4.94231 138.974 5.08744C139.108 5.23561 139.162 5.43249 139.143 5.67911L139.019 7.41656H140.539C140.733 7.41656 140.886 7.4734 141.001 7.58503C141.112 7.69768 141.167 7.8489 141.167 8.04172C141.167 8.50349 140.932 8.73386 140.462 8.73386H138.928L138.651 12.5975ZM152.314 7.94024C152.748 8.42433 152.966 9.13778 152.966 10.0836C152.966 10.3089 152.962 10.4774 152.951 10.5901L152.644 14.9256C152.623 15.1844 152.539 15.3843 152.391 15.5254C152.243 15.6654 152.052 15.7365 151.816 15.7365C151.581 15.7365 151.396 15.6654 151.263 15.5254C151.13 15.3843 151.074 15.1844 151.094 14.9266L151.403 10.6743C151.411 10.5951 151.417 10.4774 151.417 10.3191C151.417 9.17229 150.916 8.59787 149.914 8.59787C149.279 8.59787 148.776 8.80592 148.402 9.22202C148.028 9.63811 147.817 10.2186 147.765 10.9605L147.473 14.9256C147.453 15.1844 147.369 15.3843 147.221 15.5254C147.072 15.6654 146.88 15.7365 146.645 15.7365C146.411 15.7365 146.228 15.6654 146.1 15.5254C145.973 15.3843 145.914 15.1844 145.925 14.9266L146.431 8.02548C146.441 7.78902 146.528 7.59924 146.692 7.45919C146.856 7.31812 147.05 7.24809 147.274 7.24809C147.488 7.24809 147.658 7.31609 147.781 7.45107C147.904 7.58605 147.959 7.77177 147.95 8.00722L147.904 8.59787C148.16 8.15945 148.504 7.81947 148.94 7.57793C149.374 7.33537 149.863 7.2146 150.405 7.2146C151.243 7.2146 151.88 7.45614 152.314 7.94024ZM159.735 13.5423C160.114 12.9578 160.303 12.1712 160.303 11.1797C160.303 10.3363 160.126 9.6919 159.773 9.24739C159.422 8.80389 158.916 8.58163 158.263 8.58163C157.546 8.58163 156.999 8.87392 156.62 9.45848C156.243 10.0441 156.053 10.8316 156.053 11.8211C156.053 12.6766 156.229 13.3231 156.582 13.7625C156.935 14.2 157.444 14.4202 158.108 14.4202C158.815 14.4202 159.357 14.1269 159.735 13.5423H159.735ZM156.191 15.2808C155.65 14.955 155.229 14.4882 154.932 13.8803C154.637 13.2713 154.489 12.5701 154.489 11.7704C154.489 10.8468 154.644 10.0451 154.956 9.35801C155.269 8.67196 155.707 8.14321 156.276 7.77177C156.844 7.40032 157.501 7.2146 158.248 7.2146C158.984 7.2146 159.624 7.38003 160.164 7.71189C160.707 8.04477 161.126 8.51161 161.423 9.11241C161.72 9.71423 161.868 10.4216 161.868 11.2304C161.868 12.153 161.713 12.9557 161.4 13.6357C161.087 14.3156 160.648 14.8424 160.08 15.2128C159.513 15.5842 158.851 15.77 158.094 15.77C157.368 15.77 156.733 15.6066 156.191 15.2808ZM169.647 7.94024C170.081 8.42433 170.299 9.13778 170.299 10.0836C170.299 10.3089 170.294 10.4774 170.283 10.5901L169.977 14.9256C169.956 15.1844 169.871 15.3843 169.724 15.5254C169.576 15.6654 169.383 15.7365 169.148 15.7365C168.913 15.7365 168.729 15.6654 168.595 15.5254C168.462 15.3843 168.407 15.1844 168.426 14.9266L168.734 10.6743C168.743 10.5951 168.749 10.4774 168.749 10.3191C168.749 9.17229 168.248 8.59787 167.246 8.59787C166.612 8.59787 166.108 8.80592 165.734 9.22202C165.36 9.63811 165.15 10.2186 165.097 10.9605L164.806 14.9256C164.786 15.1844 164.701 15.3843 164.553 15.5254C164.406 15.6654 164.212 15.7365 163.977 15.7365C163.742 15.7365 163.561 15.6654 163.433 15.5254C163.305 15.3843 163.246 15.1844 163.257 14.9266L163.762 8.02548C163.773 7.78902 163.861 7.59924 164.023 7.45919C164.188 7.31812 164.381 7.24809 164.606 7.24809C164.821 7.24809 164.99 7.31609 165.114 7.45107C165.237 7.58605 165.292 7.77177 165.282 8.00722L165.236 8.59787C165.491 8.15945 165.837 7.81947 166.271 7.57793C166.705 7.33537 167.194 7.2146 167.736 7.2146C168.575 7.2146 169.212 7.45614 169.647 7.94024ZM177.461 13.5768C177.844 13.0136 178.035 12.2372 178.035 11.2477C178.035 10.3932 177.856 9.73453 177.499 9.27276C177.14 8.81201 176.631 8.58163 175.965 8.58163C175.291 8.58163 174.758 8.8658 174.369 9.43311C173.981 10.0014 173.786 10.7758 173.786 11.7531C173.786 12.6208 173.965 13.2815 174.324 13.7362C174.681 14.1929 175.192 14.4202 175.857 14.4202C176.543 14.4202 177.077 14.1391 177.461 13.5768ZM178.09 7.70377C178.575 8.03056 178.951 8.49436 179.216 9.09617C179.484 9.69799 179.616 10.3983 179.616 11.197C179.616 12.0971 179.467 12.8928 179.171 13.5849C178.876 14.2771 178.457 14.815 177.921 15.1966C177.383 15.5781 176.766 15.7689 176.073 15.7689C175.49 15.7689 174.985 15.635 174.561 15.365C174.137 15.0951 173.811 14.7125 173.587 14.2172L173.541 14.9256C173.53 15.1844 173.449 15.3843 173.295 15.5254C173.142 15.6654 172.951 15.7365 172.728 15.7365C172.502 15.7365 172.323 15.6654 172.19 15.5254C172.057 15.3843 171.997 15.1844 172.006 14.9266L172.773 4.44705C172.783 4.21059 172.87 4.02081 173.035 3.87365C173.198 3.72751 173.393 3.65342 173.617 3.65342C173.842 3.65342 174.019 3.72142 174.145 3.8564C174.275 3.99138 174.333 4.1771 174.324 4.41356L174.016 8.54713C174.271 8.13103 174.612 7.80627 175.037 7.56879C175.461 7.33233 175.923 7.2146 176.425 7.2146C177.049 7.2146 177.603 7.378 178.09 7.70377ZM185.121 13.8803C185.49 13.4642 185.698 12.9293 185.749 12.2768L185.781 11.8049H185.351C184.594 11.8049 184.001 11.8495 183.571 11.9398C183.141 12.0291 182.835 12.1763 182.651 12.3772C182.466 12.5802 182.374 12.8624 182.374 13.2216C182.374 13.6154 182.499 13.928 182.75 14.1583C183.002 14.3887 183.341 14.5034 183.771 14.5034C184.302 14.5034 184.753 14.2964 185.122 13.8803H185.121ZM187.377 10.1171C187.377 10.3303 187.372 10.4947 187.36 10.6063L187.056 14.9601C187.044 15.2077 186.97 15.3985 186.831 15.5345C186.693 15.6685 186.512 15.7375 186.286 15.7375C186.062 15.7375 185.883 15.6604 185.75 15.5081C185.618 15.3559 185.561 15.1621 185.582 14.9256L185.611 14.3522C185.397 14.814 185.102 15.1641 184.722 15.4066C184.343 15.6482 183.903 15.77 183.403 15.77C182.655 15.77 182.039 15.5477 181.555 15.1032C181.068 14.6587 180.826 14.0883 180.826 13.3911C180.826 12.7609 180.973 12.2626 181.271 11.8972C181.567 11.5319 182.039 11.2649 182.689 11.0955C183.339 10.927 184.211 10.8417 185.305 10.8417H185.858L185.888 10.4206C185.929 9.77918 185.826 9.31031 185.582 9.01194C185.337 8.71357 184.937 8.56438 184.385 8.56438C183.709 8.56438 182.963 8.81302 182.145 9.30727C182.052 9.36308 181.951 9.3915 181.838 9.3915C181.685 9.3915 181.557 9.32959 181.454 9.20578C181.352 9.08196 181.301 8.93582 181.301 8.76735C181.301 8.4974 181.429 8.2721 181.685 8.09247C182.052 7.82251 182.489 7.60837 182.996 7.45005C183.502 7.29275 184.017 7.2146 184.538 7.2146C186.43 7.2146 187.377 8.18279 187.377 10.1171ZM195.482 7.94024C195.917 8.42433 196.134 9.13778 196.134 10.0836C196.134 10.3089 196.129 10.4774 196.119 10.5901L195.812 14.9256C195.791 15.1844 195.707 15.3843 195.558 15.5254C195.41 15.6654 195.219 15.7365 194.983 15.7365C194.748 15.7365 194.564 15.6654 194.432 15.5254C194.298 15.3843 194.241 15.1844 194.262 14.9266L194.569 10.6743C194.578 10.5951 194.584 10.4774 194.584 10.3191C194.584 9.17229 194.083 8.59787 193.081 8.59787C192.447 8.59787 191.943 8.80592 191.569 9.22202C191.197 9.63811 190.983 10.2186 190.933 10.9605L190.642 14.9256C190.621 15.1844 190.537 15.3843 190.389 15.5254C190.239 15.6654 190.048 15.7365 189.813 15.7365C189.578 15.7365 189.396 15.6654 189.268 15.5254C189.141 15.3843 189.081 15.1844 189.092 14.9266L189.598 8.02548C189.607 7.78902 189.695 7.59924 189.858 7.45919C190.023 7.31812 190.217 7.24809 190.441 7.24809C190.656 7.24809 190.825 7.31609 190.948 7.45107C191.071 7.58605 191.126 7.77177 191.118 8.00722L191.072 8.59787C191.326 8.15945 191.672 7.81947 192.106 7.57793C192.541 7.33537 193.03 7.2146 193.573 7.2146C194.411 7.2146 195.048 7.45614 195.482 7.94024ZM204.101 14.4537C204.224 14.5876 204.286 14.751 204.286 14.9428C204.286 15.1458 204.206 15.3305 204.048 15.5C203.888 15.6695 203.713 15.7537 203.519 15.7537C203.334 15.7537 203.17 15.6695 203.027 15.5L199.636 11.8038L199.405 14.9428C199.385 15.2026 199.298 15.3995 199.146 15.5335C198.993 15.6695 198.809 15.7365 198.594 15.7365C198.368 15.7365 198.184 15.6634 198.041 15.5162C197.897 15.3701 197.836 15.1681 197.857 14.9093L198.609 4.46431C198.619 4.20551 198.705 4.00863 198.868 3.87264C199.032 3.73867 199.222 3.66966 199.437 3.66966C199.662 3.66966 199.841 3.7407 199.974 3.88177C200.106 4.02284 200.169 4.22683 200.157 4.4978L199.683 11.0945L203.303 7.53429C203.447 7.378 203.61 7.29884 203.795 7.29884C203.979 7.29884 204.139 7.37495 204.279 7.52718C204.416 7.67941 204.485 7.84991 204.485 8.04172C204.485 8.24368 204.408 8.42331 204.255 8.58163L201.339 11.3999L204.101 14.4537ZM215.891 3.80058C215.976 3.89395 216.011 4.02182 215.997 4.18623L215.71 8.14219C215.689 8.49943 215.534 8.67906 215.243 8.67906C215.109 8.67906 215.007 8.63441 214.937 8.5451C214.866 8.45477 214.836 8.32183 214.852 8.14219L215.031 5.62126L213.982 7.56981C213.911 7.69464 213.839 7.78496 213.769 7.83875C213.698 7.89254 213.612 7.91994 213.513 7.91994C213.422 7.91994 213.345 7.89254 213.286 7.83875C213.225 7.78395 213.171 7.69464 213.122 7.56981L212.336 5.69128L212.156 8.14219C212.149 8.32183 212.102 8.45477 212.017 8.5451C211.932 8.63441 211.82 8.67906 211.679 8.67906C211.544 8.67906 211.441 8.63441 211.37 8.5451C211.3 8.45477 211.268 8.32183 211.276 8.14219L211.562 4.18522C211.576 4.02284 211.63 3.89395 211.727 3.79957C211.822 3.70721 211.932 3.66053 212.06 3.66053C212.245 3.66053 212.385 3.7813 212.485 4.02182L213.588 6.75284L215.084 4.02284C215.219 3.7813 215.383 3.66053 215.574 3.66053C215.7 3.66053 215.806 3.70721 215.891 3.80058H215.891ZM211.009 4.12737C211.009 4.43893 210.843 4.59421 210.511 4.59421H209.471L209.216 8.15336C209.209 8.31675 209.156 8.44564 209.052 8.53901C208.951 8.63238 208.827 8.67906 208.686 8.67906C208.544 8.67906 208.43 8.63238 208.341 8.53901C208.252 8.44564 208.212 8.31675 208.219 8.15336L208.483 4.59421H207.455C207.165 4.59421 207.02 4.45416 207.02 4.17405C207.02 3.86249 207.182 3.70721 207.508 3.70721H210.574C210.865 3.70721 211.009 3.84726 211.009 4.12737Z" fill="white"/>
-  </svg>
-`;
-
-/**
- * @function IconMoon
- * @returns {string} HTML
- */
-
-export const IconMoon = () => `
-  <svg width="24" height="24" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-    <path d="M23.9999 19.314C8.10542 22.1253 2.68539 9.23676 12.2615 0C-1.80277 0.644425 -1.87046 14.9797 2.68539 19.9371C11.4083 29.4287 23.9999 19.314 23.9999 19.314Z" fill="currentColor"/>
-    <path d="M15.1999 5.60003C14.9959 5.60003 14.8939 5.49485 14.8939 5.28449V3.46777H13.1059C12.9019 3.46777 12.7999 3.36896 12.7999 3.17135C12.7999 2.97374 12.9019 2.87494 13.1059 2.87494H14.8939V1.11559C14.8939 0.905228 14.9959 0.800049 15.1999 0.800049C15.4039 0.800049 15.5059 0.905228 15.5059 1.11559V2.87494H17.2939C17.4979 2.87494 17.5999 2.97374 17.5999 3.17135C17.5999 3.36896 17.4979 3.46777 17.2939 3.46777H15.5059V5.28449C15.5059 5.49485 15.4039 5.60003 15.1999 5.60003Z" fill="currentColor"/>
-    <path d="M21.1997 8.80015C21.0297 8.80015 20.9447 8.71251 20.9447 8.53721V7.02327H19.4547C19.2847 7.02327 19.1997 6.94093 19.1997 6.77626C19.1997 6.61158 19.2847 6.52925 19.4547 6.52925H20.9447V5.06312C20.9447 4.88782 21.0297 4.80017 21.1997 4.80017C21.3697 4.80017 21.4547 4.88782 21.4547 5.06312V6.52925H22.9447C23.1147 6.52925 23.1997 6.61158 23.1997 6.77626C23.1997 6.94093 23.1147 7.02327 22.9447 7.02327H21.4547V8.53721C21.4547 8.71251 21.3697 8.80015 21.1997 8.80015Z" fill="currentColor"/>
-    <path d="M13.1999 12.7998C13.03 12.7998 12.945 12.7121 12.945 12.5368V11.0229H11.4549C11.2849 11.0229 11.2 10.9406 11.2 10.7759C11.2 10.6112 11.2849 10.5289 11.4549 10.5289H12.945V9.06275C12.945 8.88745 13.03 8.7998 13.1999 8.7998C13.3699 8.7998 13.4549 8.88745 13.4549 9.06275V10.5289H14.945C15.1149 10.5289 15.1999 10.6112 15.1999 10.7759C15.1999 10.9406 15.1149 11.0229 14.945 11.0229H13.4549V12.5368C13.4549 12.7121 13.3699 12.7998 13.1999 12.7998Z" fill="currentColor"/>
-    <path d="M21.1997 16.7999C21.0297 16.7999 20.9447 16.7123 20.9447 16.537V15.023H19.4547C19.2847 15.023 19.1997 14.9407 19.1997 14.776C19.1997 14.6113 19.2847 14.529 19.4547 14.529H20.9447V13.0629C20.9447 12.8876 21.0297 12.7999 21.1997 12.7999C21.3697 12.7999 21.4547 12.8876 21.4547 13.0629V14.529H22.9447C23.1147 14.529 23.1997 14.6113 23.1997 14.776C23.1997 14.9407 23.1147 15.023 22.9447 15.023H21.4547V16.537C21.4547 16.7123 21.3697 16.7999 21.1997 16.7999Z" fill="currentColor"/>
-  </svg>
-`;
-
-/**
- * @function IconSun
- * @returns {string} HTML
- */
-
-export const IconSun = () => `
-  <svg viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
-    <path d="M11.9856 0.00257827C11.9612 0.0070376 11.9367 0.0137257 11.9145 0.0204144C11.6499 0.0806116 11.4631 0.31917 11.4698 0.591171V4.01571C11.4676 4.22083 11.5743 4.41257 11.7522 4.51735C11.93 4.61991 12.1479 4.61991 12.3258 4.51735C12.5037 4.41257 12.6104 4.22083 12.6082 4.01571V0.591171C12.6149 0.426187 12.5482 0.267891 12.4303 0.156415C12.3102 0.0427098 12.1479 -0.0130285 11.9856 0.00257827ZM3.8745 3.35577C3.66105 3.39368 3.48985 3.54974 3.42981 3.75709C3.36978 3.96666 3.43204 4.18961 3.5899 4.33676L6.00901 6.76248C6.14686 6.93192 6.36698 7.00996 6.58043 6.96091C6.79166 6.91186 6.95842 6.74464 7.00733 6.53284C7.05625 6.3188 6.97843 6.09808 6.80945 5.95985L4.39034 3.53414C4.2725 3.40705 4.10129 3.3424 3.92787 3.35577C3.91008 3.35577 3.89229 3.35577 3.8745 3.35577ZM20.0256 3.35577C19.8966 3.37361 19.7766 3.43827 19.6876 3.53414L17.2685 5.95985C17.0995 6.09808 17.0217 6.3188 17.0706 6.53284C17.1196 6.74464 17.2863 6.91186 17.4975 6.96091C17.711 7.00996 17.9311 6.93192 18.069 6.76248L20.4881 4.33676C20.6659 4.16732 20.7171 3.90201 20.6126 3.67905C20.5059 3.45387 20.2702 3.32456 20.0256 3.35577ZM11.9856 6.29874C11.9678 6.3032 11.9501 6.30989 11.9323 6.31657C11.8967 6.3188 11.8611 6.32549 11.8255 6.33441C11.8189 6.3411 11.8144 6.34556 11.8077 6.35225C8.78831 6.48156 6.34697 8.94741 6.34697 12.0063C6.34697 15.1477 8.90616 17.7139 12.039 17.7139C15.1718 17.7139 17.731 15.1477 17.731 12.0063C17.731 8.95855 15.3097 6.50163 12.3058 6.35225C12.2858 6.35225 12.2724 6.33441 12.2524 6.33441C12.1946 6.31212 12.1346 6.30097 12.0746 6.29874C12.0634 6.29874 12.0501 6.29874 12.039 6.29874C12.0212 6.29874 12.0034 6.29874 11.9856 6.29874ZM12.0034 7.44025C12.0145 7.44025 12.0279 7.44025 12.039 7.44025C12.0568 7.44025 12.0746 7.44025 12.0924 7.44025C14.5826 7.46924 16.5926 9.50256 16.5926 12.0063C16.5926 14.5301 14.5559 16.5724 12.039 16.5724C9.52427 16.5724 7.48537 14.5301 7.48537 12.0063C7.48537 9.49587 9.50426 7.46032 12.0034 7.44025ZM0.494869 11.4355C0.181363 11.4801 -0.0387576 11.7722 0.0057113 12.0866C0.0501802 12.4009 0.341451 12.6217 0.654957 12.5771H4.07017C4.27472 12.5793 4.46594 12.4723 4.57044 12.2939C4.67272 12.1156 4.67272 11.8971 4.57044 11.7187C4.46594 11.5403 4.27472 11.4333 4.07017 11.4355H0.654957C0.637169 11.4355 0.619382 11.4355 0.601594 11.4355C0.583807 11.4355 0.566019 11.4355 0.548231 11.4355C0.530444 11.4355 0.512656 11.4355 0.494869 11.4355ZM19.8477 11.4355C19.5342 11.4801 19.3141 11.7722 19.3586 12.0866C19.403 12.4009 19.6943 12.6217 20.0078 12.5771H23.423C23.6276 12.5793 23.8188 12.4723 23.9233 12.2939C24.0256 12.1156 24.0256 11.8971 23.9233 11.7187C23.8188 11.5403 23.6276 11.4333 23.423 11.4355H20.0078C19.99 11.4355 19.9722 11.4355 19.9544 11.4355C19.9367 11.4355 19.9189 11.4355 19.9011 11.4355C19.8833 11.4355 19.8655 11.4355 19.8477 11.4355ZM6.34697 17.0718C6.21801 17.0896 6.09795 17.1543 6.00901 17.2501L3.5899 19.6758C3.42092 19.8141 3.3431 20.0348 3.39202 20.2488C3.44093 20.4606 3.60769 20.6279 3.81892 20.6769C4.03237 20.726 4.25249 20.6479 4.39034 20.4785L6.80945 18.0528C6.97843 17.89 7.02957 17.6381 6.93841 17.4218C6.84947 17.2055 6.6338 17.0673 6.40033 17.0718C6.38255 17.0718 6.36476 17.0718 6.34697 17.0718ZM17.5531 17.0718C17.3397 17.1097 17.1685 17.2657 17.1084 17.4731C17.0484 17.6827 17.1107 17.9056 17.2685 18.0528L19.6876 20.4785C19.8255 20.6479 20.0456 20.726 20.2591 20.6769C20.4703 20.6279 20.637 20.4606 20.686 20.2488C20.7349 20.0348 20.6571 19.8141 20.4881 19.6758L18.069 17.2501C17.9622 17.1364 17.8155 17.074 17.6599 17.0718C17.6421 17.0718 17.6243 17.0718 17.6065 17.0718C17.5887 17.0718 17.5709 17.0718 17.5531 17.0718ZM11.9856 19.4083C11.9612 19.4128 11.9367 19.4195 11.9145 19.4261C11.6499 19.4863 11.4631 19.7249 11.4698 19.9969V23.4214C11.4676 23.6266 11.5743 23.8183 11.7522 23.9231C11.93 24.0256 12.1479 24.0256 12.3258 23.9231C12.5037 23.8183 12.6104 23.6266 12.6082 23.4214V19.9969C12.6149 19.8319 12.5482 19.6736 12.4303 19.5621C12.3102 19.4484 12.1479 19.3927 11.9856 19.4083Z" fill="currentColor"/>
-  </svg>
-`;
+export { IconLogo } from './IconLogo/index.js';
+export { IconLogo2 } from './IconLogo2/index.js';
+export { IconApple } from './IconApple/index.js';
+export { IconGoogle } from './IconGoogle/index.js';
+export { IconIn } from './IconIn/index.js';
+export { IconArrowDown } from './IconArrowDown/index.js';
+export { IconClose } from './IconClose/index.js';
+export { IconDeveloper } from './IconDeveloper/index.js';
+export { IconMoon } from './IconMoon/index.js';
+export { IconSun } from './IconSun/index.js';
