+++ js/utils/index.js
@@ -1,50 +1,5 @@
-/**
- * @function scrollToTarget
- * @param {string} targetId
- * @returns {void}
- */
-
-export const scrollToTarget = (targetId) => {
-  const $targetElement = document.querySelector(targetId);
-  /** @type {HTMLElement | null} */
-  const $header = document.querySelector('#header');
-  const headerHeight = $header?.offsetHeight || 0;
-
-  if (!$targetElement) return;
-
-  const targetPosition =
-    $targetElement.getBoundingClientRect().top +
-    window.scrollY -
-    headerHeight;
-
-  window.scrollTo({
-    top: targetPosition,
-    behavior: 'smooth',
-  });
-};
-
-/**
- * @function toggleModal
- * @returns {void}
- */
-
-export const toggleModal = () => {
-  const $modal = document.querySelector('#modal');
-  if (!$modal) return;
-  $modal.classList.toggle('active');
-};
-
-/**
- * @function toggleNavigation
- * @returns {void}
- */
-
-export const toggleNavigation = () => {
-  const $nav = document.querySelector('#nav');
-  const $burgerButton = document.querySelector('#burger');
-
-  if (!$nav || !$burgerButton) return;
-
-  $nav.classList.toggle('active');
-  $burgerButton.classList.toggle('active');
-};
+export { toCapitalize } from './toCapitalize/index.js';
+export { toggleNavigation } from './toggleNavigation/index.js';
+export { toggleModal } from './toggleModal/index.js';
+export { scrollToTarget } from './scrollToTarget/index.js';
+export { loadDataFromApi } from './loadDataFromApi/index.js';
