+++ js/utils/scrollToTarget/index.js
@@ -1,0 +1,1 @@
+export { scrollToTarget } from './scrollToTarget.js';
