+++ js/utils/toggleModal/index.js
@@ -1,0 +1,1 @@
+export { toggleModal } from './toggleModal.js';
