+++ js/utils/toggleModal/toggleModal.js
@@ -1,0 +1,10 @@
+/**
+ * @function toggleModal
+ * @returns {void}
+ */
+
+export const toggleModal = () => {
+  const $modal = document.querySelector('#modal');
+  if (!$modal) return;
+  $modal.classList.toggle('active');
+};
