+++ js/widgets/Care/Care.js
@@ -1,5 +1,5 @@
-import { Title } from '../../ui/Title/index.js';
-import { Image } from '../../ui/Image/index.js';
+import { Title } from '../../ui/index.js';
+import { Image } from '../../ui/index.js';
 
 /**
 * @typedef {import('./types').CareData} CareData
