+++ js/widgets/Care/types.ts
@@ -1,8 +1,1 @@
-import { TitleData } from '../../ui/Title/types';
-import { ImageData } from '../../ui/Image/types';
-
-export type CareData = {
-  title: TitleData;
-  texts: string[];
-  image: ImageData;
-}
+export type { CareData } from '../../types';
