+++ js/widgets/Cashback/types.ts
@@ -1,7 +1,1 @@
-import { TitleData } from '../../ui/Title/types';
-
-export type CashbackData = {
-  title: TitleData;
-  texts: string[];
-  buttonText: string;
-}
+export type { CashbackData } from '../../types';
