+++ js/widgets/Clients/types.ts
@@ -1,13 +1,1 @@
-type BrandLogoData = {
-  lightSource: string;
-  darkSource: string;
-}
-
-export type BrandData = {
-  name: string;
-  logo: BrandLogoData;
-}
-
-export type ClientsData = {
-  brands: BrandData[];
-}
+export type { ClientsData } from '../../types';
