+++ js/widgets/Download/types.ts
@@ -1,10 +1,1 @@
-import { TitleData } from '../../ui/Title/types';
-import { LinkData } from '../../ui/Link/types';
-import { ImageData } from '../../ui/Image/types';
-
-export type DownloadData = {
-  title: TitleData;
-  texts: string[];
-  links: LinkData[];
-  image: ImageData;
-};
+export type { DownloadData } from '../../types';
