+++ js/widgets/Footer/types.ts
@@ -1,10 +1,1 @@
-import { ColumnData } from '../../ui/Column/types';
-import { LinkData } from '../../ui/Column/types';
-
-export type FooterData = {
-  categories: ColumnData[];
-  texts: string[];
-  developer: LinkData;
-}
-
-
+export type { FooterData } from '../../types';
