+++ js/widgets/Header/types.ts
@@ -1,7 +1,1 @@
-import { NavItemData } from '../../features/Navigation/types';
-import { LangData } from '../../features/Lang/types';
-
-export type HeaderData = {
-  langs: LangData[];
-  navigation: NavItemData[];
-};
+export type { HeaderData } from '../../types';
