+++ js/widgets/index.js
@@ -1,0 +1,8 @@
+export { Header } from './Header/index.js';
+export { Download } from './Download/index.js';
+export { Warranty } from './Warranty/index.js';
+export { Care } from './Care/index.js';
+export { Cashback } from './Cashback/index.js';
+export { Clients } from './Clients/index.js';
+export { Footer } from './Footer/index.js';
+export { Modal } from './Modal/index.js';
